Network.AWS.StorageGateway

Description

AWS Storage Gateway is a service that connects an on-premises software appliance with cloud-based storage to provide seamless and secure integration between your on-premises IT environment and AWS's storage infrastructure.

Documentation

module Network.AWS.StorageGateway.ActivateGateway

module Network.AWS.StorageGateway.AddCache

module Network.AWS.StorageGateway.AddUploadBuffer

module Network.AWS.StorageGateway.AddWorkingStorage

module Network.AWS.StorageGateway.CancelArchival

module Network.AWS.StorageGateway.CancelRetrieval

module Network.AWS.StorageGateway.CreateCachediSCSIVolume

module Network.AWS.StorageGateway.CreateSnapshot

module Network.AWS.StorageGateway.CreateSnapshotFromVolumeRecoveryPoint

module Network.AWS.StorageGateway.CreateStorediSCSIVolume

module Network.AWS.StorageGateway.CreateTapes

module Network.AWS.StorageGateway.DeleteBandwidthRateLimit

module Network.AWS.StorageGateway.DeleteChapCredentials

module Network.AWS.StorageGateway.DeleteGateway

module Network.AWS.StorageGateway.DeleteSnapshotSchedule

module Network.AWS.StorageGateway.DeleteTape

module Network.AWS.StorageGateway.DeleteTapeArchive

module Network.AWS.StorageGateway.DeleteVolume

module Network.AWS.StorageGateway.DescribeBandwidthRateLimit

module Network.AWS.StorageGateway.DescribeCache

module Network.AWS.StorageGateway.DescribeCachediSCSIVolumes

module Network.AWS.StorageGateway.DescribeChapCredentials

module Network.AWS.StorageGateway.DescribeGatewayInformation

module Network.AWS.StorageGateway.DescribeMaintenanceStartTime

module Network.AWS.StorageGateway.DescribeSnapshotSchedule

module Network.AWS.StorageGateway.DescribeStorediSCSIVolumes

module Network.AWS.StorageGateway.DescribeTapeArchives

module Network.AWS.StorageGateway.DescribeTapeRecoveryPoints

module Network.AWS.StorageGateway.DescribeTapes

module Network.AWS.StorageGateway.DescribeUploadBuffer

module Network.AWS.StorageGateway.DescribeVTLDevices

module Network.AWS.StorageGateway.DescribeWorkingStorage

module Network.AWS.StorageGateway.DisableGateway

module Network.AWS.StorageGateway.ListGateways

module Network.AWS.StorageGateway.ListLocalDisks

module Network.AWS.StorageGateway.ListVolumeRecoveryPoints

module Network.AWS.StorageGateway.ListVolumes

module Network.AWS.StorageGateway.ResetCache

module Network.AWS.StorageGateway.RetrieveTapeArchive

module Network.AWS.StorageGateway.RetrieveTapeRecoveryPoint

module Network.AWS.StorageGateway.ShutdownGateway

module Network.AWS.StorageGateway.StartGateway

module Network.AWS.StorageGateway.Types

module Network.AWS.StorageGateway.UpdateBandwidthRateLimit

module Network.AWS.StorageGateway.UpdateChapCredentials

module Network.AWS.StorageGateway.UpdateGatewayInformation

module Network.AWS.StorageGateway.UpdateGatewaySoftwareNow

module Network.AWS.StorageGateway.UpdateMaintenanceStartTime

module Network.AWS.StorageGateway.UpdateSnapshotSchedule

module Network.AWS.StorageGateway.UpdateVTLDeviceType