Network.AWS.StorageGateway.ActivateGateway

Description

This operation activates the gateway you previously deployed on your host. For more information, see Downloading and Deploying AWS Storage Gateway VM. In the activation process you specify information such as the region you want to use for storing snapshots, the time zone for scheduled snapshots and the gateway schedule window, an activation key, and a name for your gateway. The activation process also associates your gateway with your account (see UpdateGatewayInformation). You must power on the gateway VM before you can activate your gateway. Example Request The following example shows a request that activates a gateway. POST HTTP1.1 Host: storagegateway.us-east-1.amazonaws.com x-amz-Date: 20120425T120000Z Authorization: CSOC7TJPLR0OOKIRLGOHVAICUFVV4KQNSO5AEMVJF66Q9ASUAAJG Content-type: application/x-amz-json-1.1 x-amz-target: StorageGateway_20120630.ActivateGateway { ActivationKey: "29AV1-3OFV9-VVIUB-NKT0I-LRO6V", GatewayName: "mygateway", GatewayTimezone: GMT-12:00, GatewayRegion: "us-east-1", GatewayType: STORED } HTTP/1.1 200 OK x-amzn-RequestId: CSOC7TJPLR0OOKIRLGOHVAICUFVV4KQNSO5AEMVJF66Q9ASUAAJG Date: Wed, 25 Apr 2012 12:00:02 GMT Content-type: application/x-amz-json-1.1 Content-length: 80 { GatewayARN: "arn:aws:storagegateway:us-east-1:111122223333:gateway/mygateway" }.

Request

data ActivateGateway

A JSON object containing one or more of the following fields: ActivateGatewayInput$ActivationKey GatewayName ActivateGatewayInput$GatewayRegion ActivateGatewayInput$GatewayTimezone ActivateGatewayInput$GatewayType.

Request constructor

mkActivateGateway

Arguments

:: Text
agActivationKey

-> Text
agGatewayName

-> Text
agGatewayTimezone

-> Text
agGatewayRegion

-> ActivateGateway

Smart constructor for the minimum required parameters to construct a valid ActivateGateway request.

The fields accessible through corresponding lenses are:

ActivationKey :: Text

GatewayName :: Text

GatewayTimezone :: Text

GatewayRegion :: Text

GatewayType :: Maybe Text

Request lenses

agActivationKey :: Lens' ActivateGateway Text

Your gateway activation key. You can obtain the activation key by sending an HTTP GET request with redirects enabled to the gateway IP address (port 80). The redirect URL returned in the response provides you the activation key for your gateway in the query string parameter activationKey. It may also include other activation-related parameters, however, these are merely defaults -- the arguments you pass to the ActivateGateway API call determine the actual configuration of your gateway.

agGatewayName :: Lens' ActivateGateway Text

A unique identifier for your gateway. This name becomes part of the gateway Amazon Resources Name (ARN) which is what you use as an input to other operations.

agGatewayTimezone :: Lens' ActivateGateway Text

One of the values that indicates the time zone you want to set for the gateway. The time zone is used, for example, for scheduling snapshots and your gateway's maintenance schedule.

agGatewayRegion :: Lens' ActivateGateway Text

One of the values that indicates the region where you want to store the snapshot backups. The gateway region specified must be the same region as the region in your Host header in the request. For more information about available regions and endpoints for AWS Storage Gateway, see Regions and Endpoints in the Amazon Web Services Glossary. Valid Values: "us-east-1", "us-west-1", "us-west-2", "eu-west-1", "ap-northeast-1", "ap-southest-1", "sa-east-1".

agGatewayType :: Lens' ActivateGateway (Maybe Text)

One of the values that defines the type of gateway to activate. The type specified is critical to all later functions of the gateway and cannot be changed after activation. The default value is STORED.

Response

data ActivateGatewayResponse

AWS Storage Gateway returns the Amazon Resource Name (ARN) of the activated gateway. It is a string made of information such as your account, gateway name, and region. This ARN is used to reference the gateway in other API operations as well as resource-based authorization.

Response constructor

mkActivateGatewayResponse :: ActivateGatewayResponse

Smart constructor for the minimum required parameters to construct a valid ActivateGatewayResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

GatewayARN :: Maybe Text

Response lenses

agrGatewayARN :: Lens' ActivateGatewayResponse (Maybe Text)

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.