Network.AWS.StorageGateway.AddCache

Description

This operation configures one or more gateway local disks as cache for a cached-volume gateway. This operation is supported only for the gateway-cached volume architecture (see Storage Gateway Concepts).

In the request, you specify the gateway Amazon Resource Name (ARN) to which you want to add cache, and one or more disk IDs that you want to configure as cache.

http://docs.aws.amazon.com/storagegateway/latest/APIReference/API_AddCache.html

Request

data AddCache

Request constructor

addCache

Arguments

:: Text
acGatewayARN

-> AddCache

AddCache constructor.

The fields accessible through corresponding lenses are:

acDiskIds :: [Text]

acGatewayARN :: Text

Request lenses

acDiskIds :: Lens' AddCache [Text]

acGatewayARN :: Lens' AddCache Text

Response

data AddCacheResponse

Response constructor

addCacheResponse :: AddCacheResponse

AddCacheResponse constructor.

The fields accessible through corresponding lenses are:

acrGatewayARN :: Maybe Text

Response lenses

acrGatewayARN :: Lens' AddCacheResponse (Maybe Text)