Network.AWS.StorageGateway.AddTagsToResource

Description

Adds one or more tags to the specified resource. You use tags to add metadata to resources, which you can use to categorize these resources. For example, you can categorize resources by purpose, owner, environment, or team. Each tag consists of a key and a value, which you define. You can add tags to the following AWS Storage Gateway resources:

Storage gateways of all types

Storage Volumes

Virtual Tapes

You can create a maximum of 10 tags for each resource. Virtual tapes and storage volumes that are recovered to a new gateway maintain their tags.

Creating a Request

addTagsToResource

Arguments

:: Text
attrResourceARN

-> AddTagsToResource

Creates a value of AddTagsToResource with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

attrResourceARN - The Amazon Resource Name (ARN) of the resource you want to add tags to.

attrTags - The key-value pair that represents the tag you want to add to the resource. The value can be an empty string.

data AddTagsToResource

AddTagsToResourceInput

See: addTagsToResource smart constructor.

Request Lenses

attrResourceARN :: Lens' AddTagsToResource Text

The Amazon Resource Name (ARN) of the resource you want to add tags to.

attrTags :: Lens' AddTagsToResource [Tag]

The key-value pair that represents the tag you want to add to the resource. The value can be an empty string.

Destructuring the Response

addTagsToResourceResponse

Arguments

:: Int
attrrsResponseStatus

-> AddTagsToResourceResponse

Creates a value of AddTagsToResourceResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

attrrsResourceARN - The Amazon Resource Name (ARN) of the resource you want to add tags to.

attrrsResponseStatus - -- | The response status code.

data AddTagsToResourceResponse

AddTagsToResourceOutput

See: addTagsToResourceResponse smart constructor.

Response Lenses

attrrsResourceARN :: Lens' AddTagsToResourceResponse (Maybe Text)

The Amazon Resource Name (ARN) of the resource you want to add tags to.

attrrsResponseStatus :: Lens' AddTagsToResourceResponse Int

- | The response status code.