Network.AWS.StorageGateway.AddUploadBuffer

Description

This operation configures one or more gateway local disks as upload buffer for a specified gateway. This operation is supported for both the gateway-stored and gateway-cached volume architectures. In the request, you specify the gateway Amazon Resource Name (ARN) to which you want to add upload buffer, and one or more disk IDs that you want to configure as upload buffer.

Request

data AddUploadBuffer

Request constructor

mkAddUploadBuffer

Arguments

:: Text
aubGatewayARN

-> [Text]
aubDiskIds

-> AddUploadBuffer

Smart constructor for the minimum required parameters to construct a valid AddUploadBuffer request.

The fields accessible through corresponding lenses are:

GatewayARN :: Text

DiskIds :: [Text]

Request lenses

aubGatewayARN :: Lens' AddUploadBuffer Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

aubDiskIds :: Lens' AddUploadBuffer [Text]

Response

data AddUploadBufferResponse

Response constructor

mkAddUploadBufferResponse :: AddUploadBufferResponse

Smart constructor for the minimum required parameters to construct a valid AddUploadBufferResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

GatewayARN :: Maybe Text

Response lenses

aubrGatewayARN :: Lens' AddUploadBufferResponse (Maybe Text)

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.