Network.AWS.StorageGateway.AddWorkingStorage

Description

This operation configures one or more gateway local disks as working storage for a gateway. This operation is supported only for the gateway-stored volume architecture. This operation is deprecated method in cached-volumes API version (20120630). Use AddUploadBuffer instead.

Working storage is also referred to as upload buffer. You can also use the AddUploadBuffer operation to add upload buffer to a stored-volume gateway.

In the request, you specify the gateway Amazon Resource Name (ARN) to which you want to add working storage, and one or more disk IDs that you want to configure as working storage.

http://docs.aws.amazon.com/storagegateway/latest/APIReference/API_AddWorkingStorage.html

Request

data AddWorkingStorage

Request constructor

addWorkingStorage

Arguments

:: Text
awsGatewayARN

-> AddWorkingStorage

AddWorkingStorage constructor.

The fields accessible through corresponding lenses are:

awsDiskIds :: [Text]

awsGatewayARN :: Text

Request lenses

awsDiskIds :: Lens' AddWorkingStorage [Text]

An array of strings that identify disks that are to be configured as working storage. Each string have a minimum length of 1 and maximum length of 300. You can get the disk IDs from the ListLocalDisks API.

awsGatewayARN :: Lens' AddWorkingStorage Text

Response

data AddWorkingStorageResponse

Response constructor

addWorkingStorageResponse :: AddWorkingStorageResponse

AddWorkingStorageResponse constructor.

The fields accessible through corresponding lenses are:

awsrGatewayARN :: Maybe Text

Response lenses

awsrGatewayARN :: Lens' AddWorkingStorageResponse (Maybe Text)