Network.AWS.StorageGateway.CancelArchival

Description

Cancels archiving of a virtual tape to the virtual tape shelf (VTS) after the archiving process is initiated. This operation is only supported in the tape gateway architecture.

Creating a Request

cancelArchival

Arguments

:: Text
caGatewayARN

-> Text
caTapeARN

-> CancelArchival

Creates a value of CancelArchival with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

caGatewayARN - Undocumented member.

caTapeARN - The Amazon Resource Name (ARN) of the virtual tape you want to cancel archiving for.

data CancelArchival

CancelArchivalInput

See: cancelArchival smart constructor.

Request Lenses

caGatewayARN :: Lens' CancelArchival Text

Undocumented member.

caTapeARN :: Lens' CancelArchival Text

The Amazon Resource Name (ARN) of the virtual tape you want to cancel archiving for.

Destructuring the Response

cancelArchivalResponse

Arguments

:: Int
carsResponseStatus

-> CancelArchivalResponse

Creates a value of CancelArchivalResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

carsTapeARN - The Amazon Resource Name (ARN) of the virtual tape for which archiving was canceled.

carsResponseStatus - -- | The response status code.

data CancelArchivalResponse

CancelArchivalOutput

See: cancelArchivalResponse smart constructor.

Response Lenses

carsTapeARN :: Lens' CancelArchivalResponse (Maybe Text)

The Amazon Resource Name (ARN) of the virtual tape for which archiving was canceled.

carsResponseStatus :: Lens' CancelArchivalResponse Int

- | The response status code.