Network.AWS.StorageGateway.CancelRetrieval

Request

data CancelRetrieval

Request constructor

cancelRetrieval

Arguments

:: Text
crGatewayARN

-> Text
crTapeARN

-> CancelRetrieval

Smart constructor for the minimum required parameters to construct a valid CancelRetrieval request.

The fields accessible through corresponding lenses are:

GatewayARN :: Text

TapeARN :: Text

Request lenses

crGatewayARN :: Lens' CancelRetrieval Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

crTapeARN :: Lens' CancelRetrieval Text

Response

data CancelRetrievalResponse

Response constructor

cancelRetrievalResponse :: CancelRetrievalResponse

Smart constructor for the minimum required parameters to construct a valid CancelRetrievalResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

TapeARN :: Maybe Text

Response lenses

crrTapeARN :: Lens' CancelRetrievalResponse (Maybe Text)