Network.AWS.StorageGateway.CreateNFSFileShare

Description

Creates a file share on an existing file gateway. In Storage Gateway, a file share is a file system mount point backed by Amazon S3 cloud storage. Storage Gateway exposes file shares using a Network File System (NFS) interface. This operation is only supported in the file gateway architecture.

Important: File gateway requires AWS Security Token Service (AWS STS) to be activated to enable you create a file share. Make sure AWS STS is activated in the region you are creating your file gateway in. If AWS STS is not activated in the region, activate it. For information about how to activate AWS STS, see Activating and Deactivating AWS STS in an AWS Region in the AWS Identity and Access Management User Guide.

File gateway does not support creating hard or symbolic links on a file share.

Creating a Request

createNFSFileShare

Arguments

:: Text
cnfsfsClientToken

-> Text
cnfsfsGatewayARN

-> Text
cnfsfsRole

-> Text
cnfsfsLocationARN

-> CreateNFSFileShare

Creates a value of CreateNFSFileShare with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cnfsfsKMSKey - The KMS key used for Amazon S3 server side encryption. This value can only be set when KmsEncrypted is true. Optional.

cnfsfsKMSEncrypted - True to use Amazon S3 server side encryption with your own AWS KMS key, or false to use a key managed by Amazon S3. Optional.

cnfsfsDefaultStorageClass - The default storage class for objects put into an Amazon S3 bucket by file gateway. Possible values are S3_STANDARD or S3_STANDARD_IA. If this field is not populated, the default value S3_STANDARD is used. Optional.

cnfsfsSquash - Maps a user to anonymous user. Valid options are the following: * RootSquash - Only root is mapped to anonymous user. * NoSquash - No one is mapped to anonymous user. * AllSquash - Everyone is mapped to anonymous user.

cnfsfsNFSFileShareDefaults - File share default values. Optional.

cnfsfsClientList - The list of clients that are allowed to access the file gateway. The list must contain either valid IP addresses or valid CIDR blocks.

cnfsfsReadOnly - Sets the write status of a file share: "true" if the write status is read-only, and otherwise "false".

cnfsfsClientToken - A unique string value that you supply that is used by file gateway to ensure idempotent file share creation.

cnfsfsGatewayARN - The Amazon Resource Name (ARN) of the file gateway on which you want to create a file share.

cnfsfsRole - The ARN of the AWS Identity and Access Management (IAM) role that a file gateway assumes when it accesses the underlying storage.

cnfsfsLocationARN - The ARN of the backed storage used for storing file data.

data CreateNFSFileShare

CreateNFSFileShareInput

See: createNFSFileShare smart constructor.

Request Lenses

cnfsfsKMSKey :: Lens' CreateNFSFileShare (Maybe Text)

The KMS key used for Amazon S3 server side encryption. This value can only be set when KmsEncrypted is true. Optional.

cnfsfsKMSEncrypted :: Lens' CreateNFSFileShare (Maybe Bool)

True to use Amazon S3 server side encryption with your own AWS KMS key, or false to use a key managed by Amazon S3. Optional.

cnfsfsDefaultStorageClass :: Lens' CreateNFSFileShare (Maybe Text)

The default storage class for objects put into an Amazon S3 bucket by file gateway. Possible values are S3_STANDARD or S3_STANDARD_IA. If this field is not populated, the default value S3_STANDARD is used. Optional.

cnfsfsSquash :: Lens' CreateNFSFileShare (Maybe Text)

Maps a user to anonymous user. Valid options are the following: * RootSquash - Only root is mapped to anonymous user. * NoSquash - No one is mapped to anonymous user. * AllSquash - Everyone is mapped to anonymous user.

cnfsfsNFSFileShareDefaults :: Lens' CreateNFSFileShare (Maybe NFSFileShareDefaults)

File share default values. Optional.

cnfsfsClientList :: Lens' CreateNFSFileShare (Maybe (NonEmpty Text))

The list of clients that are allowed to access the file gateway. The list must contain either valid IP addresses or valid CIDR blocks.

cnfsfsReadOnly :: Lens' CreateNFSFileShare (Maybe Bool)

Sets the write status of a file share: "true" if the write status is read-only, and otherwise "false".

cnfsfsClientToken :: Lens' CreateNFSFileShare Text

A unique string value that you supply that is used by file gateway to ensure idempotent file share creation.

cnfsfsGatewayARN :: Lens' CreateNFSFileShare Text

The Amazon Resource Name (ARN) of the file gateway on which you want to create a file share.

cnfsfsRole :: Lens' CreateNFSFileShare Text

The ARN of the AWS Identity and Access Management (IAM) role that a file gateway assumes when it accesses the underlying storage.

cnfsfsLocationARN :: Lens' CreateNFSFileShare Text

The ARN of the backed storage used for storing file data.

Destructuring the Response

createNFSFileShareResponse

Arguments

:: Int
cnfsfsrsResponseStatus

-> CreateNFSFileShareResponse

Creates a value of CreateNFSFileShareResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cnfsfsrsFileShareARN - The Amazon Resource Name (ARN) of the newly created file share.

cnfsfsrsResponseStatus - -- | The response status code.

data CreateNFSFileShareResponse

CreateNFSFileShareOutput

See: createNFSFileShareResponse smart constructor.

Response Lenses

cnfsfsrsFileShareARN :: Lens' CreateNFSFileShareResponse (Maybe Text)

The Amazon Resource Name (ARN) of the newly created file share.

cnfsfsrsResponseStatus :: Lens' CreateNFSFileShareResponse Int

- | The response status code.