Network.AWS.StorageGateway.CreateSnapshot

Description

This operation initiates a snapshot of a volume. AWS Storage Gateway provides the ability to back up point-in-time snapshots of your data to Amazon Simple Storage (S3) for durable off-site recovery, as well as import the data to an Amazon Elastic Block Store (EBS) volume in Amazon Elastic Compute Cloud (EC2). You can take snapshots of your gateway volume on a scheduled or ad-hoc basis. This API enables you to take ad-hoc snapshot. For more information, see <http://docs.aws.amazon.com/storagegateway/latest/userguide/WorkingWithSnapshots.html Working With Snapshots in the AWS Storage Gateway Console>. In the CreateSnapshot request you identify the volume by providing its Amazon Resource Name (ARN). You must also provide description for the snapshot. When AWS Storage Gateway takes the snapshot of specified volume, the snapshot and description appears in the AWS Storage Gateway Console. In response, AWS Storage Gateway returns you a snapshot ID. You can use this snapshot ID to check the snapshot progress or later use it when you want to create a volume from a snapshot.

http://docs.aws.amazon.com/storagegateway/latest/APIReference/API_CreateSnapshot.html

Request

data CreateSnapshot

Request constructor

createSnapshot

Arguments

:: Text
csVolumeARN

-> Text
csSnapshotDescription

-> CreateSnapshot

CreateSnapshot constructor.

The fields accessible through corresponding lenses are:

csSnapshotDescription :: Text

csVolumeARN :: Text

Request lenses

csSnapshotDescription :: Lens' CreateSnapshot Text

Textual description of the snapshot that appears in the Amazon EC2 console, Elastic Block Store snapshots panel in the Description field, and in the AWS Storage Gateway snapshot Details pane, Description field.

csVolumeARN :: Lens' CreateSnapshot Text

The Amazon Resource Name (ARN) of the volume. Use the ListVolumes operation to return a list of gateway volumes.

Response

data CreateSnapshotResponse

Response constructor

createSnapshotResponse :: CreateSnapshotResponse

CreateSnapshotResponse constructor.

The fields accessible through corresponding lenses are:

csrSnapshotId :: Maybe Text

csrVolumeARN :: Maybe Text

Response lenses

csrSnapshotId :: Lens' CreateSnapshotResponse (Maybe Text)

The snapshot ID that is used to refer to the snapshot in future operations such as describing snapshots (Amazon Elastic Compute Cloud API DescribeSnapshots) or creating a volume from a snapshot (CreateStorediSCSIVolume).

csrVolumeARN :: Lens' CreateSnapshotResponse (Maybe Text)

The Amazon Resource Name (ARN) of the volume of which the snapshot was taken.