Network.AWS.StorageGateway.CreateSnapshotFromVolumeRecoveryPoint

Description

This operation initiates a snapshot of a gateway from a volume recovery point. This operation is supported only for the gateway-cached volume architecture (see ).

A volume recovery point is a point in time at which all data of the volume is consistent and from which you can create a snapshot. To get a list of volume recovery point for gateway-cached volumes, use ListVolumeRecoveryPoints .

In the CreateSnapshotFromVolumeRecoveryPoint request, you identify the volume by providing its Amazon Resource Name (ARN). You must also provide a description for the snapshot. When AWS Storage Gateway takes a snapshot of the specified volume, the snapshot and its description appear in the AWS Storage Gateway console. In response, AWS Storage Gateway returns you a snapshot ID. You can use this snapshot ID to check the snapshot progress or later use it when you want to create a volume from a snapshot.

http://docs.aws.amazon.com/storagegateway/latest/APIReference/API_CreateSnapshotFromVolumeRecoveryPoint.html

Request

data CreateSnapshotFromVolumeRecoveryPoint

Request constructor

createSnapshotFromVolumeRecoveryPoint

Arguments

:: Text
csfvrpVolumeARN

-> Text
csfvrpSnapshotDescription

-> CreateSnapshotFromVolumeRecoveryPoint

CreateSnapshotFromVolumeRecoveryPoint constructor.

The fields accessible through corresponding lenses are:

csfvrpSnapshotDescription :: Text

csfvrpVolumeARN :: Text

Request lenses

csfvrpSnapshotDescription :: Lens' CreateSnapshotFromVolumeRecoveryPoint Text

csfvrpVolumeARN :: Lens' CreateSnapshotFromVolumeRecoveryPoint Text

Response

data CreateSnapshotFromVolumeRecoveryPointResponse

Response constructor

createSnapshotFromVolumeRecoveryPointResponse :: CreateSnapshotFromVolumeRecoveryPointResponse

CreateSnapshotFromVolumeRecoveryPointResponse constructor.

The fields accessible through corresponding lenses are:

csfvrprSnapshotId :: Maybe Text

csfvrprVolumeARN :: Maybe Text

csfvrprVolumeRecoveryPointTime :: Maybe Text

Response lenses

csfvrprSnapshotId :: Lens' CreateSnapshotFromVolumeRecoveryPointResponse (Maybe Text)

csfvrprVolumeARN :: Lens' CreateSnapshotFromVolumeRecoveryPointResponse (Maybe Text)

csfvrprVolumeRecoveryPointTime :: Lens' CreateSnapshotFromVolumeRecoveryPointResponse (Maybe Text)