Network.AWS.StorageGateway.CreateStorediSCSIVolume

Description

This operation creates a volume on a specified gateway. This operation is supported only for the gateway-cached volume architecture. The size of the volume to create is inferred from the disk size. You can choose to preserve existing data on the disk, create volume from an existing snapshot, or create an empty volume. If you choose to create an empty gateway volume, then any existing data on the disk is erased. In the request you must specify the gateway and the disk information on which you are creating the volume. In response, AWS Storage Gateway creates the volume and returns volume information such as the volume Amazon Resource Name (ARN), its size, and the iSCSI target ARN that initiators can use to connect to the volume target. Example Request The following example shows a request that specifies that a local disk of a gateway be configured as a volume. POST / HTTP/1.1 Host: storagegateway.us-east-1.amazonaws.com x-amz-Date: 20120425T120000Z Authorization: CSOC7TJPLR0OOKIRLGOHVAICUFVV4KQNSO5AEMVJF66Q9ASUAAJG Content-type: application/x-amz-json-1.1 x-amz-target: StorageGateway_20120630.CreateStorediSCSIVolume { GatewayARN: "arn:aws:storagegateway:us-east-1:111122223333:gateway/mygateway", DiskId: "pci-0000:03:00.0-scsi-0:0:0:0", PreserveExistingData: true, TargetName: "myvolume", NetworkInterfaceId: "10.1.1.1" } HTTP/1.1 200 OK x-amzn-RequestId: CSOC7TJPLR0OOKIRLGOHVAICUFVV4KQNSO5AEMVJF66Q9ASUAAJG Date: Wed, 25 Apr 2012 12:00:02 GMT Content-type: application/x-amz-json-1.1 Content-length: 215 { VolumeARN: "arn:aws:storagegateway:us-east-1:111122223333:gatewaymygatewayvolume/vol-1122AABB", VolumeSizeInBytes: 1099511627776, TargetARN: "arn:aws:storagegateway:us-east-1:111122223333:gatewaymygatewaytarget/iqn.1997-05.com.amazon:myvolume" }.

Request

data CreateStorediSCSIVolume

A JSON object containing one or more of the following fields: CreateStorediSCSIVolumeInput$DiskId CreateStorediSCSIVolumeInput$NetworkInterfaceId CreateStorediSCSIVolumeInput$PreserveExistingData CreateStorediSCSIVolumeInput$SnapshotId CreateStorediSCSIVolumeInput$TargetName.

Request constructor

createStorediSCSIVolume

Arguments

:: Text
csscsivGatewayARN

-> Text
csscsivDiskId

-> Bool
csscsivPreserveExistingData

-> Text
csscsivTargetName

-> Text
csscsivNetworkInterfaceId

-> CreateStorediSCSIVolume

Smart constructor for the minimum required parameters to construct a valid CreateStorediSCSIVolume request.

The fields accessible through corresponding lenses are:

GatewayARN :: Text

DiskId :: Text

SnapshotId :: Maybe Text

PreserveExistingData :: Bool

TargetName :: Text

NetworkInterfaceId :: Text

Request lenses

csscsivGatewayARN :: Lens' CreateStorediSCSIVolume Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

csscsivDiskId :: Lens' CreateStorediSCSIVolume Text

The unique identifier for the gateway local disk that is configured as a stored volume. Use ListLocalDisks to list disk IDs for a gateway.

csscsivSnapshotId :: Lens' CreateStorediSCSIVolume (Maybe Text)

The snapshot ID (e.g. "snap-1122aabb") of the snapshot to restore as the new stored volume. Specify this field if you want to create the iSCSI storage volume from a snapshot otherwise do not include this field. To list snapshots for your account use DescribeSnapshots in the Amazon Elastic Compute Cloud API Reference.

csscsivPreserveExistingData :: Lens' CreateStorediSCSIVolume Bool

Specify this field as true if you want to preserve the data on the local disk. Otherwise, specifying this field as false creates an empty volume. Valid Values: true, false.

csscsivTargetName :: Lens' CreateStorediSCSIVolume Text

The name of the iSCSI target used by initiators to connect to the target and as a suffix for the target ARN. For example, specifying TargetName as myvolume results in the target ARN of arn:aws:storagegateway:us-east-1:111122223333:gatewaymygatewaytarget/iqn.1997-05.com.amazon:myvolume. The target name must be unique across all volumes of a gateway.

csscsivNetworkInterfaceId :: Lens' CreateStorediSCSIVolume Text

The network interface of the gateway on which to expose the iSCSI target. Only IPv4 addresses are accepted. Use DescribeGatewayInformation to get a list of the network interfaces available on a gateway. Valid Values: A valid IP address.

Response

data CreateStorediSCSIVolumeResponse

A JSON object containing the following fields:.

Response constructor

createStorediSCSIVolumeResponse :: CreateStorediSCSIVolumeResponse

Smart constructor for the minimum required parameters to construct a valid CreateStorediSCSIVolumeResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

VolumeARN :: Maybe Text

VolumeSizeInBytes :: Maybe Integer

TargetARN :: Maybe Text

Response lenses

csscsivrVolumeARN :: Lens' CreateStorediSCSIVolumeResponse (Maybe Text)

The Amazon Resource Name (ARN) of the configured volume.

csscsivrVolumeSizeInBytes :: Lens' CreateStorediSCSIVolumeResponse (Maybe Integer)

The size of the volume in bytes.

csscsivrTargetARN :: Lens' CreateStorediSCSIVolumeResponse (Maybe Text)

he Amazon Resource Name (ARN) of the volume target that includes the iSCSI name that initiators can use to connect to the target.