Network.AWS.StorageGateway.CreateTapeWithBarcode

Description

Creates a virtual tape by using your own barcode. You write data to the virtual tape and then archive the tape. A barcode is unique and can not be reused if it has already been used on a tape . This applies to barcodes used on deleted tapes. This operation is only supported in the tape gateway. architecture.

Creating a Request

createTapeWithBarcode

Arguments

:: Text
ctwbGatewayARN

-> Integer
ctwbTapeSizeInBytes

-> Text
ctwbTapeBarcode

-> CreateTapeWithBarcode

Creates a value of CreateTapeWithBarcode with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ctwbGatewayARN - The unique Amazon Resource Name (ARN) that represents the gateway to associate the virtual tape with. Use the ListGateways operation to return a list of gateways for your account and region.

ctwbTapeSizeInBytes - The size, in bytes, of the virtual tape that you want to create.

ctwbTapeBarcode - The barcode that you want to assign to the tape.

data CreateTapeWithBarcode

CreateTapeWithBarcodeInput

See: createTapeWithBarcode smart constructor.

Request Lenses

ctwbGatewayARN :: Lens' CreateTapeWithBarcode Text

The unique Amazon Resource Name (ARN) that represents the gateway to associate the virtual tape with. Use the ListGateways operation to return a list of gateways for your account and region.

ctwbTapeSizeInBytes :: Lens' CreateTapeWithBarcode Integer

The size, in bytes, of the virtual tape that you want to create.

ctwbTapeBarcode :: Lens' CreateTapeWithBarcode Text

The barcode that you want to assign to the tape.

Destructuring the Response

createTapeWithBarcodeResponse

Arguments

:: Int
ctwbrsResponseStatus

-> CreateTapeWithBarcodeResponse

Creates a value of CreateTapeWithBarcodeResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ctwbrsTapeARN - A unique Amazon Resource Name (ARN) that represents the virtual tape that was created.

ctwbrsResponseStatus - -- | The response status code.

data CreateTapeWithBarcodeResponse

CreateTapeOutput

See: createTapeWithBarcodeResponse smart constructor.

Response Lenses

ctwbrsTapeARN :: Lens' CreateTapeWithBarcodeResponse (Maybe Text)

A unique Amazon Resource Name (ARN) that represents the virtual tape that was created.

ctwbrsResponseStatus :: Lens' CreateTapeWithBarcodeResponse Int

- | The response status code.