Network.AWS.StorageGateway.CreateTapes

Description

Creates one or more virtual tapes. You write data to the virtual tapes and then archive the tapes.

Cache storage must be allocated to the gateway before you can create virtual tapes. Use the AddCache operation to add cache storage to a gateway.

http://docs.aws.amazon.com/storagegateway/latest/APIReference/API_CreateTapes.html

Request

data CreateTapes

Request constructor

createTapes

Arguments

:: Text
ctGatewayARN

-> Integer
ctTapeSizeInBytes

-> Text
ctClientToken

-> Natural
ctNumTapesToCreate

-> Text
ctTapeBarcodePrefix

-> CreateTapes

CreateTapes constructor.

The fields accessible through corresponding lenses are:

ctClientToken :: Text

ctGatewayARN :: Text

ctNumTapesToCreate :: Natural

ctTapeBarcodePrefix :: Text

ctTapeSizeInBytes :: Integer

Request lenses

ctClientToken :: Lens' CreateTapes Text

A unique identifier that you use to retry a request. If you retry a request, use the same ClientToken you specified in the initial request.

Using the same ClientToken prevents creating the tape multiple times.

ctGatewayARN :: Lens' CreateTapes Text

The unique Amazon Resource Name(ARN) that represents the gateway to associate the virtual tapes with. Use the ListGateways operation to return a list of gateways for your account and region.

ctNumTapesToCreate :: Lens' CreateTapes Natural

The number of virtual tapes you want to create.

ctTapeBarcodePrefix :: Lens' CreateTapes Text

A prefix you append to the barcode of the virtual tape you are creating. This makes a barcode unique.

The prefix must be 1 to 4 characters in length and must be upper-case letters A-Z.

ctTapeSizeInBytes :: Lens' CreateTapes Integer

The size, in bytes, of the virtual tapes you want to create.

The size must be gigabyte (1024*1024*1024 byte) aligned.

Response

data CreateTapesResponse

Response constructor

createTapesResponse :: CreateTapesResponse

CreateTapesResponse constructor.

The fields accessible through corresponding lenses are:

ctrTapeARNs :: [Text]

Response lenses

ctrTapeARNs :: Lens' CreateTapesResponse [Text]

A list of unique Amazon Resource Named (ARN) the represents the virtual tapes that were created.