Network.AWS.StorageGateway.DeleteGateway

Description

This operation deletes a gateway. To specify which gateway to delete, use the Amazon Resource Name (ARN) of the gateway in your request. The operation deletes the gateway; however, it does not delete the gateway virtual machine (VM) from your host computer.

After you delete a gateway, you cannot reactivate it. Completed snapshots of the gateway volumes are not deleted upon deleting the gateway, however, pending snapshots will not complete. After you delete a gateway, your next step is to remove it from your environment.

You no longer pay software charges after the gateway is deleted; however, your existing Amazon EBS snapshots persist and you will continue to be billed for these snapshots. You can choose to remove all remaining Amazon EBS snapshots by canceling your Amazon EC2 subscription. If you prefer not to cancel your Amazon EC2 subscription, you can delete your snapshots using the Amazon EC2 console. For more information, see the AWS Storage Gateway DetailPage.

http://docs.aws.amazon.com/storagegateway/latest/APIReference/API_DeleteGateway.html

Request

data DeleteGateway

Request constructor

deleteGateway

Arguments

:: Text
dgGatewayARN

-> DeleteGateway

DeleteGateway constructor.

The fields accessible through corresponding lenses are:

dgGatewayARN :: Text

Request lenses

dgGatewayARN :: Lens' DeleteGateway Text

Response

data DeleteGatewayResponse

Response constructor

deleteGatewayResponse :: DeleteGatewayResponse

DeleteGatewayResponse constructor.

The fields accessible through corresponding lenses are:

dgrGatewayARN :: Maybe Text

Response lenses

dgrGatewayARN :: Lens' DeleteGatewayResponse (Maybe Text)