Network.AWS.StorageGateway.DeleteTape

Description

Deletes the specified virtual tape.

http://docs.aws.amazon.com/storagegateway/latest/APIReference/API_DeleteTape.html

Request

data DeleteTape

Request constructor

deleteTape

Arguments

:: Text
dt1GatewayARN

-> Text
dt1TapeARN

-> DeleteTape

DeleteTape constructor.

The fields accessible through corresponding lenses are:

dt1GatewayARN :: Text

dt1TapeARN :: Text

Request lenses

dt1GatewayARN :: Lens' DeleteTape Text

The unique Amazon Resource Name (ARN) of the gateway that the virtual tape to delete is associated with. Use the ListGateways operation to return a list of gateways for your account and region.

dt1TapeARN :: Lens' DeleteTape Text

The Amazon Resource Name (ARN) of the virtual tape to delete.

Response

data DeleteTapeResponse

Response constructor

deleteTapeResponse :: DeleteTapeResponse

DeleteTapeResponse constructor.

The fields accessible through corresponding lenses are:

dtrTapeARN :: Maybe Text

Response lenses

dtrTapeARN :: Lens' DeleteTapeResponse (Maybe Text)

The Amazon Resource Name (ARN) of the deleted virtual tape.