Network.AWS.StorageGateway.DeleteTapeArchive

Description

Deletes the specified virtual tape from the virtual tape shelf (VTS).

http://docs.aws.amazon.com/storagegateway/latest/APIReference/API_DeleteTapeArchive.html

Request

data DeleteTapeArchive

Request constructor

deleteTapeArchive

Arguments

:: Text
dtaTapeARN

-> DeleteTapeArchive

DeleteTapeArchive constructor.

The fields accessible through corresponding lenses are:

dtaTapeARN :: Text

Request lenses

dtaTapeARN :: Lens' DeleteTapeArchive Text

The Amazon Resource Name (ARN) of the virtual tape to delete from the virtual tape shelf (VTS).

Response

data DeleteTapeArchiveResponse

Response constructor

deleteTapeArchiveResponse :: DeleteTapeArchiveResponse

DeleteTapeArchiveResponse constructor.

The fields accessible through corresponding lenses are:

dtarTapeARN :: Maybe Text

Response lenses

dtarTapeARN :: Lens' DeleteTapeArchiveResponse (Maybe Text)

The Amazon Resource Name (ARN) of the virtual tape that was deleted from the virtual tape shelf (VTS).