Network.AWS.StorageGateway.DeleteVolume

Description

This operation delete the specified gateway volume that you previously created using the CreateStorediSCSIVolume API. For gateway-stored volumes, the local disk that was configured as the storage volume is not deleted. You can reuse the local disk to create another storage volume.

Before you delete a gateway volume, make sure there are no iSCSI connections to the volume you are deleting. You should also make sure there is no snapshot in progress. You can use the Amazon Elastic Compute Cloud (Amazon EC2) API to query snapshots on the volume you are deleting and check the snapshot status. For more information, go to DescribeSnapshots in the AmazonElastic Compute Cloud API Reference.

In the request, you must provide the Amazon Resource Name (ARN) of the storage volume you want to delete.

http://docs.aws.amazon.com/storagegateway/latest/APIReference/API_DeleteVolume.html

Request

data DeleteVolume

Request constructor

deleteVolume

Arguments

:: Text
dvVolumeARN

-> DeleteVolume

DeleteVolume constructor.

The fields accessible through corresponding lenses are:

dvVolumeARN :: Text

Request lenses

dvVolumeARN :: Lens' DeleteVolume Text

The Amazon Resource Name (ARN) of the volume. Use the ListVolumes operation to return a list of gateway volumes.

Response

data DeleteVolumeResponse

Response constructor

deleteVolumeResponse :: DeleteVolumeResponse

DeleteVolumeResponse constructor.

The fields accessible through corresponding lenses are:

dvrVolumeARN :: Maybe Text

Response lenses

dvrVolumeARN :: Lens' DeleteVolumeResponse (Maybe Text)

The Amazon Resource Name (ARN) of the storage volume that was deleted. It is the same ARN you provided in the request.