Network.AWS.StorageGateway.DescribeBandwidthRateLimit

Description

This operation returns the bandwidth rate limits of a gateway. By default, these limits are not set, which means no bandwidth rate limiting is in effect. This operation only returns a value for a bandwidth rate limit only if the limit is set. If no limits are set for the gateway, then this operation returns only the gateway ARN in the response body. To specify which gateway to describe, use the Amazon Resource Name (ARN) of the gateway in your request. Example Request The following example shows a request that returns the bandwidth throttle properties of a gateway. POST / HTTP/1.1 Host: storagegateway.us-east-1.amazonaws.com x-amz-Date: 20120425T120000Z Authorization: CSOC7TJPLR0OOKIRLGOHVAICUFVV4KQNSO5AEMVJF66Q9ASUAAJG Content-type: application/x-amz-json-1.1 x-amz-target: StorageGateway_20120630.DescribeBandwidthRateLimit { GatewayARN: "arn:aws:storagegateway:us-east-1:111122223333:gateway/mygate way" } HTTP/1.1 200 OK x-amzn-RequestId: CSOC7TJPLR0OOKIRLGOHVAICUFVV4KQNSO5AEMVJF66Q9ASUAAJG Date: Wed, 25 Apr 2012 12:00:02 GMT Content-type: application/x-amz-json-1.1 Content-length: 169 { GatewayARN: "arn:aws:storagegateway:us-east-1:111122223333:gateway/mygateway", AverageUploadRateLimitInBitsPerSec: 102400, AverageDownloadRateLimitInBitsPerSec: 51200 }.

Request

data DescribeBandwidthRateLimit

A JSON object containing the of the gateway.

Request constructor

describeBandwidthRateLimit

Arguments

:: Text
dbrl1GatewayARN

-> DescribeBandwidthRateLimit

Smart constructor for the minimum required parameters to construct a valid DescribeBandwidthRateLimit request.

The fields accessible through corresponding lenses are:

GatewayARN :: Text

Request lenses

dbrl1GatewayARN :: Lens' DescribeBandwidthRateLimit Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

Response

data DescribeBandwidthRateLimitResponse

A JSON object containing the following fields:.

Response constructor

describeBandwidthRateLimitResponse :: DescribeBandwidthRateLimitResponse

Smart constructor for the minimum required parameters to construct a valid DescribeBandwidthRateLimitResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

GatewayARN :: Maybe Text

AverageUploadRateLimitInBitsPerSec :: Maybe Integer

AverageDownloadRateLimitInBitsPerSec :: Maybe Integer

Response lenses

dbrlrrGatewayARN :: Lens' DescribeBandwidthRateLimitResponse (Maybe Text)

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

dbrlrrAverageUploadRateLimitInBitsPerSec :: Lens' DescribeBandwidthRateLimitResponse (Maybe Integer)

The average upload bandwidth rate limit in bits per second. This field does not appear in the response if the upload rate limit is not set.

dbrlrrAverageDownloadRateLimitInBitsPerSec :: Lens' DescribeBandwidthRateLimitResponse (Maybe Integer)

The average download bandwidth rate limit in bits per second. This field does not appear in the response if the download rate limit is not set.