Network.AWS.StorageGateway.DescribeCache

Description

This operation returns information about the cache of a gateway. This operation is supported only for the gateway-cached volume architecture. The response includes disk IDs that are configured as cache, and it includes the amount of cache allocated and used. Example Request The following example shows a request to obtain a description of a gateway's working storage. POST HTTP1.1 Host: storagegateway.us-east-1.amazonaws.com Content-Type: application/x-amz-json-1.1 Authorization: AWS4-HMAC-SHA256 Credential=AKIAIOSFODNN7EXAMPLE20120425us-east-1storagegatewayaws4_request, SignedHeaders=content-type;host;x-amz-date;x-amz-target, Signature=9cd5a3584d1d67d57e61f120f35102d6b3649066abdd4bf4bbcf05bd9f2f8fe2 x-amz-date: 20120912T120000Z x-amz-target: StorageGateway_20120630.DescribeCache { GatewayARN:"arn:aws:storagegateway:us-east-1:111122223333:gateway/mygateway" } HTTP/1.1 200 OK x-amzn-RequestId: gur28r2rqlgb8vvs0mq17hlgij1q8glle1qeu3kpgg6f0kstauu0 Date: Wed, 12 Sep 2012 12:00:02 GMT Content-Type: application/x-amz-json-1.1 Content-length: 271 { CacheAllocationInBytes: 2199023255552, CacheDirtyPercentage: 0.07, CacheHitPercentage: 99.68, CacheMissPercentage: 0.32, CacheUsedPercentage: 0.07, DiskIds: [ "pci-0000:03:00.0-scsi-0:0:0:0", "pci-0000:04:00.0-scsi-0:1:0:0" ], GatewayARN: "arn:aws:storagegateway:us-east-1:111122223333:gateway/mygateway" }.

Request

data DescribeCache

Request constructor

describeCache

Arguments

:: Text
dcGatewayARN

-> DescribeCache

Smart constructor for the minimum required parameters to construct a valid DescribeCache request.

The fields accessible through corresponding lenses are:

GatewayARN :: Text

Request lenses

dcGatewayARN :: Lens' DescribeCache Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

Response

data DescribeCacheResponse

Response constructor

describeCacheResponse :: DescribeCacheResponse

Smart constructor for the minimum required parameters to construct a valid DescribeCacheResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

GatewayARN :: Maybe Text

DiskIds :: [Text]

CacheAllocatedInBytes :: Maybe Integer

CacheUsedPercentage :: Maybe Double

CacheDirtyPercentage :: Maybe Double

CacheHitPercentage :: Maybe Double

CacheMissPercentage :: Maybe Double

Response lenses

dcrGatewayARN :: Lens' DescribeCacheResponse (Maybe Text)

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

dcrDiskIds :: Lens' DescribeCacheResponse [Text]

dcrCacheAllocatedInBytes :: Lens' DescribeCacheResponse (Maybe Integer)

dcrCacheUsedPercentage :: Lens' DescribeCacheResponse (Maybe Double)

dcrCacheDirtyPercentage :: Lens' DescribeCacheResponse (Maybe Double)

dcrCacheHitPercentage :: Lens' DescribeCacheResponse (Maybe Double)

dcrCacheMissPercentage :: Lens' DescribeCacheResponse (Maybe Double)