Network.AWS.StorageGateway.DescribeCachediSCSIVolumes

Description

This operation returns a description of the gateway volumes specified in the request. This operation is supported only for the gateway-cached volume architecture.

The list of gateway volumes in the request must be from one gateway. In the response Amazon Storage Gateway returns volume information sorted by volume Amazon Resource Name (ARN).

http://docs.aws.amazon.com/storagegateway/latest/APIReference/API_DescribeCachediSCSIVolumes.html

Request

data DescribeCachediSCSIVolumes

Request constructor

describeCachediSCSIVolumes :: DescribeCachediSCSIVolumes

DescribeCachediSCSIVolumes constructor.

The fields accessible through corresponding lenses are:

dcscsivVolumeARNs :: [Text]

Request lenses

dcscsivVolumeARNs :: Lens' DescribeCachediSCSIVolumes [Text]

Response

data DescribeCachediSCSIVolumesResponse

Response constructor

describeCachediSCSIVolumesResponse :: DescribeCachediSCSIVolumesResponse

DescribeCachediSCSIVolumesResponse constructor.

The fields accessible through corresponding lenses are:

dcscsivrCachediSCSIVolumes :: [CachediSCSIVolume]

Response lenses

dcscsivrCachediSCSIVolumes :: Lens' DescribeCachediSCSIVolumesResponse [CachediSCSIVolume]

An array of objects where each object contains metadata about one cached volume.