Network.AWS.StorageGateway.DescribeChapCredentials

Description

This operation returns an array of Challenge-Handshake Authentication Protocol (CHAP) credentials information for a specified iSCSI target, one for each target-initiator pair.

http://docs.aws.amazon.com/storagegateway/latest/APIReference/API_DescribeChapCredentials.html

Request

data DescribeChapCredentials

Request constructor

describeChapCredentials

Arguments

:: Text
dccTargetARN

-> DescribeChapCredentials

DescribeChapCredentials constructor.

The fields accessible through corresponding lenses are:

dccTargetARN :: Text

Request lenses

dccTargetARN :: Lens' DescribeChapCredentials Text

The Amazon Resource Name (ARN) of the iSCSI volume target. Use the DescribeStorediSCSIVolumes operation to return to retrieve the TargetARN for specified VolumeARN.

Response

data DescribeChapCredentialsResponse

Response constructor

describeChapCredentialsResponse :: DescribeChapCredentialsResponse

DescribeChapCredentialsResponse constructor.

The fields accessible through corresponding lenses are:

dccrChapCredentials :: [ChapInfo]

Response lenses

dccrChapCredentials :: Lens' DescribeChapCredentialsResponse [ChapInfo]

An array of ChapInfo objects that represent CHAP credentials. Each object in the array contains CHAP credential information for one target-initiator pair. If no CHAP credentials are set, an empty array is returned. CHAP credential information is provided in a JSON object with the following fields:

InitiatorName: The iSCSI initiator that connects to the target.

SecretToAuthenticateInitiator: The secret key that the initiator (e.g. Windows client) must provide to participate in mutual CHAP with the target.

SecretToAuthenticateTarget: The secret key that the target must provide to participate in mutual CHAP with the initiator (e.g. Windows client).

TargetARN: The Amazon Resource Name (ARN) of the storage volume.