Network.AWS.StorageGateway.DescribeGatewayInformation

Description

This operation returns metadata about a gateway such as its name, network interfaces, configured time zone, and the state (whether the gateway is running or not). To specify which gateway to describe, use the Amazon Resource Name (ARN) of the gateway in your request.

http://docs.aws.amazon.com/storagegateway/latest/APIReference/API_DescribeGatewayInformation.html

Request

data DescribeGatewayInformation

Request constructor

describeGatewayInformation

Arguments

:: Text
dgiGatewayARN

-> DescribeGatewayInformation

DescribeGatewayInformation constructor.

The fields accessible through corresponding lenses are:

dgiGatewayARN :: Text

Request lenses

dgiGatewayARN :: Lens' DescribeGatewayInformation Text

Response

data DescribeGatewayInformationResponse

Response constructor

describeGatewayInformationResponse :: DescribeGatewayInformationResponse

DescribeGatewayInformationResponse constructor.

The fields accessible through corresponding lenses are:

dgirGatewayARN :: Maybe Text

dgirGatewayId :: Maybe Text

dgirGatewayNetworkInterfaces :: [NetworkInterface]

dgirGatewayState :: Maybe Text

dgirGatewayTimezone :: Maybe Text

dgirGatewayType :: Maybe Text

dgirNextUpdateAvailabilityDate :: Maybe Text

Response lenses

dgirGatewayARN :: Lens' DescribeGatewayInformationResponse (Maybe Text)

dgirGatewayId :: Lens' DescribeGatewayInformationResponse (Maybe Text)

The gateway ID.

dgirGatewayNetworkInterfaces :: Lens' DescribeGatewayInformationResponse [NetworkInterface]

A NetworkInterface array that contains descriptions of the gateway network interfaces.

dgirGatewayState :: Lens' DescribeGatewayInformationResponse (Maybe Text)

One of the values that indicates the operating state of the gateway.

dgirGatewayTimezone :: Lens' DescribeGatewayInformationResponse (Maybe Text)

One of the values that indicates the time zone configured for the gateway.

dgirGatewayType :: Lens' DescribeGatewayInformationResponse (Maybe Text)

TBD

dgirNextUpdateAvailabilityDate :: Lens' DescribeGatewayInformationResponse (Maybe Text)

The date at which an update to the gateway is available. This date is in the time zone of the gateway. If the gateway is not available for an update this field is not returned in the response.