Network.AWS.StorageGateway.DescribeMaintenanceStartTime

Description

This operation returns your gateway's weekly maintenance start time including the day and time of the week. Note that values are in terms of the gateway's time zone.

http://docs.aws.amazon.com/storagegateway/latest/APIReference/API_DescribeMaintenanceStartTime.html

Request

data DescribeMaintenanceStartTime

Request constructor

describeMaintenanceStartTime

Arguments

:: Text
dmstGatewayARN

-> DescribeMaintenanceStartTime

DescribeMaintenanceStartTime constructor.

The fields accessible through corresponding lenses are:

dmstGatewayARN :: Text

Request lenses

dmstGatewayARN :: Lens' DescribeMaintenanceStartTime Text

Response

data DescribeMaintenanceStartTimeResponse

Response constructor

describeMaintenanceStartTimeResponse :: DescribeMaintenanceStartTimeResponse

DescribeMaintenanceStartTimeResponse constructor.

The fields accessible through corresponding lenses are:

dmstrDayOfWeek :: Maybe Natural

dmstrGatewayARN :: Maybe Text

dmstrHourOfDay :: Maybe Natural

dmstrMinuteOfHour :: Maybe Natural

dmstrTimezone :: Maybe Text

Response lenses

dmstrDayOfWeek :: Lens' DescribeMaintenanceStartTimeResponse (Maybe Natural)

dmstrGatewayARN :: Lens' DescribeMaintenanceStartTimeResponse (Maybe Text)

dmstrHourOfDay :: Lens' DescribeMaintenanceStartTimeResponse (Maybe Natural)

dmstrMinuteOfHour :: Lens' DescribeMaintenanceStartTimeResponse (Maybe Natural)

dmstrTimezone :: Lens' DescribeMaintenanceStartTimeResponse (Maybe Text)