Network.AWS.StorageGateway.DescribeSnapshotSchedule

Description

This operation describes the snapshot schedule for the specified gateway volume. The snapshot schedule information includes intervals at which snapshots are automatically initiated on the volume. Example Request The following example shows a request that retrieves the snapshot schedule for a volume. POST HTTP1.1 Host: storagegateway.us-east-1.amazonaws.com x-amz-Date: 20120425T120000Z Authorization: CSOC7TJPLR0OOKIRLGOHVAICUFVV4KQNSO5AEMVJF66Q9ASUAAJG Content-type: application/x-amz-json-1.1 x-amz-target: StorageGateway_20120630.DescribeSnapshotSchedule { VolumeARN: "arn:aws:storagegateway:us-east-1:111122223333:gatewaymygatewayvolume/vol-1122AABB" } HTTP/1.1 200 OK x-amzn-RequestId: CSOC7TJPLR0OOKIRLGOHVAICUFVV4KQNSO5AEMVJF66Q9ASUAAJG Date: Wed, 25 Apr 2012 12:00:02 GMT Content-type: application/x-amz-json-1.1 Content-length: 211 { VolumeARN: "arn:aws:storagegateway:us-east-1:111122223333:gatewaymygatewayvolume/vol-1122AABB", StartAt: 6, RecurrenceInHours: 24, Description: "sgw-AABB1122:vol-AABB1122:Schedule", Timezone: "GMT+7:00" }.

Request

data DescribeSnapshotSchedule

A JSON object containing the DescribeSnapshotScheduleInput$VolumeARN of the volume.

Request constructor

describeSnapshotSchedule

Arguments

:: Text
dss1VolumeARN

-> DescribeSnapshotSchedule

Smart constructor for the minimum required parameters to construct a valid DescribeSnapshotSchedule request.

The fields accessible through corresponding lenses are:

VolumeARN :: Text

Request lenses

dss1VolumeARN :: Lens' DescribeSnapshotSchedule Text

The Amazon Resource Name (ARN) of the volume. Use the ListVolumes operation to return a list of gateway volumes.

Response

data DescribeSnapshotScheduleResponse

Response constructor

describeSnapshotScheduleResponse :: DescribeSnapshotScheduleResponse

Smart constructor for the minimum required parameters to construct a valid DescribeSnapshotScheduleResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

VolumeARN :: Maybe Text

StartAt :: Maybe Integer

RecurrenceInHours :: Maybe Integer

Description :: Maybe Text

Timezone :: Maybe Text

Response lenses

dssrrVolumeARN :: Lens' DescribeSnapshotScheduleResponse (Maybe Text)

dssrrStartAt :: Lens' DescribeSnapshotScheduleResponse (Maybe Integer)

dssrrRecurrenceInHours :: Lens' DescribeSnapshotScheduleResponse (Maybe Integer)

dssrrDescription :: Lens' DescribeSnapshotScheduleResponse (Maybe Text)

dssrrTimezone :: Lens' DescribeSnapshotScheduleResponse (Maybe Text)