Network.AWS.StorageGateway.DescribeStorediSCSIVolumes

Description

This operation returns description of the gateway volumes specified in the request. The list of gateway volumes in the request must be from one gateway. In the response Amazon Storage Gateway returns volume information sorted by volume ARNs.

http://docs.aws.amazon.com/storagegateway/latest/APIReference/API_DescribeStorediSCSIVolumes.html

Request

data DescribeStorediSCSIVolumes

Request constructor

describeStorediSCSIVolumes :: DescribeStorediSCSIVolumes

DescribeStorediSCSIVolumes constructor.

The fields accessible through corresponding lenses are:

dsscsivVolumeARNs :: [Text]

Request lenses

dsscsivVolumeARNs :: Lens' DescribeStorediSCSIVolumes [Text]

An array of strings where each string represents the Amazon Resource Name (ARN) of a stored volume. All of the specified stored volumes must from the same gateway. Use ListVolumes to get volume ARNs for a gateway.

Response

data DescribeStorediSCSIVolumesResponse

Response constructor

describeStorediSCSIVolumesResponse :: DescribeStorediSCSIVolumesResponse

DescribeStorediSCSIVolumesResponse constructor.

The fields accessible through corresponding lenses are:

dsscsivrStorediSCSIVolumes :: [StorediSCSIVolume]

Response lenses

dsscsivrStorediSCSIVolumes :: Lens' DescribeStorediSCSIVolumesResponse [StorediSCSIVolume]