Network.AWS.StorageGateway.DescribeTapeArchives

Description

Returns a description of specified virtual tapes in the virtual tape shelf (VTS).

If a specific TapeARN is not specified, AWS Storage Gateway returns a description of all virtual tapes found in the VTS associated with your account.

http://docs.aws.amazon.com/storagegateway/latest/APIReference/API_DescribeTapeArchives.html

Request

data DescribeTapeArchives

Request constructor

describeTapeArchives :: DescribeTapeArchives

DescribeTapeArchives constructor.

The fields accessible through corresponding lenses are:

dtaLimit :: Maybe Natural

dtaMarker :: Maybe Text

dtaTapeARNs :: [Text]

Request lenses

dtaLimit :: Lens' DescribeTapeArchives (Maybe Natural)

Specifies that the number of virtual tapes descried be limited to the specified number.

dtaMarker :: Lens' DescribeTapeArchives (Maybe Text)

An opaque string that indicates the position at which to begin describing virtual tapes.

dtaTapeARNs :: Lens' DescribeTapeArchives [Text]

Specifies one or more unique Amazon Resource Names (ARNs) that represent the virtual tapes you want to describe.

Response

data DescribeTapeArchivesResponse

Response constructor

describeTapeArchivesResponse :: DescribeTapeArchivesResponse

DescribeTapeArchivesResponse constructor.

The fields accessible through corresponding lenses are:

dtarMarker :: Maybe Text

dtarTapeArchives :: [TapeArchive]

Response lenses

dtarMarker :: Lens' DescribeTapeArchivesResponse (Maybe Text)

An opaque string that indicates the position at which the virtual tapes that were fetched for description ended. Use this marker in your next request to fetch the next set of virtual tapes in the virtual tape shelf (VTS). If there are no more virtual tapes to describe, this field does not appear in the response.

dtarTapeArchives :: Lens' DescribeTapeArchivesResponse [TapeArchive]

An array of virtual tape objects in the virtual tape shelf (VTS). The description includes of the Amazon Resource Name(ARN) of the virtual tapes. The information returned includes the Amazon Resource Names (ARNs) of the tapes, size of the tapes, status of the tapes, progress of the description and tape barcode.