Network.AWS.StorageGateway.DescribeTapeRecoveryPoints

Request

data DescribeTapeRecoveryPoints

Request constructor

describeTapeRecoveryPoints

Arguments

:: Text
dtrpGatewayARN

-> DescribeTapeRecoveryPoints

Smart constructor for the minimum required parameters to construct a valid DescribeTapeRecoveryPoints request.

The fields accessible through corresponding lenses are:

GatewayARN :: Text

Marker :: Maybe Text

Limit :: Maybe Integer

Request lenses

dtrpGatewayARN :: Lens' DescribeTapeRecoveryPoints Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

dtrpMarker :: Lens' DescribeTapeRecoveryPoints (Maybe Text)

dtrpLimit :: Lens' DescribeTapeRecoveryPoints (Maybe Integer)

Response

data DescribeTapeRecoveryPointsResponse

Response constructor

describeTapeRecoveryPointsResponse :: DescribeTapeRecoveryPointsResponse

Smart constructor for the minimum required parameters to construct a valid DescribeTapeRecoveryPointsResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

GatewayARN :: Maybe Text

TapeRecoveryPointInfos :: [TapeRecoveryPointInfo]

Marker :: Maybe Text

Response lenses

dtrprGatewayARN :: Lens' DescribeTapeRecoveryPointsResponse (Maybe Text)

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

dtrprTapeRecoveryPointInfos :: Lens' DescribeTapeRecoveryPointsResponse [TapeRecoveryPointInfo]

dtrprMarker :: Lens' DescribeTapeRecoveryPointsResponse (Maybe Text)