Network.AWS.StorageGateway.DescribeTapes

Request

data DescribeTapes

Request constructor

describeTapes

Arguments

:: Text
dt1GatewayARN

-> DescribeTapes

Smart constructor for the minimum required parameters to construct a valid DescribeTapes request.

The fields accessible through corresponding lenses are:

GatewayARN :: Text

TapeARNs :: [Text]

Marker :: Maybe Text

Limit :: Maybe Integer

Request lenses

dt1GatewayARN :: Lens' DescribeTapes Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

dt1TapeARNs :: Lens' DescribeTapes [Text]

dt1Marker :: Lens' DescribeTapes (Maybe Text)

dt1Limit :: Lens' DescribeTapes (Maybe Integer)

Response

data DescribeTapesResponse

Response constructor

describeTapesResponse :: DescribeTapesResponse

Smart constructor for the minimum required parameters to construct a valid DescribeTapesResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Tapes :: [Tape]

Marker :: Maybe Text

Response lenses

dtrrTapes :: Lens' DescribeTapesResponse [Tape]

dtrrMarker :: Lens' DescribeTapesResponse (Maybe Text)