Network.AWS.StorageGateway.DescribeUploadBuffer

Description

This operation returns information about the upload buffer of a gateway. This operation is supported for both the gateway-stored and gateway-cached volume architectures.

The response includes disk IDs that are configured as upload buffer space, and it includes the amount of upload buffer space allocated and used.

http://docs.aws.amazon.com/storagegateway/latest/APIReference/API_DescribeUploadBuffer.html

Request

data DescribeUploadBuffer

Request constructor

describeUploadBuffer

Arguments

:: Text
dubGatewayARN

-> DescribeUploadBuffer

DescribeUploadBuffer constructor.

The fields accessible through corresponding lenses are:

dubGatewayARN :: Text

Request lenses

dubGatewayARN :: Lens' DescribeUploadBuffer Text

Response

data DescribeUploadBufferResponse

Response constructor

describeUploadBufferResponse :: DescribeUploadBufferResponse

DescribeUploadBufferResponse constructor.

The fields accessible through corresponding lenses are:

dubrDiskIds :: [Text]

dubrGatewayARN :: Maybe Text

dubrUploadBufferAllocatedInBytes :: Maybe Integer

dubrUploadBufferUsedInBytes :: Maybe Integer

Response lenses

dubrDiskIds :: Lens' DescribeUploadBufferResponse [Text]

dubrGatewayARN :: Lens' DescribeUploadBufferResponse (Maybe Text)

dubrUploadBufferAllocatedInBytes :: Lens' DescribeUploadBufferResponse (Maybe Integer)

dubrUploadBufferUsedInBytes :: Lens' DescribeUploadBufferResponse (Maybe Integer)