Network.AWS.StorageGateway.DescribeVTLDevices

Description

Returns a description of virtual tape library (VTL) devices for the specified gateway. In the response, AWS Storage Gateway returns VTL device information.

The list of VTL devices must be from one gateway.

http://docs.aws.amazon.com/storagegateway/latest/APIReference/API_DescribeVTLDevices.html

Request

data DescribeVTLDevices

Request constructor

describeVTLDevices

Arguments

:: Text
dvtldGatewayARN

-> DescribeVTLDevices

DescribeVTLDevices constructor.

The fields accessible through corresponding lenses are:

dvtldGatewayARN :: Text

dvtldLimit :: Maybe Natural

dvtldMarker :: Maybe Text

dvtldVTLDeviceARNs :: [Text]

Request lenses

dvtldGatewayARN :: Lens' DescribeVTLDevices Text

dvtldLimit :: Lens' DescribeVTLDevices (Maybe Natural)

Specifies that the number of VTL devices described be limited to the specified number.

dvtldMarker :: Lens' DescribeVTLDevices (Maybe Text)

An opaque string that indicates the position at which to begin describing the VTL devices.

dvtldVTLDeviceARNs :: Lens' DescribeVTLDevices [Text]

An array of strings, where each string represents the Amazon Resource Name (ARN) of a VTL device.

All of the specified VTL devices must be from the same gateway. If no VTL devices are specified, the result will contain all devices on the specified gateway.

Response

data DescribeVTLDevicesResponse

Response constructor

describeVTLDevicesResponse :: DescribeVTLDevicesResponse

DescribeVTLDevicesResponse constructor.

The fields accessible through corresponding lenses are:

dvtldrGatewayARN :: Maybe Text

dvtldrMarker :: Maybe Text

dvtldrVTLDevices :: [VTLDevice]

Response lenses

dvtldrGatewayARN :: Lens' DescribeVTLDevicesResponse (Maybe Text)

dvtldrMarker :: Lens' DescribeVTLDevicesResponse (Maybe Text)

An opaque string that indicates the position at which the VTL devices that were fetched for description ended. Use the marker in your next request to fetch the next set of VTL devices in the list. If there are no more VTL devices to describe, this field does not appear in the response.

dvtldrVTLDevices :: Lens' DescribeVTLDevicesResponse [VTLDevice]

An array of VTL device objects composed of the Amazon Resource Name(ARN) of the VTL devices.