Network.AWS.StorageGateway.ListFileShares

Description

Gets a list of the file shares for a specific file gateway, or the list of file shares that belong to the calling user account. This operation is only supported in the file gateway architecture.

Creating a Request

listFileShares :: ListFileShares

Creates a value of ListFileShares with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lfsGatewayARN - The Amazon resource Name (ARN) of the gateway whose file shares you want to list. If this field is not present, all file shares under your account are listed.

lfsMarker - Opaque pagination token returned from a previous ListFileShares operation. If present, Marker specifies where to continue the list from after a previous call to ListFileShares. Optional.

lfsLimit - The maximum number of file shares to return in the response. The value must be an integer with a value greater than zero. Optional.

data ListFileShares

ListFileShareInput

See: listFileShares smart constructor.

Request Lenses

lfsGatewayARN :: Lens' ListFileShares (Maybe Text)

The Amazon resource Name (ARN) of the gateway whose file shares you want to list. If this field is not present, all file shares under your account are listed.

lfsMarker :: Lens' ListFileShares (Maybe Text)

Opaque pagination token returned from a previous ListFileShares operation. If present, Marker specifies where to continue the list from after a previous call to ListFileShares. Optional.

lfsLimit :: Lens' ListFileShares (Maybe Natural)

The maximum number of file shares to return in the response. The value must be an integer with a value greater than zero. Optional.

Destructuring the Response

listFileSharesResponse

Arguments

:: Int
lfsrsResponseStatus

-> ListFileSharesResponse

Creates a value of ListFileSharesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lfsrsFileShareInfoList - An array of information about the file gateway's file shares.

lfsrsMarker - If the request includes Marker , the response returns that value in this field.

lfsrsNextMarker - If a value is present, there are more file shares to return. In a subsequent request, use NextMarker as the value for Marker to retrieve the next set of file shares.

lfsrsResponseStatus - -- | The response status code.

data ListFileSharesResponse

ListFileShareOutput

See: listFileSharesResponse smart constructor.

Response Lenses

lfsrsFileShareInfoList :: Lens' ListFileSharesResponse [FileShareInfo]

An array of information about the file gateway's file shares.

lfsrsMarker :: Lens' ListFileSharesResponse (Maybe Text)

If the request includes Marker , the response returns that value in this field.

lfsrsNextMarker :: Lens' ListFileSharesResponse (Maybe Text)

If a value is present, there are more file shares to return. In a subsequent request, use NextMarker as the value for Marker to retrieve the next set of file shares.

lfsrsResponseStatus :: Lens' ListFileSharesResponse Int

- | The response status code.