Network.AWS.StorageGateway.ListGateways

Description

Lists gateways owned by an AWS account in a region specified in the request. The returned list is ordered by gateway Amazon Resource Name (ARN).

By default, the operation returns a maximum of 100 gateways. This operation supports pagination that allows you to optionally reduce the number of gateways returned in a response.

If you have more gateways than are returned in a response (that is, the response returns only a truncated list of your gateways), the response contains a marker that you can specify in your next request to fetch the next page of gateways.

This operation returns paginated results.

Creating a Request

listGateways :: ListGateways

Creates a value of ListGateways with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lgMarker - An opaque string that indicates the position at which to begin the returned list of gateways.

lgLimit - Specifies that the list of gateways returned be limited to the specified number of items.

data ListGateways

A JSON object containing zero or more of the following fields:

'ListGatewaysInput$Limit'

'ListGatewaysInput$Marker'

See: listGateways smart constructor.

Request Lenses

lgMarker :: Lens' ListGateways (Maybe Text)

An opaque string that indicates the position at which to begin the returned list of gateways.

lgLimit :: Lens' ListGateways (Maybe Natural)

Specifies that the list of gateways returned be limited to the specified number of items.

Destructuring the Response

listGatewaysResponse

Arguments

:: Int
lgrsResponseStatus

-> ListGatewaysResponse

Creates a value of ListGatewaysResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lgrsMarker - Undocumented member.

lgrsGateways - Undocumented member.

lgrsResponseStatus - -- | The response status code.

data ListGatewaysResponse

See: listGatewaysResponse smart constructor.

Response Lenses

lgrsMarker :: Lens' ListGatewaysResponse (Maybe Text)

Undocumented member.

lgrsGateways :: Lens' ListGatewaysResponse [GatewayInfo]

Undocumented member.

lgrsResponseStatus :: Lens' ListGatewaysResponse Int

- | The response status code.