Network.AWS.StorageGateway.ListLocalDisks

Description

This operation returns a list of the local disks of a gateway. To specify which gateway to describe you use the Amazon Resource Name (ARN) of the gateway in the body of the request. The request returns all disks, specifying which are configured as working storage, stored volume or not configured at all.

http://docs.aws.amazon.com/storagegateway/latest/APIReference/API_ListLocalDisks.html

Request

data ListLocalDisks

Request constructor

listLocalDisks

Arguments

:: Text
lldGatewayARN

-> ListLocalDisks

ListLocalDisks constructor.

The fields accessible through corresponding lenses are:

lldGatewayARN :: Text

Request lenses

lldGatewayARN :: Lens' ListLocalDisks Text

Response

data ListLocalDisksResponse

Response constructor

listLocalDisksResponse :: ListLocalDisksResponse

ListLocalDisksResponse constructor.

The fields accessible through corresponding lenses are:

lldrDisks :: [Disk]

lldrGatewayARN :: Maybe Text

Response lenses

lldrDisks :: Lens' ListLocalDisksResponse [Disk]

lldrGatewayARN :: Lens' ListLocalDisksResponse (Maybe Text)