Network.AWS.StorageGateway.ListTapes

Description

Lists virtual tapes in your virtual tape library (VTL) and your virtual tape shelf (VTS). You specify the tapes to list by specifying one or more tape Amazon Resource Names (ARNs). If you don't specify a tape ARN, the operation lists all virtual tapes in both your VTL and VTS.

This operation supports pagination. By default, the operation returns a maximum of up to 100 tapes. You can optionally specify the Limit parameter in the body to limit the number of tapes in the response. If the number of tapes returned in the response is truncated, the response includes a Marker element that you can use in your subsequent request to retrieve the next set of tapes. This operation is only supported in the tape gateway architecture.

Creating a Request

listTapes :: ListTapes

Creates a value of ListTapes with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ltMarker - A string that indicates the position at which to begin the returned list of tapes.

ltLimit - An optional number limit for the tapes in the list returned by this call.

ltTapeARNs - Undocumented member.

data ListTapes

A JSON object that contains one or more of the following fields:

'ListTapesInput$Limit'

'ListTapesInput$Marker'

'ListTapesInput$TapeARNs'

See: listTapes smart constructor.

Request Lenses

ltMarker :: Lens' ListTapes (Maybe Text)

A string that indicates the position at which to begin the returned list of tapes.

ltLimit :: Lens' ListTapes (Maybe Natural)

An optional number limit for the tapes in the list returned by this call.

ltTapeARNs :: Lens' ListTapes [Text]

Undocumented member.

Destructuring the Response

listTapesResponse

Arguments

:: Int
ltrsResponseStatus

-> ListTapesResponse

Creates a value of ListTapesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ltrsMarker - A string that indicates the position at which to begin returning the next list of tapes. Use the marker in your next request to continue pagination of tapes. If there are no more tapes to list, this element does not appear in the response body.

ltrsTapeInfos - Undocumented member.

ltrsResponseStatus - -- | The response status code.

data ListTapesResponse

A JSON object containing the following fields:

'ListTapesOutput$Marker'

'ListTapesOutput$VolumeInfos'

See: listTapesResponse smart constructor.

Response Lenses

ltrsMarker :: Lens' ListTapesResponse (Maybe Text)

A string that indicates the position at which to begin returning the next list of tapes. Use the marker in your next request to continue pagination of tapes. If there are no more tapes to list, this element does not appear in the response body.

ltrsTapeInfos :: Lens' ListTapesResponse [TapeInfo]

Undocumented member.

ltrsResponseStatus :: Lens' ListTapesResponse Int

- | The response status code.