Network.AWS.StorageGateway.ListVolumeRecoveryPoints

Description

This operation lists the recovery points for a specified gateway. This operation is supported only for the gateway-cached volume architecture.

Each gateway-cached volume has one recovery point. A volume recovery point is a point in time at which all data of the volume is consistent and from which you can create a snapshot. To create a snapshot from a volume recovery point use the CreateSnapshotFromVolumeRecoveryPoint operation.

http://docs.aws.amazon.com/storagegateway/latest/APIReference/API_ListVolumeRecoveryPoints.html

Request

data ListVolumeRecoveryPoints

Request constructor

listVolumeRecoveryPoints

Arguments

:: Text
lvrpGatewayARN

-> ListVolumeRecoveryPoints

ListVolumeRecoveryPoints constructor.

The fields accessible through corresponding lenses are:

lvrpGatewayARN :: Text

Request lenses

lvrpGatewayARN :: Lens' ListVolumeRecoveryPoints Text

Response

data ListVolumeRecoveryPointsResponse

Response constructor

listVolumeRecoveryPointsResponse :: ListVolumeRecoveryPointsResponse

ListVolumeRecoveryPointsResponse constructor.

The fields accessible through corresponding lenses are:

lvrprGatewayARN :: Maybe Text

lvrprVolumeRecoveryPointInfos :: [VolumeRecoveryPointInfo]

Response lenses

lvrprGatewayARN :: Lens' ListVolumeRecoveryPointsResponse (Maybe Text)

lvrprVolumeRecoveryPointInfos :: Lens' ListVolumeRecoveryPointsResponse [VolumeRecoveryPointInfo]