Network.AWS.StorageGateway.ListVolumes

Description

This operation lists the iSCSI stored volumes of a gateway. Results are sorted by volume ARN. The response includes only the volume ARNs. If you want additional volume information, use the DescribeStorediSCSIVolumes API.

The operation supports pagination. By default, the operation returns a maximum of up to 100 volumes. You can optionally specify the Limit field in the body to limit the number of volumes in the response. If the number of volumes returned in the response is truncated, the response includes a Marker field. You can use this Marker value in your subsequent request to retrieve the next set of volumes.

http://docs.aws.amazon.com/storagegateway/latest/APIReference/API_ListVolumes.html

Request

data ListVolumes

Request constructor

listVolumes

Arguments

:: Text
lvGatewayARN

-> ListVolumes

ListVolumes constructor.

The fields accessible through corresponding lenses are:

lvGatewayARN :: Text

lvLimit :: Maybe Natural

lvMarker :: Maybe Text

Request lenses

lvGatewayARN :: Lens' ListVolumes Text

lvLimit :: Lens' ListVolumes (Maybe Natural)

Specifies that the list of volumes returned be limited to the specified number of items.

lvMarker :: Lens' ListVolumes (Maybe Text)

A string that indicates the position at which to begin the returned list of volumes. Obtain the marker from the response of a previous List iSCSI Volumes request.

Response

data ListVolumesResponse

Response constructor

listVolumesResponse :: ListVolumesResponse

ListVolumesResponse constructor.

The fields accessible through corresponding lenses are:

lvrGatewayARN :: Maybe Text

lvrMarker :: Maybe Text

lvrVolumeInfos :: [VolumeInfo]

Response lenses

lvrGatewayARN :: Lens' ListVolumesResponse (Maybe Text)

lvrMarker :: Lens' ListVolumesResponse (Maybe Text)

lvrVolumeInfos :: Lens' ListVolumesResponse [VolumeInfo]