Network.AWS.StorageGateway.ResetCache

Description

This operation resets all cache disks and makes the disks available for reconfiguration as cache storage. When a cache is reset, the gateway loses its cache storage. At this point you can reconfigure the disks as cache disks.

http://docs.aws.amazon.com/storagegateway/latest/APIReference/API_ResetCache.html

Request

data ResetCache

Request constructor

resetCache

Arguments

:: Text
rcGatewayARN

-> ResetCache

ResetCache constructor.

The fields accessible through corresponding lenses are:

rcGatewayARN :: Text

Request lenses

rcGatewayARN :: Lens' ResetCache Text

Response

data ResetCacheResponse

Response constructor

resetCacheResponse :: ResetCacheResponse

ResetCacheResponse constructor.

The fields accessible through corresponding lenses are:

rcrGatewayARN :: Maybe Text

Response lenses

rcrGatewayARN :: Lens' ResetCacheResponse (Maybe Text)