Network.AWS.StorageGateway.RetrieveTapeArchive

Request

data RetrieveTapeArchive

Request constructor

retrieveTapeArchive

Arguments

:: Text
rtaTapeARN

-> Text
rtaGatewayARN

-> RetrieveTapeArchive

Smart constructor for the minimum required parameters to construct a valid RetrieveTapeArchive request.

The fields accessible through corresponding lenses are:

TapeARN :: Text

GatewayARN :: Text

Request lenses

rtaTapeARN :: Lens' RetrieveTapeArchive Text

rtaGatewayARN :: Lens' RetrieveTapeArchive Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

Response

data RetrieveTapeArchiveResponse

Response constructor

retrieveTapeArchiveResponse :: RetrieveTapeArchiveResponse

Smart constructor for the minimum required parameters to construct a valid RetrieveTapeArchiveResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

TapeARN :: Maybe Text

Response lenses

rtarTapeARN :: Lens' RetrieveTapeArchiveResponse (Maybe Text)