Network.AWS.StorageGateway.RetrieveTapeRecoveryPoint

Description

Retrieves the recovery point for the specified virtual tape.

A recovery point is a point in time view of a virtual tape at which all the data on the tape is consistent. If your gateway crashes, virtual tapes that have recovery points can be recovered to a new gateway.

The virtual tape can be retrieved to only one gateway. The retrieved tape is read-only. The virtual tape can be retrieved to only a gateway-VTL. There is no charge for retrieving recovery points.

http://docs.aws.amazon.com/storagegateway/latest/APIReference/API_RetrieveTapeRecoveryPoint.html

Request

data RetrieveTapeRecoveryPoint

Request constructor

retrieveTapeRecoveryPoint

Arguments

:: Text
rtrpTapeARN

-> Text
rtrpGatewayARN

-> RetrieveTapeRecoveryPoint

RetrieveTapeRecoveryPoint constructor.

The fields accessible through corresponding lenses are:

rtrpGatewayARN :: Text

rtrpTapeARN :: Text

Request lenses

rtrpGatewayARN :: Lens' RetrieveTapeRecoveryPoint Text

rtrpTapeARN :: Lens' RetrieveTapeRecoveryPoint Text

The Amazon Resource Name (ARN) of the virtual tape for which you want to retrieve the recovery point.

Response

data RetrieveTapeRecoveryPointResponse

Response constructor

retrieveTapeRecoveryPointResponse :: RetrieveTapeRecoveryPointResponse

RetrieveTapeRecoveryPointResponse constructor.

The fields accessible through corresponding lenses are:

rtrprTapeARN :: Maybe Text

Response lenses

rtrprTapeARN :: Lens' RetrieveTapeRecoveryPointResponse (Maybe Text)

The Amazon Resource Name (ARN) of the virtual tape for which the recovery point was retrieved.