Network.AWS.StorageGateway.SetLocalConsolePassword

Description

Sets the password for your VM local console. When you log in to the local console for the first time, you log in to the VM with the default credentials. We recommend that you set a new password. You don't need to know the default password to set a new password.

Creating a Request

setLocalConsolePassword

Arguments

:: Text
slcpGatewayARN

-> Text
slcpLocalConsolePassword

-> SetLocalConsolePassword

Creates a value of SetLocalConsolePassword with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

slcpGatewayARN - Undocumented member.

slcpLocalConsolePassword - The password you want to set for your VM local console.

data SetLocalConsolePassword

SetLocalConsolePasswordInput

See: setLocalConsolePassword smart constructor.

Request Lenses

slcpGatewayARN :: Lens' SetLocalConsolePassword Text

Undocumented member.

slcpLocalConsolePassword :: Lens' SetLocalConsolePassword Text

The password you want to set for your VM local console.

Destructuring the Response

setLocalConsolePasswordResponse

Arguments

:: Int
slcprsResponseStatus

-> SetLocalConsolePasswordResponse

Creates a value of SetLocalConsolePasswordResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

slcprsGatewayARN - Undocumented member.

slcprsResponseStatus - -- | The response status code.

data SetLocalConsolePasswordResponse

See: setLocalConsolePasswordResponse smart constructor.

Response Lenses

slcprsGatewayARN :: Lens' SetLocalConsolePasswordResponse (Maybe Text)

Undocumented member.

slcprsResponseStatus :: Lens' SetLocalConsolePasswordResponse Int

- | The response status code.