Network.AWS.StorageGateway.ShutdownGateway

Description

This operation shuts down a gateway. To specify which gateway to shut down, use the Amazon Resource Name (ARN) of the gateway in the body of your request.

The operation shuts down the gateway service component running in the storage gateway's virtual machine (VM) and not the VM.

If you want to shut down the VM, it is recommended that you first shut down the gateway component in the VM to avoid unpredictable conditions. After the gateway is shutdown, you cannot call any other API except StartGateway, DescribeGatewayInformation, and ListGateways. For more information, see ActivateGateway. Your applications cannot read from or write to the gateway's storage volumes, and there are no snapshots taken.

When you make a shutdown request, you will get a '200 OK' success response immediately. However, it might take some time for the gateway to shut down. You can call the DescribeGatewayInformation API to check the status. For more information, see ActivateGateway. If do not intend to use the gateway again, you must delete the gateway (using DeleteGateway) to no longer pay software charges associated with the gateway.

http://docs.aws.amazon.com/storagegateway/latest/APIReference/API_ShutdownGateway.html

Request

data ShutdownGateway

Request constructor

shutdownGateway

Arguments

:: Text
sg1GatewayARN

-> ShutdownGateway

ShutdownGateway constructor.

The fields accessible through corresponding lenses are:

sg1GatewayARN :: Text

Request lenses

sg1GatewayARN :: Lens' ShutdownGateway Text

Response

data ShutdownGatewayResponse

Response constructor

shutdownGatewayResponse :: ShutdownGatewayResponse

ShutdownGatewayResponse constructor.

The fields accessible through corresponding lenses are:

sgr1GatewayARN :: Maybe Text

Response lenses

sgr1GatewayARN :: Lens' ShutdownGatewayResponse (Maybe Text)