Network.AWS.StorageGateway.Types

Service

data StorageGateway

Version 2013-06-30 of the Amazon Storage Gateway service.

Error

data JSONError :: *

ChapInfo

data ChapInfo

chapInfo :: ChapInfo

ChapInfo constructor.

The fields accessible through corresponding lenses are:

ciInitiatorName :: Maybe Text

ciSecretToAuthenticateInitiator :: Maybe Text

ciSecretToAuthenticateTarget :: Maybe Text

ciTargetARN :: Maybe Text

ciInitiatorName :: Lens' ChapInfo (Maybe Text)

The iSCSI initiator that connects to the target.

ciSecretToAuthenticateInitiator :: Lens' ChapInfo (Maybe Text)

The secret key that the initiator (e.g. Windows client) must provide to participate in mutual CHAP with the target.

ciSecretToAuthenticateTarget :: Lens' ChapInfo (Maybe Text)

The secret key that the target must provide to participate in mutual CHAP with the initiator (e.g. Windows client).

ciTargetARN :: Lens' ChapInfo (Maybe Text)

The Amazon Resource Name (ARN) of the volume.

Valid Values: 50 to 500 lowercase letters, numbers, periods (.), and hyphens (-).

VolumeiSCSIAttributes

data VolumeiSCSIAttributes

volumeiSCSIAttributes :: VolumeiSCSIAttributes

VolumeiSCSIAttributes constructor.

The fields accessible through corresponding lenses are:

vscsiaChapEnabled :: Maybe Bool

vscsiaLunNumber :: Maybe Natural

vscsiaNetworkInterfaceId :: Maybe Text

vscsiaNetworkInterfacePort :: Maybe Int

vscsiaTargetARN :: Maybe Text

vscsiaChapEnabled :: Lens' VolumeiSCSIAttributes (Maybe Bool)

Indicates whether mutual CHAP is enabled for the iSCSI target.

vscsiaLunNumber :: Lens' VolumeiSCSIAttributes (Maybe Natural)

The logical disk number.

vscsiaNetworkInterfaceId :: Lens' VolumeiSCSIAttributes (Maybe Text)

The network interface identifier.

vscsiaNetworkInterfacePort :: Lens' VolumeiSCSIAttributes (Maybe Int)

The port used to communicate with iSCSI targets.

vscsiaTargetARN :: Lens' VolumeiSCSIAttributes (Maybe Text)

The Amazon Resource Name (ARN) of the volume target.

DeviceiSCSIAttributes

data DeviceiSCSIAttributes

deviceiSCSIAttributes :: DeviceiSCSIAttributes

DeviceiSCSIAttributes constructor.

The fields accessible through corresponding lenses are:

dscsiaChapEnabled :: Maybe Bool

dscsiaNetworkInterfaceId :: Maybe Text

dscsiaNetworkInterfacePort :: Maybe Int

dscsiaTargetARN :: Maybe Text

dscsiaChapEnabled :: Lens' DeviceiSCSIAttributes (Maybe Bool)

Indicates whether mutual CHAP is enabled for the iSCSI target.

dscsiaNetworkInterfaceId :: Lens' DeviceiSCSIAttributes (Maybe Text)

The network interface identifier of the VTL device.

dscsiaNetworkInterfacePort :: Lens' DeviceiSCSIAttributes (Maybe Int)

The port used to communicate with iSCSI VTL device targets.

dscsiaTargetARN :: Lens' DeviceiSCSIAttributes (Maybe Text)

Specifies the unique Amazon Resource Name(ARN) that encodes the iSCSI qualified name(iqn) of a tape drive or media changer target.

Error'

data Error'

error :: Error'

Error' constructor.

The fields accessible through corresponding lenses are:

eErrorCode :: Maybe ErrorCode

eErrorDetails :: HashMap Text Text

eErrorCode :: Lens' Error' (Maybe ErrorCode)

Additional information about the error.

eErrorDetails :: Lens' Error' (HashMap Text Text)

Human-readable text that provides detail about the error that occurred.

Disk

data Disk

disk :: Disk

Disk constructor.

The fields accessible through corresponding lenses are:

dDiskAllocationResource :: Maybe Text

dDiskAllocationType :: Maybe Text

dDiskId :: Maybe Text

dDiskNode :: Maybe Text

dDiskPath :: Maybe Text

dDiskSizeInBytes :: Maybe Integer

dDiskStatus :: Maybe Text

dDiskAllocationResource :: Lens' Disk (Maybe Text)

dDiskAllocationType :: Lens' Disk (Maybe Text)

dDiskId :: Lens' Disk (Maybe Text)

dDiskNode :: Lens' Disk (Maybe Text)

dDiskPath :: Lens' Disk (Maybe Text)

dDiskSizeInBytes :: Lens' Disk (Maybe Integer)

dDiskStatus :: Lens' Disk (Maybe Text)

Tape

data Tape

tape :: Tape

Tape constructor.

The fields accessible through corresponding lenses are:

tProgress :: Maybe Double

tTapeARN :: Maybe Text

tTapeBarcode :: Maybe Text

tTapeSizeInBytes :: Maybe Integer

tTapeStatus :: Maybe Text

tVTLDevice :: Maybe Text

tProgress :: Lens' Tape (Maybe Double)

For archiving virtual tapes, indicates how much data remains to be uploaded before archiving is complete.

Range: 0 (not started) to 100 (complete).

tTapeARN :: Lens' Tape (Maybe Text)

The Amazon Resource Name (ARN) of the virtual tape.

tTapeBarcode :: Lens' Tape (Maybe Text)

The barcode that identifies a specific virtual tape.

tTapeSizeInBytes :: Lens' Tape (Maybe Integer)

The size, in bytes, of the virtual tape.

tTapeStatus :: Lens' Tape (Maybe Text)

The current state of the virtual tape.

tVTLDevice :: Lens' Tape (Maybe Text)

The virtual tape library (VTL) device that the virtual tape is associated with.

NetworkInterface

data NetworkInterface

networkInterface :: NetworkInterface

NetworkInterface constructor.

The fields accessible through corresponding lenses are:

niIpv4Address :: Maybe Text

niIpv6Address :: Maybe Text

niMacAddress :: Maybe Text

niIpv4Address :: Lens' NetworkInterface (Maybe Text)

The Internet Protocol version 4 (IPv4) address of the interface.

niIpv6Address :: Lens' NetworkInterface (Maybe Text)

The Internet Protocol version 6 (IPv6) address of the interface. Currentlynot supported.

niMacAddress :: Lens' NetworkInterface (Maybe Text)

The Media Access Control (MAC) address of the interface.

This is currently unsupported and will not be returned in output.

VTLDevice

data VTLDevice

vtldevice :: VTLDevice

VTLDevice constructor.

The fields accessible through corresponding lenses are:

vtldDeviceiSCSIAttributes :: Maybe DeviceiSCSIAttributes

vtldVTLDeviceARN :: Maybe Text

vtldVTLDeviceProductIdentifier :: Maybe Text

vtldVTLDeviceType :: Maybe Text

vtldVTLDeviceVendor :: Maybe Text

vtldDeviceiSCSIAttributes :: Lens' VTLDevice (Maybe DeviceiSCSIAttributes)

A list of iSCSI information about a VTL device.

vtldVTLDeviceARN :: Lens' VTLDevice (Maybe Text)

Specifies the unique Amazon Resource Name (ARN) of the device (tape drive or media changer).

vtldVTLDeviceProductIdentifier :: Lens' VTLDevice (Maybe Text)

vtldVTLDeviceType :: Lens' VTLDevice (Maybe Text)

vtldVTLDeviceVendor :: Lens' VTLDevice (Maybe Text)

TapeRecoveryPointInfo

data TapeRecoveryPointInfo

tapeRecoveryPointInfo :: TapeRecoveryPointInfo

TapeRecoveryPointInfo constructor.

The fields accessible through corresponding lenses are:

trpiTapeARN :: Maybe Text

trpiTapeRecoveryPointTime :: Maybe UTCTime

trpiTapeSizeInBytes :: Maybe Integer

trpiTapeStatus :: Maybe Text

trpiTapeARN :: Lens' TapeRecoveryPointInfo (Maybe Text)

The Amazon Resource Name (ARN) of the virtual tape.

trpiTapeRecoveryPointTime :: Lens' TapeRecoveryPointInfo (Maybe UTCTime)

The time when the point-in-time view of the virtual tape was replicated for later recovery.

The string format of the tape recovery point time is in the ISO8601 extended YYYY-MM-DD'T'HH:MM:SS'Z' format.

trpiTapeSizeInBytes :: Lens' TapeRecoveryPointInfo (Maybe Integer)

The size, in bytes, of the virtual tapes to recover.

trpiTapeStatus :: Lens' TapeRecoveryPointInfo (Maybe Text)

VolumeRecoveryPointInfo

data VolumeRecoveryPointInfo

volumeRecoveryPointInfo :: VolumeRecoveryPointInfo

VolumeRecoveryPointInfo constructor.

The fields accessible through corresponding lenses are:

vrpiVolumeARN :: Maybe Text

vrpiVolumeRecoveryPointTime :: Maybe Text

vrpiVolumeSizeInBytes :: Maybe Integer

vrpiVolumeUsageInBytes :: Maybe Integer

vrpiVolumeARN :: Lens' VolumeRecoveryPointInfo (Maybe Text)

vrpiVolumeRecoveryPointTime :: Lens' VolumeRecoveryPointInfo (Maybe Text)

vrpiVolumeSizeInBytes :: Lens' VolumeRecoveryPointInfo (Maybe Integer)

vrpiVolumeUsageInBytes :: Lens' VolumeRecoveryPointInfo (Maybe Integer)

TapeArchive

data TapeArchive

tapeArchive :: TapeArchive

TapeArchive constructor.

The fields accessible through corresponding lenses are:

taCompletionTime :: Maybe UTCTime

taRetrievedTo :: Maybe Text

taTapeARN :: Maybe Text

taTapeBarcode :: Maybe Text

taTapeSizeInBytes :: Maybe Integer

taTapeStatus :: Maybe Text

taCompletionTime :: Lens' TapeArchive (Maybe UTCTime)

The time that the archiving of the virtual tape was completed.

The string format of the completion time is in the ISO8601 extended YYYY-MM-DD'T'HH:MM:SS'Z' format.

taRetrievedTo :: Lens' TapeArchive (Maybe Text)

The Amazon Resource Name (ARN) of the gateway-VTL that the virtual tape is being retrieved to.

The virtual tape is retrieved from the virtual tape shelf (VTS).

taTapeARN :: Lens' TapeArchive (Maybe Text)

The Amazon Resource Name (ARN) of an archived virtual tape.

taTapeBarcode :: Lens' TapeArchive (Maybe Text)

The barcode that identifies the archived virtual tape.

taTapeSizeInBytes :: Lens' TapeArchive (Maybe Integer)

The size, in bytes, of the archived virtual tape.

taTapeStatus :: Lens' TapeArchive (Maybe Text)

The current state of the archived virtual tape.

ErrorCode

data ErrorCode

Constructors

ActivationKeyExpired
ActivationKeyExpired

ActivationKeyInvalid
ActivationKeyInvalid

ActivationKeyNotFound
ActivationKeyNotFound

AuthenticationFailure
AuthenticationFailure

BandwidthThrottleScheduleNotFound
BandwidthThrottleScheduleNotFound

Blocked
Blocked

CannotExportSnapshot
CannotExportSnapshot

ChapCredentialNotFound
ChapCredentialNotFound

DiskAlreadyAllocated
DiskAlreadyAllocated

DiskDoesNotExist
DiskDoesNotExist

DiskSizeGreaterThanVolumeMaxSize
DiskSizeGreaterThanVolumeMaxSize

DiskSizeLessThanVolumeSize
DiskSizeLessThanVolumeSize

DiskSizeNotGigAligned
DiskSizeNotGigAligned

DuplicateCertificateInfo
DuplicateCertificateInfo

DuplicateSchedule
DuplicateSchedule

EndpointNotFound
EndpointNotFound

GatewayInternalError
GatewayInternalError

GatewayNotConnected
GatewayNotConnected

GatewayNotFound
GatewayNotFound

GatewayProxyNetworkConnectionBusy
GatewayProxyNetworkConnectionBusy

IAMNotSupported
IAMNotSupported

InitiatorInvalid
InitiatorInvalid

InitiatorNotFound
InitiatorNotFound

InternalError
InternalError

InvalidEndpoint
InvalidEndpoint

InvalidGateway
InvalidGateway

InvalidParameters
InvalidParameters

InvalidSchedule
InvalidSchedule

LocalStorageLimitExceeded
LocalStorageLimitExceeded

LunAlreadyAllocated
LunAlreadyAllocated

LunInvalid
LunInvalid

MaximumContentLengthExceeded
MaximumContentLengthExceeded

MaximumTapeCartridgeCountExceeded
MaximumTapeCartridgeCountExceeded

MaximumVolumeCountExceeded
MaximumVolumeCountExceeded

NetworkConfigurationChanged
NetworkConfigurationChanged

NoDisksAvailable
NoDisksAvailable

NotImplemented
NotImplemented

NotSupported
NotSupported

OperationAborted
OperationAborted

OutdatedGateway
OutdatedGateway

ParametersNotImplemented
ParametersNotImplemented

RegionInvalid
RegionInvalid

RequestTimeout
RequestTimeout

ServiceUnavailable
ServiceUnavailable

SnapshotDeleted
SnapshotDeleted

SnapshotIdInvalid
SnapshotIdInvalid

SnapshotInProgress
SnapshotInProgress

SnapshotNotFound
SnapshotNotFound

SnapshotScheduleNotFound
SnapshotScheduleNotFound

StagingAreaFull
StagingAreaFull

StorageFailure
StorageFailure

TapeCartridgeNotFound
TapeCartridgeNotFound

TargetAlreadyExists
TargetAlreadyExists

TargetInvalid
TargetInvalid

TargetNotFound
TargetNotFound

UnauthorizedOperation
UnauthorizedOperation

VolumeAlreadyExists
VolumeAlreadyExists

VolumeIdInvalid
VolumeIdInvalid

VolumeInUse
VolumeInUse

VolumeNotFound
VolumeNotFound

VolumeNotReady
VolumeNotReady

StorediSCSIVolume

data StorediSCSIVolume

storediSCSIVolume :: StorediSCSIVolume

StorediSCSIVolume constructor.

The fields accessible through corresponding lenses are:

sscsivPreservedExistingData :: Maybe Bool

sscsivSourceSnapshotId :: Maybe Text

sscsivVolumeARN :: Maybe Text

sscsivVolumeDiskId :: Maybe Text

sscsivVolumeId :: Maybe Text

sscsivVolumeProgress :: Maybe Double

sscsivVolumeSizeInBytes :: Maybe Integer

sscsivVolumeStatus :: Maybe Text

sscsivVolumeType :: Maybe Text

sscsivVolumeiSCSIAttributes :: Maybe VolumeiSCSIAttributes

sscsivPreservedExistingData :: Lens' StorediSCSIVolume (Maybe Bool)

sscsivSourceSnapshotId :: Lens' StorediSCSIVolume (Maybe Text)

sscsivVolumeARN :: Lens' StorediSCSIVolume (Maybe Text)

sscsivVolumeDiskId :: Lens' StorediSCSIVolume (Maybe Text)

sscsivVolumeId :: Lens' StorediSCSIVolume (Maybe Text)

sscsivVolumeProgress :: Lens' StorediSCSIVolume (Maybe Double)

sscsivVolumeSizeInBytes :: Lens' StorediSCSIVolume (Maybe Integer)

sscsivVolumeStatus :: Lens' StorediSCSIVolume (Maybe Text)

sscsivVolumeType :: Lens' StorediSCSIVolume (Maybe Text)

sscsivVolumeiSCSIAttributes :: Lens' StorediSCSIVolume (Maybe VolumeiSCSIAttributes)

CachediSCSIVolume

data CachediSCSIVolume

cachediSCSIVolume :: CachediSCSIVolume

CachediSCSIVolume constructor.

The fields accessible through corresponding lenses are:

cscsivSourceSnapshotId :: Maybe Text

cscsivVolumeARN :: Maybe Text

cscsivVolumeId :: Maybe Text

cscsivVolumeProgress :: Maybe Double

cscsivVolumeSizeInBytes :: Maybe Integer

cscsivVolumeStatus :: Maybe Text

cscsivVolumeType :: Maybe Text

cscsivVolumeiSCSIAttributes :: Maybe VolumeiSCSIAttributes

cscsivSourceSnapshotId :: Lens' CachediSCSIVolume (Maybe Text)

cscsivVolumeARN :: Lens' CachediSCSIVolume (Maybe Text)

cscsivVolumeId :: Lens' CachediSCSIVolume (Maybe Text)

cscsivVolumeProgress :: Lens' CachediSCSIVolume (Maybe Double)

cscsivVolumeSizeInBytes :: Lens' CachediSCSIVolume (Maybe Integer)

cscsivVolumeStatus :: Lens' CachediSCSIVolume (Maybe Text)

cscsivVolumeType :: Lens' CachediSCSIVolume (Maybe Text)

cscsivVolumeiSCSIAttributes :: Lens' CachediSCSIVolume (Maybe VolumeiSCSIAttributes)

VolumeInfo

data VolumeInfo

volumeInfo :: VolumeInfo

VolumeInfo constructor.

The fields accessible through corresponding lenses are:

viVolumeARN :: Maybe Text

viVolumeType :: Maybe Text

viVolumeARN :: Lens' VolumeInfo (Maybe Text)

viVolumeType :: Lens' VolumeInfo (Maybe Text)

GatewayInfo

data GatewayInfo

gatewayInfo :: GatewayInfo

GatewayInfo constructor.

The fields accessible through corresponding lenses are:

giGatewayARN :: Maybe Text

giGatewayOperationalState :: Maybe Text

giGatewayType :: Maybe Text

giGatewayARN :: Lens' GatewayInfo (Maybe Text)

giGatewayOperationalState :: Lens' GatewayInfo (Maybe Text)

giGatewayType :: Lens' GatewayInfo (Maybe Text)