Network.AWS.StorageGateway.UpdateChapCredentials

Description

This operation updates the Challenge-Handshake Authentication Protocol (CHAP) credentials for a specified iSCSI target. By default, a gateway does not have CHAP enabled; however, for added security, you might use it.

When you update CHAP credentials, all existing connections on the target are closed and initiators must reconnect with the new credentials.

http://docs.aws.amazon.com/storagegateway/latest/APIReference/API_UpdateChapCredentials.html

Request

data UpdateChapCredentials

Request constructor

updateChapCredentials

Arguments

:: Text
uccTargetARN

-> Text
uccSecretToAuthenticateInitiator

-> Text
uccInitiatorName

-> UpdateChapCredentials

UpdateChapCredentials constructor.

The fields accessible through corresponding lenses are:

uccInitiatorName :: Text

uccSecretToAuthenticateInitiator :: Text

uccSecretToAuthenticateTarget :: Maybe Text

uccTargetARN :: Text

Request lenses

uccInitiatorName :: Lens' UpdateChapCredentials Text

The iSCSI initiator that connects to the target.

uccSecretToAuthenticateInitiator :: Lens' UpdateChapCredentials Text

The secret key that the initiator (e.g. Windows client) must provide to participate in mutual CHAP with the target.

uccSecretToAuthenticateTarget :: Lens' UpdateChapCredentials (Maybe Text)

The secret key that the target must provide to participate in mutual CHAP with the initiator (e.g. Windows client).

uccTargetARN :: Lens' UpdateChapCredentials Text

The Amazon Resource Name (ARN) of the iSCSI volume target. Use the DescribeStorediSCSIVolumes operation to return to retrieve the TargetARN for specified VolumeARN.

Response

data UpdateChapCredentialsResponse

Response constructor

updateChapCredentialsResponse :: UpdateChapCredentialsResponse

UpdateChapCredentialsResponse constructor.

The fields accessible through corresponding lenses are:

uccrInitiatorName :: Maybe Text

uccrTargetARN :: Maybe Text

Response lenses

uccrInitiatorName :: Lens' UpdateChapCredentialsResponse (Maybe Text)

The iSCSI initiator that connects to the target. This is the same initiator name specified in the request.

uccrTargetARN :: Lens' UpdateChapCredentialsResponse (Maybe Text)

The Amazon Resource Name (ARN) of the target. This is the same target specified in the request.