Network.AWS.StorageGateway.UpdateGatewayInformation

Description

Updates a gateway's metadata, which includes the gateway's name and time zone. To specify which gateway to update, use the Amazon Resource Name (ARN) of the gateway in your request.

Creating a Request

updateGatewayInformation

Arguments

:: Text
ugiGatewayARN

-> UpdateGatewayInformation

Creates a value of UpdateGatewayInformation with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ugiGatewayName - Undocumented member.

ugiGatewayTimezone - Undocumented member.

ugiGatewayARN - Undocumented member.

data UpdateGatewayInformation

See: updateGatewayInformation smart constructor.

Request Lenses

ugiGatewayName :: Lens' UpdateGatewayInformation (Maybe Text)

Undocumented member.

ugiGatewayTimezone :: Lens' UpdateGatewayInformation (Maybe Text)

Undocumented member.

ugiGatewayARN :: Lens' UpdateGatewayInformation Text

Undocumented member.

Destructuring the Response

updateGatewayInformationResponse

Arguments

:: Int
ugirsResponseStatus

-> UpdateGatewayInformationResponse

Creates a value of UpdateGatewayInformationResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ugirsGatewayARN - Undocumented member.

ugirsGatewayName - Undocumented member.

ugirsResponseStatus - -- | The response status code.

data UpdateGatewayInformationResponse

A JSON object containing the ARN of the gateway that was updated.

See: updateGatewayInformationResponse smart constructor.

Response Lenses

ugirsGatewayARN :: Lens' UpdateGatewayInformationResponse (Maybe Text)

Undocumented member.

ugirsGatewayName :: Lens' UpdateGatewayInformationResponse (Maybe Text)

Undocumented member.

ugirsResponseStatus :: Lens' UpdateGatewayInformationResponse Int

- | The response status code.