Network.AWS.StorageGateway.UpdateGatewaySoftwareNow

Description

This operation updates the gateway virtual machine (VM) software. The request immediately triggers the software update. Before sending this request, you should make sure all your applications have finished writing to your gateway's storage volumes in order to avoid data loss. During the update, applications cannot use the gateway's storage volumes. --> When you make this request, you get a 200 OK success response immediately. However, it might take some time for the update to complete. You can call DescribeGatewayInformation to verify the gateway is in the STATE_RUNNING state. A software update forces a system restart of your gateway. You can minimize the chance of any disruption to your applications by increasing your iSCSI Initiators' timeouts. For more information about increasing iSCSI Initiator timeouts for Windows and Linux, see Customizing Your Windows iSCSI Settings and Customizing Your Linux iSCSI Settings, respectively. Example Request The following example shows a request that initiates a gateway VM update. POST HTTP1.1 Host: storagegateway.us-east-1.amazonaws.com x-amz-Date: 20120425T120000Z Authorization: CSOC7TJPLR0OOKIRLGOHVAICUFVV4KQNSO5AEMVJF66Q9ASUAAJG Content-type: application/x-amz-json-1.1 x-amz-target: StorageGateway_20120630.UpdateGatewaySoftwareNow { GatewayARN: "arn:aws:storagegateway:us-east-1:111122223333:gateway/mygateway" } HTTP/1.1 200 OK x-amzn-RequestId: CSOC7TJPLR0OOKIRLGOHVAICUFVV4KQNSO5AEMVJF66Q9ASUAAJG Date: Wed, 25 Apr 2012 12:00:02 GMT Content-type: application/x-amz-json-1.1 Content-length: 80 { GatewayARN: "arn:aws:storagegateway:us-east-1:111122223333:gateway/mygateway" }.

Request

data UpdateGatewaySoftwareNow

A JSON object containing the of the gateway to update.

Request constructor

mkUpdateGatewaySoftwareNow

Arguments

:: Text
ugsnGatewayARN

-> UpdateGatewaySoftwareNow

Smart constructor for the minimum required parameters to construct a valid UpdateGatewaySoftwareNow request.

The fields accessible through corresponding lenses are:

GatewayARN :: Text

Request lenses

ugsnGatewayARN :: Lens' UpdateGatewaySoftwareNow Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

Response

data UpdateGatewaySoftwareNowResponse

A JSON object containing the of the gateway that was updated.

Response constructor

mkUpdateGatewaySoftwareNowResponse :: UpdateGatewaySoftwareNowResponse

Smart constructor for the minimum required parameters to construct a valid UpdateGatewaySoftwareNowResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

GatewayARN :: Maybe Text

Response lenses

ugsnrGatewayARN :: Lens' UpdateGatewaySoftwareNowResponse (Maybe Text)

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.