Network.AWS.StorageGateway.UpdateMaintenanceStartTime

Description

This operation updates a gateway's weekly maintenance start time information, including day and time of the week. The maintenance time is the time in your gateway's time zone. Example Request The following example shows a request that updates the maintenance start time of mygateway. POST HTTP1.1 Host: storagegateway.us-east-1.amazonaws.com x-amz-Date: 20120425T120000Z Authorization: CSOC7TJPLR0OOKIRLGOHVAICUFVV4KQNSO5AEMVJF66Q9ASUAAJG Content-type: application/x-amz-json-1.1 x-amz-target: StorageGateway_20120630.UpdateMaintenanceStartTime { GatewayARN: "arn:aws:storagegateway:us-east-1:111122223333:gateway/mygateway", HourOfDay: 0, MinuteOfHour: 30, DayOfWeek: 2 } HTTP/1.1 200 OK x-amzn-RequestId: CSOC7TJPLR0OOKIRLGOHVAICUFVV4KQNSO5AEMVJF66Q9ASUAAJG Date: Wed, 25 Apr 2012 12:00:02 GMT Content-type: application/x-amz-json-1.1 Content-length: 80 { GatewayARN: "arn:aws:storagegateway:us-east-1:111122223333:gateway/mygateway" }.

Request

data UpdateMaintenanceStartTime

A JSON object containing the following fields: UpdateMaintenanceStartTimeInput$DayOfWeek UpdateMaintenanceStartTimeInput$HourOfDay UpdateMaintenanceStartTimeInput$MinuteOfHour.

Request constructor

updateMaintenanceStartTime

Arguments

:: Text
umstGatewayARN

-> Integer
umstHourOfDay

-> Integer
umstMinuteOfHour

-> Integer
umstDayOfWeek

-> UpdateMaintenanceStartTime

Smart constructor for the minimum required parameters to construct a valid UpdateMaintenanceStartTime request.

The fields accessible through corresponding lenses are:

GatewayARN :: Text

HourOfDay :: Integer

MinuteOfHour :: Integer

DayOfWeek :: Integer

Request lenses

umstGatewayARN :: Lens' UpdateMaintenanceStartTime Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

umstHourOfDay :: Lens' UpdateMaintenanceStartTime Integer

The hour component of the maintenance start time represented as hh, where hh is the hour (00 to 23). The hour of the day is in the time zone of the gateway.

umstMinuteOfHour :: Lens' UpdateMaintenanceStartTime Integer

The minute component of the maintenance start time represented as mm, where mm is the minute (00 to 59). The minute of the hour is in the time zone of the gateway.

umstDayOfWeek :: Lens' UpdateMaintenanceStartTime Integer

The maintenance start time day of the week.

Response

data UpdateMaintenanceStartTimeResponse

A JSON object containing the of the gateway whose maintenance start time is updated.

Response constructor

updateMaintenanceStartTimeResponse

Arguments

:: Text
umstrGatewayARN

-> UpdateMaintenanceStartTimeResponse

Smart constructor for the minimum required parameters to construct a valid UpdateMaintenanceStartTimeResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

GatewayARN :: Text

Response lenses

umstrGatewayARN :: Lens' UpdateMaintenanceStartTimeResponse Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.