Network.AWS.StorageGateway.UpdateNFSFileShare

Description

Updates a file share. This operation is only supported in the file gateway architecture.

Updates the following file share setting:

Default storage class for your S3 bucket

Metadata defaults for your S3 bucket

Allowed NFS clients for your file share

Squash settings

Write status of your file share

Creating a Request

updateNFSFileShare

Arguments

:: Text
unfsfsFileShareARN

-> UpdateNFSFileShare

Creates a value of UpdateNFSFileShare with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

unfsfsKMSKey - The KMS key used for Amazon S3 server side encryption. This value can only be set when KmsEncrypted is true. Optional.

unfsfsKMSEncrypted - True to use Amazon S3 server side encryption with your own AWS KMS key, or false to use a key managed by Amazon S3. Optional.

unfsfsDefaultStorageClass - The default storage class for objects put into an Amazon S3 bucket by a file gateway. Possible values are S3_STANDARD or S3_STANDARD_IA. If this field is not populated, the default value S3_STANDARD is used. Optional.

unfsfsSquash - The user mapped to anonymous user. Valid options are the following: * RootSquash - Only root is mapped to anonymous user. * NoSquash - No one is mapped to anonymous user * AllSquash - Everyone is mapped to anonymous user.

unfsfsNFSFileShareDefaults - The default values for the file share. Optional.

unfsfsClientList - The list of clients that are allowed to access the file gateway. The list must contain either valid IP addresses or valid CIDR blocks.

unfsfsReadOnly - Sets the write status of a file share: "true" if the write status is read-only, otherwise "false".

unfsfsFileShareARN - The Amazon Resource Name (ARN) of the file share to be updated.

data UpdateNFSFileShare

UpdateNFSFileShareInput

See: updateNFSFileShare smart constructor.

Request Lenses

unfsfsKMSKey :: Lens' UpdateNFSFileShare (Maybe Text)

The KMS key used for Amazon S3 server side encryption. This value can only be set when KmsEncrypted is true. Optional.

unfsfsKMSEncrypted :: Lens' UpdateNFSFileShare (Maybe Bool)

True to use Amazon S3 server side encryption with your own AWS KMS key, or false to use a key managed by Amazon S3. Optional.

unfsfsDefaultStorageClass :: Lens' UpdateNFSFileShare (Maybe Text)

The default storage class for objects put into an Amazon S3 bucket by a file gateway. Possible values are S3_STANDARD or S3_STANDARD_IA. If this field is not populated, the default value S3_STANDARD is used. Optional.

unfsfsSquash :: Lens' UpdateNFSFileShare (Maybe Text)

The user mapped to anonymous user. Valid options are the following: * RootSquash - Only root is mapped to anonymous user. * NoSquash - No one is mapped to anonymous user * AllSquash - Everyone is mapped to anonymous user.

unfsfsNFSFileShareDefaults :: Lens' UpdateNFSFileShare (Maybe NFSFileShareDefaults)

The default values for the file share. Optional.

unfsfsClientList :: Lens' UpdateNFSFileShare (Maybe (NonEmpty Text))

The list of clients that are allowed to access the file gateway. The list must contain either valid IP addresses or valid CIDR blocks.

unfsfsReadOnly :: Lens' UpdateNFSFileShare (Maybe Bool)

Sets the write status of a file share: "true" if the write status is read-only, otherwise "false".

unfsfsFileShareARN :: Lens' UpdateNFSFileShare Text

The Amazon Resource Name (ARN) of the file share to be updated.

Destructuring the Response

updateNFSFileShareResponse

Arguments

:: Int
unfsfsrsResponseStatus

-> UpdateNFSFileShareResponse

Creates a value of UpdateNFSFileShareResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

unfsfsrsFileShareARN - The Amazon Resource Name (ARN) of the updated file share.

unfsfsrsResponseStatus - -- | The response status code.

data UpdateNFSFileShareResponse

UpdateNFSFileShareOutput

See: updateNFSFileShareResponse smart constructor.

Response Lenses

unfsfsrsFileShareARN :: Lens' UpdateNFSFileShareResponse (Maybe Text)

The Amazon Resource Name (ARN) of the updated file share.

unfsfsrsResponseStatus :: Lens' UpdateNFSFileShareResponse Int

- | The response status code.