Network.AWS.StorageGateway.UpdateSnapshotSchedule

Description

This operation updates a snapshot schedule configured for a gateway volume. The default snapshot schedule for volume is once every 24 hours, starting at the creation time of the volume. You can use this API to change the snapshot schedule configured for the volume. In the request you must identify the gateway volume whose snapshot schedule you want to update, and the schedule information, including when you want the snapshot to begin on a day and the frequency (in hours) of snapshots.

http://docs.aws.amazon.com/storagegateway/latest/APIReference/API_UpdateSnapshotSchedule.html

Request

data UpdateSnapshotSchedule

Request constructor

updateSnapshotSchedule

Arguments

:: Text
ussVolumeARN

-> Natural
ussStartAt

-> Natural
ussRecurrenceInHours

-> UpdateSnapshotSchedule

UpdateSnapshotSchedule constructor.

The fields accessible through corresponding lenses are:

ussDescription :: Maybe Text

ussRecurrenceInHours :: Natural

ussStartAt :: Natural

ussVolumeARN :: Text

Request lenses

ussDescription :: Lens' UpdateSnapshotSchedule (Maybe Text)

Optional description of the snapshot that overwrites the existing description.

ussRecurrenceInHours :: Lens' UpdateSnapshotSchedule Natural

Frequency of snapshots. Specify the number of hours between snapshots.

ussStartAt :: Lens' UpdateSnapshotSchedule Natural

The hour of the day at which the snapshot schedule begins represented as hh, where hh is the hour (0 to 23). The hour of the day is in the time zone of the gateway.

ussVolumeARN :: Lens' UpdateSnapshotSchedule Text

The Amazon Resource Name (ARN) of the volume. Use the ListVolumes operation to return a list of gateway volumes.

Response

data UpdateSnapshotScheduleResponse

Response constructor

updateSnapshotScheduleResponse :: UpdateSnapshotScheduleResponse

UpdateSnapshotScheduleResponse constructor.

The fields accessible through corresponding lenses are:

ussrVolumeARN :: Maybe Text

Response lenses

ussrVolumeARN :: Lens' UpdateSnapshotScheduleResponse (Maybe Text)