Network.AWS.StorageGateway.V2013_06_30

Description

AWS Storage Gateway is a service that connects an on-premises software appliance with cloud-based storage to provide seamless and secure integration between your on-premises IT environment and AWS's storage infrastructure.

Documentation

activateGateway

Arguments

:: Text
_agiActivationKey

-> Text
_agiGatewayName

-> Text
_agiGatewayTimezone

-> Text
_agiGatewayRegion

-> ActivateGateway

Minimum specification for a ActivateGateway request.

data ActivateGateway

Constructors

ActivateGateway

Fields

_agiActivationKey :: Text

Your gateway activation key. You can obtain the activation key by sending an HTTP GET request with redirects enabled to the gateway IP address (port 80). The redirect URL returned in the response provides you the activation key for your gateway in the query string parameter activationKey. It may also include other activation-related parameters, however, these are merely defaults -- the arguments you pass to the ActivateGateway API call determine the actual configuration of your gateway.

_agiGatewayName :: Text

A unique identifier for your gateway. This name becomes part of the gateway Amazon Resources Name (ARN) which is what you use as an input to other operations.

_agiGatewayTimezone :: Text

One of the values that indicates the time zone you want to set for the gateway. The time zone is used, for example, for scheduling snapshots and your gateway's maintenance schedule.

_agiGatewayRegion :: Text

One of the values that indicates the region where you want to store the snapshot backups. The gateway region specified must be the same region as the region in your Host header in the request. For more information about available regions and endpoints for AWS Storage Gateway, see Regions and Endpoints in the Amazon Web Services Glossary. Valid Values: "us-east-1", "us-west-1", "us-west-2", "eu-west-1", "ap-northeast-1", "ap-southest-1", "sa-east-1".

_agiGatewayType :: Maybe Text

One of the values that defines the type of gateway to activate. The type specified is critical to all later functions of the gateway and cannot be changed after activation. The default value is STORED.

agiGatewayType :: Lens' ActivateGateway (Maybe Text)

agiGatewayTimezone :: Lens' ActivateGateway Text

agiGatewayRegion :: Lens' ActivateGateway Text

agiGatewayName :: Lens' ActivateGateway Text

agiActivationKey :: Lens' ActivateGateway Text

data ActivateGatewayResponse

Constructors

ActivateGatewayResponse

Fields

_agoGatewayARN :: Maybe Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

agoGatewayARN :: Iso' ActivateGatewayResponse (Maybe Text)

data AddCache

Constructors

AddCache

Fields

_aciDiskIds :: [Text]

_aciGatewayARN :: Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

aciGatewayARN :: Lens' AddCache Text

aciDiskIds :: Lens' AddCache [Text]

data AddCacheResponse

Constructors

AddCacheResponse

Fields

_acoGatewayARN :: Maybe Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

acoGatewayARN :: Iso' AddCacheResponse (Maybe Text)

data AddUploadBuffer

Constructors

AddUploadBuffer

Fields

_aubiDiskIds :: [Text]

_aubiGatewayARN :: Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

aubiGatewayARN :: Lens' AddUploadBuffer Text

aubiDiskIds :: Lens' AddUploadBuffer [Text]

data AddUploadBufferResponse

Constructors

AddUploadBufferResponse

Fields

_auboGatewayARN :: Maybe Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

auboGatewayARN :: Iso' AddUploadBufferResponse (Maybe Text)

data AddWorkingStorage

Constructors

AddWorkingStorage

Fields

_awsiDiskIds :: [Text]

An array of strings that identify disks that are to be configured as working storage. Each string have a minimum length of 1 and maximum length of 300. You can get the disk IDs from the ListLocalDisks API.

_awsiGatewayARN :: Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

awsiGatewayARN :: Lens' AddWorkingStorage Text

awsiDiskIds :: Lens' AddWorkingStorage [Text]

data AddWorkingStorageResponse

Constructors

AddWorkingStorageResponse

Fields

_awsoGatewayARN :: Maybe Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

awsoGatewayARN :: Iso' AddWorkingStorageResponse (Maybe Text)

data CancelArchival

Constructors

CancelArchival

Fields

_caiGatewayARN :: Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

_caiTapeARN :: Text

caiTapeARN :: Lens' CancelArchival Text

caiGatewayARN :: Lens' CancelArchival Text

data CancelArchivalResponse

Constructors

CancelArchivalResponse

Fields

_caoTapeARN :: Maybe Text

caoTapeARN :: Iso' CancelArchivalResponse (Maybe Text)

data CancelRetrieval

Constructors

CancelRetrieval

Fields

_criGatewayARN :: Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

_criTapeARN :: Text

criTapeARN :: Lens' CancelRetrieval Text

criGatewayARN :: Lens' CancelRetrieval Text

data CancelRetrievalResponse

Constructors

CancelRetrievalResponse

Fields

_croTapeARN :: Maybe Text

croTapeARN :: Iso' CancelRetrievalResponse (Maybe Text)

createCachediSCSIVolume

Arguments

:: Text
_ccscsiviClientToken

-> Text
_ccscsiviGatewayARN

-> Integer
_ccscsiviVolumeSizeInBytes

-> Text
_ccscsiviNetworkInterfaceId

-> Text
_ccscsiviTargetName

-> CreateCachediSCSIVolume

Minimum specification for a CreateCachediSCSIVolume request.

data CreateCachediSCSIVolume

Constructors

CreateCachediSCSIVolume

Fields

_ccscsiviClientToken :: Text

_ccscsiviGatewayARN :: Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

_ccscsiviVolumeSizeInBytes :: Integer

_ccscsiviNetworkInterfaceId :: Text

_ccscsiviTargetName :: Text

_ccscsiviSnapshotId :: Maybe Text

ccscsiviVolumeSizeInBytes :: Lens' CreateCachediSCSIVolume Integer

ccscsiviTargetName :: Lens' CreateCachediSCSIVolume Text

ccscsiviSnapshotId :: Lens' CreateCachediSCSIVolume (Maybe Text)

ccscsiviNetworkInterfaceId :: Lens' CreateCachediSCSIVolume Text

ccscsiviGatewayARN :: Lens' CreateCachediSCSIVolume Text

ccscsiviClientToken :: Lens' CreateCachediSCSIVolume Text

data CreateCachediSCSIVolumeResponse

Constructors

CreateCachediSCSIVolumeResponse

Fields

_ccscsivoTargetARN :: Maybe Text

_ccscsivoVolumeARN :: Maybe Text

ccscsivoVolumeARN :: Lens' CreateCachediSCSIVolumeResponse (Maybe Text)

ccscsivoTargetARN :: Lens' CreateCachediSCSIVolumeResponse (Maybe Text)

data CreateSnapshot

Constructors

CreateSnapshot

Fields

_csiSnapshotDescription :: Text

Textual description of the snapshot that appears in the Amazon EC2 console, Elastic Block Store snapshots panel in the Description field, and in the AWS Storage Gateway snapshot Details pane, Description field.

_csiVolumeARN :: Text

The Amazon Resource Name (ARN) of the volume. Use the ListVolumes operation to return a list of gateway volumes.

csiVolumeARN :: Lens' CreateSnapshot Text

csiSnapshotDescription :: Lens' CreateSnapshot Text

data CreateSnapshotResponse

Constructors

CreateSnapshotResponse

Fields

_csoSnapshotId :: Maybe Text

The snapshot ID that is used to refer to the snapshot in future operations such as describing snapshots (Amazon Elastic Compute Cloud API DescribeSnapshots) or creating a volume from a snapshot (CreateStorediSCSIVolume).

_csoVolumeARN :: Maybe Text

The Amazon Resource Name (ARN) of the volume of which the snapshot was taken.

csoVolumeARN :: Lens' CreateSnapshotResponse (Maybe Text)

csoSnapshotId :: Lens' CreateSnapshotResponse (Maybe Text)

data CreateSnapshotFromVolumeRecoveryPoint

Constructors

CreateSnapshotFromVolumeRecoveryPoint

Fields

_csfvrpiSnapshotDescription :: Text

_csfvrpiVolumeARN :: Text

csfvrpiVolumeARN :: Lens' CreateSnapshotFromVolumeRecoveryPoint Text

csfvrpiSnapshotDescription :: Lens' CreateSnapshotFromVolumeRecoveryPoint Text

data CreateSnapshotFromVolumeRecoveryPointResponse

Constructors

CreateSnapshotFromVolumeRecoveryPointResponse

Fields

_csfvrpoSnapshotId :: Maybe Text

_csfvrpoVolumeRecoveryPointTime :: Maybe Text

_csfvrpoVolumeARN :: Maybe Text

csfvrpoVolumeRecoveryPointTime :: Lens' CreateSnapshotFromVolumeRecoveryPointResponse (Maybe Text)

csfvrpoVolumeARN :: Lens' CreateSnapshotFromVolumeRecoveryPointResponse (Maybe Text)

csfvrpoSnapshotId :: Lens' CreateSnapshotFromVolumeRecoveryPointResponse (Maybe Text)

createStorediSCSIVolume

Arguments

:: Bool
_csscsiviPreserveExistingData

-> Text
_csscsiviDiskId

-> Text
_csscsiviGatewayARN

-> Text
_csscsiviNetworkInterfaceId

-> Text
_csscsiviTargetName

-> CreateStorediSCSIVolume

Minimum specification for a CreateStorediSCSIVolume request.

data CreateStorediSCSIVolume

Constructors

CreateStorediSCSIVolume

Fields

_csscsiviPreserveExistingData :: Bool

Specify this field as true if you want to preserve the data on the local disk. Otherwise, specifying this field as false creates an empty volume. Valid Values: true, false.

_csscsiviDiskId :: Text

The unique identifier for the gateway local disk that is configured as a stored volume. Use ListLocalDisks to list disk IDs for a gateway.

_csscsiviGatewayARN :: Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

_csscsiviNetworkInterfaceId :: Text

The network interface of the gateway on which to expose the iSCSI target. Only IPv4 addresses are accepted. Use DescribeGatewayInformation to get a list of the network interfaces available on a gateway. Valid Values: A valid IP address.

_csscsiviTargetName :: Text

The name of the iSCSI target used by initiators to connect to the target and as a suffix for the target ARN. For example, specifying TargetName as myvolume results in the target ARN of arn:aws:storagegateway:us-east-1:111122223333:gatewaymygatewaytarget/iqn.1997-05.com.amazon:myvolume. The target name must be unique across all volumes of a gateway.

_csscsiviSnapshotId :: Maybe Text

The snapshot ID (e.g. "snap-1122aabb") of the snapshot to restore as the new stored volume. Specify this field if you want to create the iSCSI storage volume from a snapshot otherwise do not include this field. To list snapshots for your account use DescribeSnapshots in the Amazon Elastic Compute Cloud API Reference.

csscsiviTargetName :: Lens' CreateStorediSCSIVolume Text

csscsiviSnapshotId :: Lens' CreateStorediSCSIVolume (Maybe Text)

csscsiviPreserveExistingData :: Lens' CreateStorediSCSIVolume Bool

csscsiviNetworkInterfaceId :: Lens' CreateStorediSCSIVolume Text

csscsiviGatewayARN :: Lens' CreateStorediSCSIVolume Text

csscsiviDiskId :: Lens' CreateStorediSCSIVolume Text

data CreateStorediSCSIVolumeResponse

Constructors

CreateStorediSCSIVolumeResponse

Fields

_csscsivoVolumeSizeInBytes :: Maybe Integer

The size of the volume in bytes.

_csscsivoTargetARN :: Maybe Text

he Amazon Resource Name (ARN) of the volume target that includes the iSCSI name that initiators can use to connect to the target.

_csscsivoVolumeARN :: Maybe Text

The Amazon Resource Name (ARN) of the configured volume.

csscsivoVolumeSizeInBytes :: Lens' CreateStorediSCSIVolumeResponse (Maybe Integer)

csscsivoVolumeARN :: Lens' CreateStorediSCSIVolumeResponse (Maybe Text)

csscsivoTargetARN :: Lens' CreateStorediSCSIVolumeResponse (Maybe Text)

data CreateTapes

Constructors

CreateTapes

Fields

_ctiClientToken :: Text

_ctiGatewayARN :: Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

_ctiNumTapesToCreate :: Integer

_ctiTapeBarcodePrefix :: Text

_ctiTapeSizeInBytes :: Integer

ctiTapeSizeInBytes :: Lens' CreateTapes Integer

ctiTapeBarcodePrefix :: Lens' CreateTapes Text

ctiNumTapesToCreate :: Lens' CreateTapes Integer

ctiGatewayARN :: Lens' CreateTapes Text

ctiClientToken :: Lens' CreateTapes Text

data CreateTapesResponse

Constructors

CreateTapesResponse

Fields

_ctoTapeARNs :: [Text]

ctoTapeARNs :: Iso' CreateTapesResponse [Text]

data DeleteBandwidthRateLimit

Constructors

DeleteBandwidthRateLimit

Fields

_dbrljBandwidthType :: Text

_dbrljGatewayARN :: Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

dbrljGatewayARN :: Lens' DeleteBandwidthRateLimit Text

dbrljBandwidthType :: Lens' DeleteBandwidthRateLimit Text

data DeleteBandwidthRateLimitResponse

Constructors

DeleteBandwidthRateLimitResponse

Fields

_dbrlpGatewayARN :: Maybe Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

dbrlpGatewayARN :: Iso' DeleteBandwidthRateLimitResponse (Maybe Text)

data DeleteChapCredentials

Constructors

DeleteChapCredentials

Fields

_dccjInitiatorName :: Text

The iSCSI initiator that connects to the target.

_dccjTargetARN :: Text

The Amazon Resource Name (ARN) of the iSCSI volume target. Use the DescribeStorediSCSIVolumes operation to return to retrieve the TargetARN for specified VolumeARN.

dccjTargetARN :: Lens' DeleteChapCredentials Text

dccjInitiatorName :: Lens' DeleteChapCredentials Text

data DeleteChapCredentialsResponse

Constructors

DeleteChapCredentialsResponse

Fields

_dccpInitiatorName :: Maybe Text

The iSCSI initiator that connects to the target.

_dccpTargetARN :: Maybe Text

The Amazon Resource Name (ARN) of the target.

dccpTargetARN :: Lens' DeleteChapCredentialsResponse (Maybe Text)

dccpInitiatorName :: Lens' DeleteChapCredentialsResponse (Maybe Text)

data DeleteGateway

Constructors

DeleteGateway

Fields

_dgiGatewayARN :: Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

dgiGatewayARN :: Iso' DeleteGateway Text

data DeleteGatewayResponse

Constructors

DeleteGatewayResponse

Fields

_dgoGatewayARN :: Maybe Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

dgoGatewayARN :: Iso' DeleteGatewayResponse (Maybe Text)

data DeleteSnapshotSchedule

Constructors

DeleteSnapshotSchedule

Fields

_dssjVolumeARN :: Text

dssjVolumeARN :: Iso' DeleteSnapshotSchedule Text

data DeleteSnapshotScheduleResponse

Constructors

DeleteSnapshotScheduleResponse

Fields

_dsspVolumeARN :: Maybe Text

dsspVolumeARN :: Iso' DeleteSnapshotScheduleResponse (Maybe Text)

data DeleteTape

Constructors

DeleteTape

Fields

_dtjGatewayARN :: Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

_dtjTapeARN :: Text

dtjTapeARN :: Lens' DeleteTape Text

dtjGatewayARN :: Lens' DeleteTape Text

data DeleteTapeResponse

Constructors

DeleteTapeResponse

Fields

_dtpTapeARN :: Maybe Text

dtpTapeARN :: Iso' DeleteTapeResponse (Maybe Text)

data DeleteTapeArchive

Constructors

DeleteTapeArchive

Fields

_dtajTapeARN :: Text

dtajTapeARN :: Iso' DeleteTapeArchive Text

data DeleteTapeArchiveResponse

Constructors

DeleteTapeArchiveResponse

Fields

_dtapTapeARN :: Maybe Text

dtapTapeARN :: Iso' DeleteTapeArchiveResponse (Maybe Text)

data DeleteVolume

Constructors

DeleteVolume

Fields

_dviVolumeARN :: Text

The Amazon Resource Name (ARN) of the volume. Use the ListVolumes operation to return a list of gateway volumes.

dviVolumeARN :: Iso' DeleteVolume Text

data DeleteVolumeResponse

Constructors

DeleteVolumeResponse

Fields

_dvoVolumeARN :: Maybe Text

The Amazon Resource Name (ARN) of the storage volume that was deleted. It is the same ARN you provided in the request.

dvoVolumeARN :: Iso' DeleteVolumeResponse (Maybe Text)

data DescribeBandwidthRateLimit

Constructors

DescribeBandwidthRateLimit

Fields

_dbrliGatewayARN :: Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

dbrliGatewayARN :: Iso' DescribeBandwidthRateLimit Text

data DescribeBandwidthRateLimitResponse

Constructors

DescribeBandwidthRateLimitResponse

Fields

_dbrloAverageDownloadRateLimitInBitsPerSec :: Maybe Integer

The average download bandwidth rate limit in bits per second. This field does not appear in the response if the download rate limit is not set.

_dbrloAverageUploadRateLimitInBitsPerSec :: Maybe Integer

The average upload bandwidth rate limit in bits per second. This field does not appear in the response if the upload rate limit is not set.

_dbrloGatewayARN :: Maybe Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

dbrloGatewayARN :: Lens' DescribeBandwidthRateLimitResponse (Maybe Text)

dbrloAverageUploadRateLimitInBitsPerSec :: Lens' DescribeBandwidthRateLimitResponse (Maybe Integer)

dbrloAverageDownloadRateLimitInBitsPerSec :: Lens' DescribeBandwidthRateLimitResponse (Maybe Integer)

data DescribeCache

Constructors

DescribeCache

Fields

_dciGatewayARN :: Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

dciGatewayARN :: Iso' DescribeCache Text

data DescribeCacheResponse

Constructors

DescribeCacheResponse

Fields

_dcoDiskIds :: [Text]

_dcoCacheUsedPercentage :: Maybe Double

_dcoCacheHitPercentage :: Maybe Double

_dcoCacheMissPercentage :: Maybe Double

_dcoCacheDirtyPercentage :: Maybe Double

_dcoGatewayARN :: Maybe Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

_dcoCacheAllocatedInBytes :: Maybe Integer

dcoGatewayARN :: Lens' DescribeCacheResponse (Maybe Text)

dcoDiskIds :: Lens' DescribeCacheResponse [Text]

dcoCacheUsedPercentage :: Lens' DescribeCacheResponse (Maybe Double)

dcoCacheMissPercentage :: Lens' DescribeCacheResponse (Maybe Double)

dcoCacheHitPercentage :: Lens' DescribeCacheResponse (Maybe Double)

dcoCacheDirtyPercentage :: Lens' DescribeCacheResponse (Maybe Double)

dcoCacheAllocatedInBytes :: Lens' DescribeCacheResponse (Maybe Integer)

data DescribeCachediSCSIVolumes

Constructors

DescribeCachediSCSIVolumes

Fields

_dcscsiviVolumeARNs :: [Text]

dcscsiviVolumeARNs :: Iso' DescribeCachediSCSIVolumes [Text]

data DescribeCachediSCSIVolumesResponse

Constructors

DescribeCachediSCSIVolumesResponse

Fields

_dcscsivoCachediSCSIVolumes :: [CachediSCSIVolumeInformation]

An array of objects where each object contains metadata about one cached volume.

dcscsivoCachediSCSIVolumes :: Iso' DescribeCachediSCSIVolumesResponse [CachediSCSIVolumeInformation]

data DescribeChapCredentials

Constructors

DescribeChapCredentials

Fields

_dcciTargetARN :: Text

The Amazon Resource Name (ARN) of the iSCSI volume target. Use the DescribeStorediSCSIVolumes operation to return to retrieve the TargetARN for specified VolumeARN.

dcciTargetARN :: Iso' DescribeChapCredentials Text

data DescribeChapCredentialsResponse

Constructors

DescribeChapCredentialsResponse

Fields

_dccoChapCredentials :: [ChapInfo]

An array of ChapInfo objects that represent CHAP credentials. Each object in the array contains CHAP credential information for one target-initiator pair. If no CHAP credentials are set, an empty array is returned. CHAP credential information is provided in a JSON object with the following fields: InitiatorName: The iSCSI initiator that connects to the target. SecretToAuthenticateInitiator: The secret key that the initiator (e.g. Windows client) must provide to participate in mutual CHAP with the target. SecretToAuthenticateTarget: The secret key that the target must provide to participate in mutual CHAP with the initiator (e.g. Windows client). TargetARN: The Amazon Resource Name (ARN) of the storage volume.

dccoChapCredentials :: Iso' DescribeChapCredentialsResponse [ChapInfo]

data DescribeGatewayInformation

Constructors

DescribeGatewayInformation

Fields

_dgiiGatewayARN :: Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

dgiiGatewayARN :: Iso' DescribeGatewayInformation Text

data DescribeGatewayInformationResponse

Constructors

DescribeGatewayInformationResponse

Fields

_dgioGatewayARN :: Maybe Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

_dgioGatewayId :: Maybe Text

The gateway ID.

_dgioGatewayNetworkInterfaces :: [NetworkInterface]

A NetworkInterface array that contains descriptions of the gateway network interfaces.

_dgioGatewayState :: Maybe Text

One of the values that indicates the operating state of the gateway.

_dgioGatewayTimezone :: Maybe Text

One of the values that indicates the time zone configured for the gateway.

_dgioGatewayType :: Maybe Text

TBD.

_dgioNextUpdateAvailabilityDate :: Maybe Text

The date at which an update to the gateway is available. This date is in the time zone of the gateway. If the gateway is not available for an update this field is not returned in the response. response example. -->.

dgioNextUpdateAvailabilityDate :: Lens' DescribeGatewayInformationResponse (Maybe Text)

dgioGatewayType :: Lens' DescribeGatewayInformationResponse (Maybe Text)

dgioGatewayTimezone :: Lens' DescribeGatewayInformationResponse (Maybe Text)

dgioGatewayState :: Lens' DescribeGatewayInformationResponse (Maybe Text)

dgioGatewayNetworkInterfaces :: Lens' DescribeGatewayInformationResponse [NetworkInterface]

dgioGatewayId :: Lens' DescribeGatewayInformationResponse (Maybe Text)

dgioGatewayARN :: Lens' DescribeGatewayInformationResponse (Maybe Text)

data DescribeMaintenanceStartTime

Constructors

DescribeMaintenanceStartTime

Fields

_dmstiGatewayARN :: Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

dmstiGatewayARN :: Iso' DescribeMaintenanceStartTime Text

data DescribeMaintenanceStartTimeResponse

Constructors

DescribeMaintenanceStartTimeResponse

Fields

_dmstoDayOfWeek :: Maybe Integer

_dmstoGatewayARN :: Maybe Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

_dmstoTimezone :: Maybe Text

_dmstoHourOfDay :: Maybe Integer

_dmstoMinuteOfHour :: Maybe Integer

dmstoTimezone :: Lens' DescribeMaintenanceStartTimeResponse (Maybe Text)

dmstoMinuteOfHour :: Lens' DescribeMaintenanceStartTimeResponse (Maybe Integer)

dmstoHourOfDay :: Lens' DescribeMaintenanceStartTimeResponse (Maybe Integer)

dmstoGatewayARN :: Lens' DescribeMaintenanceStartTimeResponse (Maybe Text)

dmstoDayOfWeek :: Lens' DescribeMaintenanceStartTimeResponse (Maybe Integer)

data DescribeSnapshotSchedule

Constructors

DescribeSnapshotSchedule

Fields

_dssiVolumeARN :: Text

The Amazon Resource Name (ARN) of the volume. Use the ListVolumes operation to return a list of gateway volumes.

dssiVolumeARN :: Iso' DescribeSnapshotSchedule Text

data DescribeSnapshotScheduleResponse

Constructors

DescribeSnapshotScheduleResponse

Fields

_dssoDescription :: Maybe Text

_dssoTimezone :: Maybe Text

_dssoStartAt :: Maybe Integer

_dssoRecurrenceInHours :: Maybe Integer

_dssoVolumeARN :: Maybe Text

dssoVolumeARN :: Lens' DescribeSnapshotScheduleResponse (Maybe Text)

dssoTimezone :: Lens' DescribeSnapshotScheduleResponse (Maybe Text)

dssoStartAt :: Lens' DescribeSnapshotScheduleResponse (Maybe Integer)

dssoRecurrenceInHours :: Lens' DescribeSnapshotScheduleResponse (Maybe Integer)

dssoDescription :: Lens' DescribeSnapshotScheduleResponse (Maybe Text)

data DescribeStorediSCSIVolumes

Constructors

DescribeStorediSCSIVolumes

Fields

_dsscsiviVolumeARNs :: [Text]

An array of strings where each string represents the Amazon Resource Name (ARN) of a stored volume. All of the specified stored volumes must from the same gateway. Use ListVolumes to get volume ARNs for a gateway.

dsscsiviVolumeARNs :: Iso' DescribeStorediSCSIVolumes [Text]

data DescribeStorediSCSIVolumesResponse

Constructors

DescribeStorediSCSIVolumesResponse

Fields

_dsscsivoStorediSCSIVolumes :: [StorediSCSIVolumeInformation]

dsscsivoStorediSCSIVolumes :: Iso' DescribeStorediSCSIVolumesResponse [StorediSCSIVolumeInformation]

describeTapeArchives :: DescribeTapeArchives

Minimum specification for a DescribeTapeArchives request.

data DescribeTapeArchives

Constructors

DescribeTapeArchives

Fields

_dtaiMarker :: Maybe Text

_dtaiLimit :: Maybe Integer

_dtaiTapeARNs :: [Text]

dtaiTapeARNs :: Lens' DescribeTapeArchives [Text]

dtaiMarker :: Lens' DescribeTapeArchives (Maybe Text)

dtaiLimit :: Lens' DescribeTapeArchives (Maybe Integer)

data DescribeTapeArchivesResponse

Constructors

DescribeTapeArchivesResponse

Fields

_dtaoMarker :: Maybe Text

_dtaoTapeArchives :: [TapeArchive]

dtaoTapeArchives :: Lens' DescribeTapeArchivesResponse [TapeArchive]

dtaoMarker :: Lens' DescribeTapeArchivesResponse (Maybe Text)

describeTapeRecoveryPoints

Arguments

:: Text
_dtrpiGatewayARN

-> DescribeTapeRecoveryPoints

Minimum specification for a DescribeTapeRecoveryPoints request.

data DescribeTapeRecoveryPoints

Constructors

DescribeTapeRecoveryPoints

Fields

_dtrpiGatewayARN :: Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

_dtrpiMarker :: Maybe Text

_dtrpiLimit :: Maybe Integer

dtrpiMarker :: Lens' DescribeTapeRecoveryPoints (Maybe Text)

dtrpiLimit :: Lens' DescribeTapeRecoveryPoints (Maybe Integer)

dtrpiGatewayARN :: Lens' DescribeTapeRecoveryPoints Text

data DescribeTapeRecoveryPointsResponse

Constructors

DescribeTapeRecoveryPointsResponse

Fields

_dtrpoGatewayARN :: Maybe Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

_dtrpoMarker :: Maybe Text

_dtrpoTapeRecoveryPointInfos :: [TapeRecoveryPointInfo]

dtrpoTapeRecoveryPointInfos :: Lens' DescribeTapeRecoveryPointsResponse [TapeRecoveryPointInfo]

dtrpoMarker :: Lens' DescribeTapeRecoveryPointsResponse (Maybe Text)

dtrpoGatewayARN :: Lens' DescribeTapeRecoveryPointsResponse (Maybe Text)

describeTapes

Arguments

:: Text
_dtiGatewayARN

-> DescribeTapes

Minimum specification for a DescribeTapes request.

data DescribeTapes

Constructors

DescribeTapes

Fields

_dtiGatewayARN :: Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

_dtiMarker :: Maybe Text

_dtiLimit :: Maybe Integer

_dtiTapeARNs :: [Text]

dtiTapeARNs :: Lens' DescribeTapes [Text]

dtiMarker :: Lens' DescribeTapes (Maybe Text)

dtiLimit :: Lens' DescribeTapes (Maybe Integer)

dtiGatewayARN :: Lens' DescribeTapes Text

data DescribeTapesResponse

Constructors

DescribeTapesResponse

Fields

_dtoMarker :: Maybe Text

_dtoTapes :: [Tape]

dtoTapes :: Lens' DescribeTapesResponse [Tape]

dtoMarker :: Lens' DescribeTapesResponse (Maybe Text)

data DescribeUploadBuffer

Constructors

DescribeUploadBuffer

Fields

_dubiGatewayARN :: Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

dubiGatewayARN :: Iso' DescribeUploadBuffer Text

data DescribeUploadBufferResponse

Constructors

DescribeUploadBufferResponse

Fields

_duboDiskIds :: [Text]

_duboGatewayARN :: Maybe Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

_duboUploadBufferAllocatedInBytes :: Maybe Integer

_duboUploadBufferUsedInBytes :: Maybe Integer

duboUploadBufferUsedInBytes :: Lens' DescribeUploadBufferResponse (Maybe Integer)

duboUploadBufferAllocatedInBytes :: Lens' DescribeUploadBufferResponse (Maybe Integer)

duboGatewayARN :: Lens' DescribeUploadBufferResponse (Maybe Text)

duboDiskIds :: Lens' DescribeUploadBufferResponse [Text]

describeVTLDevices

Arguments

:: Text
_dvtldiGatewayARN

-> DescribeVTLDevices

Minimum specification for a DescribeVTLDevices request.

data DescribeVTLDevices

Constructors

DescribeVTLDevices

Fields

_dvtldiGatewayARN :: Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

_dvtldiMarker :: Maybe Text

_dvtldiLimit :: Maybe Integer

_dvtldiVTLDeviceARNs :: [Text]

dvtldiVTLDeviceARNs :: Lens' DescribeVTLDevices [Text]

dvtldiMarker :: Lens' DescribeVTLDevices (Maybe Text)

dvtldiLimit :: Lens' DescribeVTLDevices (Maybe Integer)

dvtldiGatewayARN :: Lens' DescribeVTLDevices Text

data DescribeVTLDevicesResponse

Constructors

DescribeVTLDevicesResponse

Fields

_dvtldoGatewayARN :: Maybe Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

_dvtldoMarker :: Maybe Text

_dvtldoVTLDevices :: [VTLDevice]

dvtldoVTLDevices :: Lens' DescribeVTLDevicesResponse [VTLDevice]

dvtldoMarker :: Lens' DescribeVTLDevicesResponse (Maybe Text)

dvtldoGatewayARN :: Lens' DescribeVTLDevicesResponse (Maybe Text)

data DescribeWorkingStorage

Constructors

DescribeWorkingStorage

Fields

_dwsiGatewayARN :: Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

dwsiGatewayARN :: Iso' DescribeWorkingStorage Text

data DescribeWorkingStorageResponse

Constructors

DescribeWorkingStorageResponse

Fields

_dwsoDiskIds :: [Text]

An array of the gateway's local disk IDs that are configured as working storage. Each local disk ID is specified as a string (minimum length of 1 and maximum length of 300). If no local disks are configured as working storage, then the DiskIds array is empty.

_dwsoGatewayARN :: Maybe Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

_dwsoWorkingStorageAllocatedInBytes :: Maybe Integer

The total working storage in bytes allocated for the gateway. If no working storage is configured for the gateway, this field returns 0.

_dwsoWorkingStorageUsedInBytes :: Maybe Integer

The total working storage in bytes in use by the gateway. If no working storage is configured for the gateway, this field returns 0.

dwsoWorkingStorageUsedInBytes :: Lens' DescribeWorkingStorageResponse (Maybe Integer)

dwsoWorkingStorageAllocatedInBytes :: Lens' DescribeWorkingStorageResponse (Maybe Integer)

dwsoGatewayARN :: Lens' DescribeWorkingStorageResponse (Maybe Text)

dwsoDiskIds :: Lens' DescribeWorkingStorageResponse [Text]

data DisableGateway

Constructors

DisableGateway

Fields

_dgjGatewayARN :: Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

dgjGatewayARN :: Iso' DisableGateway Text

data DisableGatewayResponse

Constructors

DisableGatewayResponse

Fields

_dgpGatewayARN :: Maybe Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

dgpGatewayARN :: Iso' DisableGatewayResponse (Maybe Text)

listGateways :: ListGateways

Minimum specification for a ListGateways request.

data ListGateways

Constructors

ListGateways

Fields

_lgiMarker :: Maybe Text

An opaque string that indicates the position at which to begin the returned list of gateways.

_lgiLimit :: Maybe Integer

Specifies that the list of gateways returned be limited to the specified number of items.

lgiMarker :: Lens' ListGateways (Maybe Text)

lgiLimit :: Lens' ListGateways (Maybe Integer)

data ListGatewaysResponse

Constructors

ListGatewaysResponse

Fields

_lgoGateways :: [GatewayInformation]

_lgoMarker :: Maybe Text

lgoMarker :: Lens' ListGatewaysResponse (Maybe Text)

lgoGateways :: Lens' ListGatewaysResponse [GatewayInformation]

data ListLocalDisks

Constructors

ListLocalDisks

Fields

_lldiGatewayARN :: Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

lldiGatewayARN :: Iso' ListLocalDisks Text

data ListLocalDisksResponse

Constructors

ListLocalDisksResponse

Fields

_lldoDisks :: [DiskInformation]

_lldoGatewayARN :: Maybe Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

lldoGatewayARN :: Lens' ListLocalDisksResponse (Maybe Text)

lldoDisks :: Lens' ListLocalDisksResponse [DiskInformation]

data ListVolumeRecoveryPoints

Constructors

ListVolumeRecoveryPoints

Fields

_lvrpiGatewayARN :: Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

lvrpiGatewayARN :: Iso' ListVolumeRecoveryPoints Text

data ListVolumeRecoveryPointsResponse

Constructors

ListVolumeRecoveryPointsResponse

Fields

_lvrpoGatewayARN :: Maybe Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

_lvrpoVolumeRecoveryPointInfos :: [VolumeRecoveryPointInfo]

lvrpoVolumeRecoveryPointInfos :: Lens' ListVolumeRecoveryPointsResponse [VolumeRecoveryPointInfo]

lvrpoGatewayARN :: Lens' ListVolumeRecoveryPointsResponse (Maybe Text)

listVolumes

Arguments

:: Text
_lviGatewayARN

-> ListVolumes

Minimum specification for a ListVolumes request.

data ListVolumes

Constructors

ListVolumes

Fields

_lviGatewayARN :: Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

_lviMarker :: Maybe Text

A string that indicates the position at which to begin the returned list of volumes. Obtain the marker from the response of a previous List iSCSI Volumes request.

_lviLimit :: Maybe Integer

Specifies that the list of volumes returned be limited to the specified number of items.

lviMarker :: Lens' ListVolumes (Maybe Text)

lviLimit :: Lens' ListVolumes (Maybe Integer)

lviGatewayARN :: Lens' ListVolumes Text

data ListVolumesResponse

Constructors

ListVolumesResponse

Fields

_lvoGatewayARN :: Maybe Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

_lvoMarker :: Maybe Text

_lvoVolumeInfos :: [VolumeInformation]

lvoVolumeInfos :: Lens' ListVolumesResponse [VolumeInformation]

lvoMarker :: Lens' ListVolumesResponse (Maybe Text)

lvoGatewayARN :: Lens' ListVolumesResponse (Maybe Text)

data RetrieveTapeArchive

Constructors

RetrieveTapeArchive

Fields

_rtaiGatewayARN :: Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

_rtaiTapeARN :: Text

rtaiTapeARN :: Lens' RetrieveTapeArchive Text

rtaiGatewayARN :: Lens' RetrieveTapeArchive Text

data RetrieveTapeArchiveResponse

Constructors

RetrieveTapeArchiveResponse

Fields

_rtaoTapeARN :: Maybe Text

rtaoTapeARN :: Iso' RetrieveTapeArchiveResponse (Maybe Text)

data RetrieveTapeRecoveryPoint

Constructors

RetrieveTapeRecoveryPoint

Fields

_rtrpiGatewayARN :: Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

_rtrpiTapeARN :: Text

rtrpiTapeARN :: Lens' RetrieveTapeRecoveryPoint Text

rtrpiGatewayARN :: Lens' RetrieveTapeRecoveryPoint Text

data RetrieveTapeRecoveryPointResponse

Constructors

RetrieveTapeRecoveryPointResponse

Fields

_rtrpoTapeARN :: Maybe Text

rtrpoTapeARN :: Iso' RetrieveTapeRecoveryPointResponse (Maybe Text)

data ShutdownGateway

Constructors

ShutdownGateway

Fields

_sgjGatewayARN :: Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

sgjGatewayARN :: Iso' ShutdownGateway Text

data ShutdownGatewayResponse

Constructors

ShutdownGatewayResponse

Fields

_sgpGatewayARN :: Maybe Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

sgpGatewayARN :: Iso' ShutdownGatewayResponse (Maybe Text)

data StartGateway

Constructors

StartGateway

Fields

_sgiGatewayARN :: Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

sgiGatewayARN :: Iso' StartGateway Text

data StartGatewayResponse

Constructors

StartGatewayResponse

Fields

_sgoGatewayARN :: Maybe Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

sgoGatewayARN :: Iso' StartGatewayResponse (Maybe Text)

data StorageGateway

Supported version (2013-06-30) of the AWS Storage Gateway service.

data ErrorCode

Additional information about the error.

Constructors

ErrorCodeActivationKeyExpired
ActivationKeyExpired

ErrorCodeActivationKeyInvalid
ActivationKeyInvalid

ErrorCodeActivationKeyNotFound
ActivationKeyNotFound

ErrorCodeAuthenticationFailure
AuthenticationFailure

ErrorCodeBandwidthThrottleScheduleNotFound
BandwidthThrottleScheduleNotFound

ErrorCodeBlocked
Blocked

ErrorCodeCannotExportSnapshot
CannotExportSnapshot

ErrorCodeChapCredentialNotFound
ChapCredentialNotFound

ErrorCodeDiskAlreadyAllocated
DiskAlreadyAllocated

ErrorCodeDiskDoesNotExist
DiskDoesNotExist

ErrorCodeDiskSizeGreaterThanVolumeMaxSize
DiskSizeGreaterThanVolumeMaxSize

ErrorCodeDiskSizeLessThanVolumeSize
DiskSizeLessThanVolumeSize

ErrorCodeDiskSizeNotGigAligned
DiskSizeNotGigAligned

ErrorCodeDuplicateCertificateInfo
DuplicateCertificateInfo

ErrorCodeDuplicateSchedule
DuplicateSchedule

ErrorCodeEndpointNotFound
EndpointNotFound

ErrorCodeGatewayInternalError
GatewayInternalError

ErrorCodeGatewayNotConnected
GatewayNotConnected

ErrorCodeGatewayNotFound
GatewayNotFound

ErrorCodeGatewayProxyNetworkConnectionBusy
GatewayProxyNetworkConnectionBusy

ErrorCodeIAMNotSupported
IAMNotSupported

ErrorCodeInitiatorInvalid
InitiatorInvalid

ErrorCodeInitiatorNotFound
InitiatorNotFound

ErrorCodeInternalError
InternalError

ErrorCodeInvalidEndpoint
InvalidEndpoint

ErrorCodeInvalidGateway
InvalidGateway

ErrorCodeInvalidParameters
InvalidParameters

ErrorCodeInvalidSchedule
InvalidSchedule

ErrorCodeLocalStorageLimitExceeded
LocalStorageLimitExceeded

ErrorCodeLunAlreadyAllocated
LunAlreadyAllocated

ErrorCodeLunInvalid
LunInvalid

ErrorCodeMaximumContentLengthExceeded
MaximumContentLengthExceeded

ErrorCodeMaximumTapeCartridgeCountExceeded
MaximumTapeCartridgeCountExceeded

ErrorCodeMaximumVolumeCountExceeded
MaximumVolumeCountExceeded

ErrorCodeNetworkConfigurationChanged
NetworkConfigurationChanged

ErrorCodeNoDisksAvailable
NoDisksAvailable

ErrorCodeNotImplemented
NotImplemented

ErrorCodeNotSupported
NotSupported

ErrorCodeOperationAborted
OperationAborted

ErrorCodeOutdatedGateway
OutdatedGateway

ErrorCodeParametersNotImplemented
ParametersNotImplemented

ErrorCodeRegionInvalid
RegionInvalid

ErrorCodeRequestTimeout
RequestTimeout

ErrorCodeServiceUnavailable
ServiceUnavailable

ErrorCodeSnapshotDeleted
SnapshotDeleted

ErrorCodeSnapshotIdInvalid
SnapshotIdInvalid

ErrorCodeSnapshotInProgress
SnapshotInProgress

ErrorCodeSnapshotNotFound
SnapshotNotFound

ErrorCodeSnapshotScheduleNotFound
SnapshotScheduleNotFound

ErrorCodeStagingAreaFull
StagingAreaFull

ErrorCodeStorageFailure
StorageFailure

ErrorCodeTapeCartridgeNotFound
TapeCartridgeNotFound

ErrorCodeTargetAlreadyExists
TargetAlreadyExists

ErrorCodeTargetInvalid
TargetInvalid

ErrorCodeTargetNotFound
TargetNotFound

ErrorCodeUnauthorizedOperation
UnauthorizedOperation

ErrorCodeVolumeAlreadyExists
VolumeAlreadyExists

ErrorCodeVolumeIdInvalid
VolumeIdInvalid

ErrorCodeVolumeInUse
VolumeInUse

ErrorCodeVolumeNotFound
VolumeNotFound

ErrorCodeVolumeNotReady
VolumeNotReady

data CachediSCSIVolumeInformation

Constructors

CachediSCSIVolumeInformation

Fields

_cscsiviVolumeiSCSIAttributes :: Maybe VolumeiSCSIAttributes

Lists iSCSI information about a volume.

_cscsiviVolumeStatus :: Maybe Text

_cscsiviSourceSnapshotId :: Maybe Text

_cscsiviVolumeARN :: Maybe Text

_cscsiviVolumeProgress :: Maybe Double

_cscsiviVolumeSizeInBytes :: Maybe Integer

_cscsiviVolumeId :: Maybe Text

_cscsiviVolumeType :: Maybe Text

data ChapInfo

Describes Challenge-Handshake Authentication Protocol (CHAP) information that supports authentication between your gateway and iSCSI initiators.

Constructors

ChapInfo

Fields

_ciTargetARN :: Maybe Text

The Amazon Resource Name (ARN) of the volume. Valid Values: 50 to 500 lowercase letters, numbers, periods (.), and hyphens (-).

_ciSecretToAuthenticateInitiator :: Maybe Text

The secret key that the initiator (e.g. Windows client) must provide to participate in mutual CHAP with the target.

_ciInitiatorName :: Maybe Text

The iSCSI initiator that connects to the target.

_ciSecretToAuthenticateTarget :: Maybe Text

The secret key that the target must provide to participate in mutual CHAP with the initiator (e.g. Windows client).

data DeviceiSCSIAttributes

Constructors

DeviceiSCSIAttributes

Fields

_dscsiaTargetARN :: Maybe Text

_dscsiaChapEnabled :: Maybe Bool

_dscsiaNetworkInterfaceId :: Maybe Text

_dscsiaNetworkInterfacePort :: Maybe Integer

data DiskInformation

Constructors

DiskInformation

Fields

_dwDiskAllocationResource :: Maybe Text

_dwDiskAllocationType :: Maybe Text

_dwDiskNode :: Maybe Text

_dwDiskPath :: Maybe Text

_dwDiskSizeInBytes :: Maybe Integer

_dwDiskId :: Maybe Text

data GatewayInformation

Constructors

GatewayInformation

Fields

_gjGatewayARN :: Maybe Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

_gjGatewayType :: Maybe Text

data NetworkInterface

Describes a gateway's network interface.

Constructors

NetworkInterface

Fields

_niIpv6Address :: Maybe Text

The Internet Protocol version 6 (IPv6) address of the interface. Currently not supported.

_niMacAddress :: Maybe Text

The Media Access Control (MAC) address of the interface. This is currently unsupported and will not be returned in output.

_niIpv4Address :: Maybe Text

The Internet Protocol version 4 (IPv4) address of the interface.

data StorageGatewayError

A StorageGatewayError that provides more detail about the cause of the error.

Constructors

StorageGatewayError

Fields

_sgeErrorDetails :: Map Text Text

Human-readable text that provides detail about the error that occured.

_sgeErrorCode :: Maybe ErrorCode

Additional information about the error.

data StorediSCSIVolumeInformation

Constructors

StorediSCSIVolumeInformation

Fields

_sscsiviVolumeiSCSIAttributes :: Maybe VolumeiSCSIAttributes

Lists iSCSI information about a volume.

_sscsiviVolumeStatus :: Maybe Text

_sscsiviSourceSnapshotId :: Maybe Text

_sscsiviPreservedExistingData :: Maybe Bool

_sscsiviVolumeARN :: Maybe Text

_sscsiviVolumeProgress :: Maybe Double

_sscsiviVolumeSizeInBytes :: Maybe Integer

_sscsiviVolumeId :: Maybe Text

_sscsiviVolumeDiskId :: Maybe Text

_sscsiviVolumeType :: Maybe Text

data Tape

Constructors

Tape

Fields

_vTapeBarcode :: Maybe Text

_vTapeStatus :: Maybe Text

_vTapeARN :: Maybe Text

_vProgress :: Maybe Double

_vTapeSizeInBytes :: Maybe Integer

_vVTLDevice :: Maybe Text

data TapeArchive

Constructors

TapeArchive

Fields

_tcTapeBarcode :: Maybe Text

_tcTapeStatus :: Maybe Text

_tcTapeARN :: Maybe Text

_tcTapeSizeInBytes :: Maybe Integer

_tcCompletionTime :: Maybe ISO8601

_tcRetrievedTo :: Maybe Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

data TapeRecoveryPointInfo

Constructors

TapeRecoveryPointInfo

Fields

_trpjTapeRecoveryPointTime :: Maybe ISO8601

_trpjTapeARN :: Maybe Text

_trpjTapeSizeInBytes :: Maybe Integer

data VTLDevice

Constructors

VTLDevice

Fields

_vtleDeviceiSCSIAttributes :: Maybe DeviceiSCSIAttributes

_vtleVTLDeviceARN :: Maybe Text

data VolumeInformation

Constructors

VolumeInformation

Fields

_vlVolumeARN :: Maybe Text

_vlVolumeType :: Maybe Text

data VolumeRecoveryPointInfo

Constructors

VolumeRecoveryPointInfo

Fields

_vrpjVolumeRecoveryPointTime :: Maybe Text

_vrpjVolumeARN :: Maybe Text

_vrpjVolumeSizeInBytes :: Maybe Integer

_vrpjVolumeUsageInBytes :: Maybe Integer

data VolumeiSCSIAttributes

Lists iSCSI information about a volume.

Constructors

VolumeiSCSIAttributes

Fields

_vscsiaLunNumber :: Maybe Integer

The logical disk number.

_vscsiaTargetARN :: Maybe Text

The Amazon Resource Name (ARN) of the volume target.

_vscsiaChapEnabled :: Maybe Bool

Indicates whether mutual CHAP is enabled for the iSCSI target.

_vscsiaNetworkInterfaceId :: Maybe Text

The network interface identifier.

_vscsiaNetworkInterfacePort :: Maybe Integer

The port used to communicate with iSCSI targets.

cscsiviVolumeiSCSIAttributes :: Lens' CachediSCSIVolumeInformation (Maybe VolumeiSCSIAttributes)

cscsiviVolumeType :: Lens' CachediSCSIVolumeInformation (Maybe Text)

cscsiviVolumeStatus :: Lens' CachediSCSIVolumeInformation (Maybe Text)

cscsiviVolumeSizeInBytes :: Lens' CachediSCSIVolumeInformation (Maybe Integer)

cscsiviVolumeProgress :: Lens' CachediSCSIVolumeInformation (Maybe Double)

cscsiviVolumeId :: Lens' CachediSCSIVolumeInformation (Maybe Text)

cscsiviVolumeARN :: Lens' CachediSCSIVolumeInformation (Maybe Text)

cscsiviSourceSnapshotId :: Lens' CachediSCSIVolumeInformation (Maybe Text)

ciTargetARN :: Lens' ChapInfo (Maybe Text)

ciSecretToAuthenticateTarget :: Lens' ChapInfo (Maybe Text)

ciSecretToAuthenticateInitiator :: Lens' ChapInfo (Maybe Text)

ciInitiatorName :: Lens' ChapInfo (Maybe Text)

dscsiaTargetARN :: Lens' DeviceiSCSIAttributes (Maybe Text)

dscsiaNetworkInterfacePort :: Lens' DeviceiSCSIAttributes (Maybe Integer)

dscsiaNetworkInterfaceId :: Lens' DeviceiSCSIAttributes (Maybe Text)

dscsiaChapEnabled :: Lens' DeviceiSCSIAttributes (Maybe Bool)

dwDiskSizeInBytes :: Lens' DiskInformation (Maybe Integer)

dwDiskPath :: Lens' DiskInformation (Maybe Text)

dwDiskNode :: Lens' DiskInformation (Maybe Text)

dwDiskId :: Lens' DiskInformation (Maybe Text)

dwDiskAllocationType :: Lens' DiskInformation (Maybe Text)

dwDiskAllocationResource :: Lens' DiskInformation (Maybe Text)

gjGatewayType :: Lens' GatewayInformation (Maybe Text)

gjGatewayARN :: Lens' GatewayInformation (Maybe Text)

niMacAddress :: Lens' NetworkInterface (Maybe Text)

niIpv6Address :: Lens' NetworkInterface (Maybe Text)

niIpv4Address :: Lens' NetworkInterface (Maybe Text)

sgeErrorDetails :: Lens' StorageGatewayError (Map Text Text)

sgeErrorCode :: Lens' StorageGatewayError (Maybe ErrorCode)

sscsiviVolumeiSCSIAttributes :: Lens' StorediSCSIVolumeInformation (Maybe VolumeiSCSIAttributes)

sscsiviVolumeType :: Lens' StorediSCSIVolumeInformation (Maybe Text)

sscsiviVolumeStatus :: Lens' StorediSCSIVolumeInformation (Maybe Text)

sscsiviVolumeSizeInBytes :: Lens' StorediSCSIVolumeInformation (Maybe Integer)

sscsiviVolumeProgress :: Lens' StorediSCSIVolumeInformation (Maybe Double)

sscsiviVolumeId :: Lens' StorediSCSIVolumeInformation (Maybe Text)

sscsiviVolumeDiskId :: Lens' StorediSCSIVolumeInformation (Maybe Text)

sscsiviVolumeARN :: Lens' StorediSCSIVolumeInformation (Maybe Text)

sscsiviSourceSnapshotId :: Lens' StorediSCSIVolumeInformation (Maybe Text)

sscsiviPreservedExistingData :: Lens' StorediSCSIVolumeInformation (Maybe Bool)

vVTLDevice :: Lens' Tape (Maybe Text)

vTapeStatus :: Lens' Tape (Maybe Text)

vTapeSizeInBytes :: Lens' Tape (Maybe Integer)

vTapeBarcode :: Lens' Tape (Maybe Text)

vTapeARN :: Lens' Tape (Maybe Text)

vProgress :: Lens' Tape (Maybe Double)

tcTapeStatus :: Lens' TapeArchive (Maybe Text)

tcTapeSizeInBytes :: Lens' TapeArchive (Maybe Integer)

tcTapeBarcode :: Lens' TapeArchive (Maybe Text)

tcTapeARN :: Lens' TapeArchive (Maybe Text)

tcRetrievedTo :: Lens' TapeArchive (Maybe Text)

tcCompletionTime :: Lens' TapeArchive (Maybe ISO8601)

trpjTapeSizeInBytes :: Lens' TapeRecoveryPointInfo (Maybe Integer)

trpjTapeRecoveryPointTime :: Lens' TapeRecoveryPointInfo (Maybe ISO8601)

trpjTapeARN :: Lens' TapeRecoveryPointInfo (Maybe Text)

vtleVTLDeviceARN :: Lens' VTLDevice (Maybe Text)

vtleDeviceiSCSIAttributes :: Lens' VTLDevice (Maybe DeviceiSCSIAttributes)

vlVolumeType :: Lens' VolumeInformation (Maybe Text)

vlVolumeARN :: Lens' VolumeInformation (Maybe Text)

vrpjVolumeUsageInBytes :: Lens' VolumeRecoveryPointInfo (Maybe Integer)

vrpjVolumeSizeInBytes :: Lens' VolumeRecoveryPointInfo (Maybe Integer)

vrpjVolumeRecoveryPointTime :: Lens' VolumeRecoveryPointInfo (Maybe Text)

vrpjVolumeARN :: Lens' VolumeRecoveryPointInfo (Maybe Text)

vscsiaTargetARN :: Lens' VolumeiSCSIAttributes (Maybe Text)

vscsiaNetworkInterfacePort :: Lens' VolumeiSCSIAttributes (Maybe Integer)

vscsiaNetworkInterfaceId :: Lens' VolumeiSCSIAttributes (Maybe Text)

vscsiaLunNumber :: Lens' VolumeiSCSIAttributes (Maybe Integer)

vscsiaChapEnabled :: Lens' VolumeiSCSIAttributes (Maybe Bool)

updateBandwidthRateLimit

Arguments

:: Text
_ubrliGatewayARN

-> UpdateBandwidthRateLimit

Minimum specification for a UpdateBandwidthRateLimit request.

data UpdateBandwidthRateLimit

Constructors

UpdateBandwidthRateLimit

Fields

_ubrliGatewayARN :: Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

_ubrliAverageDownloadRateLimitInBitsPerSec :: Maybe Integer

The average download bandwidth rate limit in bits per second.

_ubrliAverageUploadRateLimitInBitsPerSec :: Maybe Integer

The average upload bandwidth rate limit in bits per second.

ubrliGatewayARN :: Lens' UpdateBandwidthRateLimit Text

ubrliAverageUploadRateLimitInBitsPerSec :: Lens' UpdateBandwidthRateLimit (Maybe Integer)

ubrliAverageDownloadRateLimitInBitsPerSec :: Lens' UpdateBandwidthRateLimit (Maybe Integer)

data UpdateBandwidthRateLimitResponse

Constructors

UpdateBandwidthRateLimitResponse

Fields

_ubrloGatewayARN :: Maybe Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

ubrloGatewayARN :: Iso' UpdateBandwidthRateLimitResponse (Maybe Text)

updateChapCredentials

Arguments

:: Text
_ucciSecretToAuthenticateInitiator

-> Text
_ucciInitiatorName

-> Text
_ucciTargetARN

-> UpdateChapCredentials

Minimum specification for a UpdateChapCredentials request.

data UpdateChapCredentials

Constructors

UpdateChapCredentials

Fields

_ucciSecretToAuthenticateInitiator :: Text

The secret key that the initiator (e.g. Windows client) must provide to participate in mutual CHAP with the target.

_ucciInitiatorName :: Text

The iSCSI initiator that connects to the target.

_ucciTargetARN :: Text

The Amazon Resource Name (ARN) of the iSCSI volume target. Use the DescribeStorediSCSIVolumes operation to return to retrieve the TargetARN for specified VolumeARN.

_ucciSecretToAuthenticateTarget :: Maybe Text

The secret key that the target must provide to participate in mutual CHAP with the initiator (e.g. Windows client).

ucciTargetARN :: Lens' UpdateChapCredentials Text

ucciSecretToAuthenticateTarget :: Lens' UpdateChapCredentials (Maybe Text)

ucciSecretToAuthenticateInitiator :: Lens' UpdateChapCredentials Text

ucciInitiatorName :: Lens' UpdateChapCredentials Text

data UpdateChapCredentialsResponse

Constructors

UpdateChapCredentialsResponse

Fields

_uccoInitiatorName :: Maybe Text

The iSCSI initiator that connects to the target. This is the same initiator name specified in the request.

_uccoTargetARN :: Maybe Text

The Amazon Resource Name (ARN) of the target. This is the same target specified in the request.

uccoTargetARN :: Lens' UpdateChapCredentialsResponse (Maybe Text)

uccoInitiatorName :: Lens' UpdateChapCredentialsResponse (Maybe Text)

updateGatewayInformation

Arguments

:: Text
_ugiiGatewayARN

-> UpdateGatewayInformation

Minimum specification for a UpdateGatewayInformation request.

data UpdateGatewayInformation

Constructors

UpdateGatewayInformation

Fields

_ugiiGatewayARN :: Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

_ugiiGatewayName :: Maybe Text

A unique identifier for your gateway. This name becomes part of the gateway Amazon Resources Name (ARN) which is what you use as an input to other operations.

_ugiiGatewayTimezone :: Maybe Text

ugiiGatewayTimezone :: Lens' UpdateGatewayInformation (Maybe Text)

ugiiGatewayName :: Lens' UpdateGatewayInformation (Maybe Text)

ugiiGatewayARN :: Lens' UpdateGatewayInformation Text

data UpdateGatewayInformationResponse

Constructors

UpdateGatewayInformationResponse

Fields

_ugioGatewayARN :: Maybe Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

ugioGatewayARN :: Iso' UpdateGatewayInformationResponse (Maybe Text)

data UpdateGatewaySoftwareNow

Constructors

UpdateGatewaySoftwareNow

Fields

_ugsniGatewayARN :: Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

ugsniGatewayARN :: Iso' UpdateGatewaySoftwareNow Text

data UpdateGatewaySoftwareNowResponse

Constructors

UpdateGatewaySoftwareNowResponse

Fields

_ugsnoGatewayARN :: Maybe Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

ugsnoGatewayARN :: Iso' UpdateGatewaySoftwareNowResponse (Maybe Text)

data UpdateMaintenanceStartTime

Constructors

UpdateMaintenanceStartTime

Fields

_umstiDayOfWeek :: Integer

The maintenance start time day of the week.

_umstiGatewayARN :: Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

_umstiHourOfDay :: Integer

The hour component of the maintenance start time represented as hh, where hh is the hour (00 to 23). The hour of the day is in the time zone of the gateway.

_umstiMinuteOfHour :: Integer

The minute component of the maintenance start time represented as mm, where mm is the minute (00 to 59). The minute of the hour is in the time zone of the gateway.

umstiMinuteOfHour :: Lens' UpdateMaintenanceStartTime Integer

umstiHourOfDay :: Lens' UpdateMaintenanceStartTime Integer

umstiGatewayARN :: Lens' UpdateMaintenanceStartTime Text

umstiDayOfWeek :: Lens' UpdateMaintenanceStartTime Integer

data UpdateMaintenanceStartTimeResponse

Constructors

UpdateMaintenanceStartTimeResponse

Fields

_umstoGatewayARN :: Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

umstoGatewayARN :: Iso' UpdateMaintenanceStartTimeResponse Text

updateSnapshotSchedule

Arguments

:: Integer
_ussiStartAt

-> Integer
_ussiRecurrenceInHours

-> Text
_ussiVolumeARN

-> UpdateSnapshotSchedule

Minimum specification for a UpdateSnapshotSchedule request.

data UpdateSnapshotSchedule

Constructors

UpdateSnapshotSchedule

Fields

_ussiStartAt :: Integer

The hour of the day at which the snapshot schedule begins represented as hh, where hh is the hour (0 to 23). The hour of the day is in the time zone of the gateway.

_ussiRecurrenceInHours :: Integer

Frequency of snapshots. Specify the number of hours between snapshots.

_ussiVolumeARN :: Text

The Amazon Resource Name (ARN) of the volume. Use the ListVolumes operation to return a list of gateway volumes.

_ussiDescription :: Maybe Text

Optional description of the snapshot that overwrites the existing description.

ussiVolumeARN :: Lens' UpdateSnapshotSchedule Text

ussiStartAt :: Lens' UpdateSnapshotSchedule Integer

ussiRecurrenceInHours :: Lens' UpdateSnapshotSchedule Integer

ussiDescription :: Lens' UpdateSnapshotSchedule (Maybe Text)

data UpdateSnapshotScheduleResponse

Constructors

UpdateSnapshotScheduleResponse

Fields

_ussoVolumeARN :: Maybe Text

ussoVolumeARN :: Iso' UpdateSnapshotScheduleResponse (Maybe Text)