Network.AWS.StorageGateway.V2013_06_30.AddWorkingStorage

Description

This operation configures one or more gateway local disks as working storage for a gateway. This operation is supported only for the gateway-stored volume architecture. Working storage is also referred to as upload buffer. You can also use the AddUploadBuffer operation to add upload buffer to a stored-volume gateway. In the request, you specify the gateway Amazon Resource Name (ARN) to which you want to add working storage, and one or more disk IDs that you want to configure as working storage. Example Request The following example shows a request that specifies that two local disks of a gateway are to be configured as working storage. POST HTTP1.1 Host: storagegateway.us-east-1.amazonaws.com x-amz-Date: 20120425T120000Z Authorization: CSOC7TJPLR0OOKIRLGOHVAICUFVV4KQNSO5AEMVJF66Q9ASUAAJG Content-type: application/x-amz-json-1.1 x-amz-target: StorageGateway_20120630.AddWorkingStorage { GatewayARN: "arn:aws:storagegateway:us-east-1:111122223333:gateway/mygateway" DiskIds: ["pci-0000:03:00.0-scsi-0:0:0:0", "pci-0000:04:00.0-scsi-1:0:0:0"] } HTTP/1.1 200 OK x-amzn-RequestId: CSOC7TJPLR0OOKIRLGOHVAICUFVV4KQNSO5AEMVJF66Q9ASUAAJG Date: Wed, 25 Apr 2012 12:00:02 GMT Content-type: application/x-amz-json-1.1 Content-length: 80 { GatewayARN: "arn:aws:storagegateway:us-east-1:111122223333:gateway/mygateway" }.

Request

data AddWorkingStorage

A JSON object containing one or more of the following fields: AddWorkingStorageInput$DiskIds.

Request constructor

mkAddWorkingStorage

Arguments

:: Text
awsGatewayARN

-> [Text]
awsDiskIds

-> AddWorkingStorage

Smart constructor for the minimum required parameters to construct a valid AddWorkingStorage request.

The fields accessible through corresponding lenses are:

GatewayARN :: Text

DiskIds :: [Text]

Request lenses

awsGatewayARN :: Lens' AddWorkingStorage Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

awsDiskIds :: Lens' AddWorkingStorage [Text]

An array of strings that identify disks that are to be configured as working storage. Each string have a minimum length of 1 and maximum length of 300. You can get the disk IDs from the ListLocalDisks API.

Response

data AddWorkingStorageResponse

A JSON object containing the of the gateway for which working storage was configured.

Response constructor

mkAddWorkingStorageResponse :: AddWorkingStorageResponse

Smart constructor for the minimum required parameters to construct a valid AddWorkingStorageResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

GatewayARN :: Maybe Text

Response lenses

awsrGatewayARN :: Lens' AddWorkingStorageResponse (Maybe Text)

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.