Network.AWS.StorageGateway.V2013_06_30.CancelArchival

Request

data CancelArchival

Request constructor

mkCancelArchival

Arguments

:: Text
caGatewayARN

-> Text
caTapeARN

-> CancelArchival

Smart constructor for the minimum required parameters to construct a valid CancelArchival request.

Request lenses

caGatewayARN :: Lens' CancelArchival Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

caTapeARN :: Lens' CancelArchival Text

Response

data CancelArchivalResponse

Response lenses

carsTapeARN :: Lens' CancelArchivalResponse (Maybe Text)