Network.AWS.StorageGateway.V2013_06_30.DeleteChapCredentials

Description

This operation deletes Challenge-Handshake Authentication Protocol (CHAP) credentials for a specified iSCSI target and initiator pair. Example Request The following example shows a request that deletes the CHAP credentials for an iSCSI target myvolume. POST HTTP1.1 Host: storagegateway.us-east-1.amazonaws.com x-amz-Date: 20120425T120000Z Authorization: CSOC7TJPLR0OOKIRLGOHVAICUFVV4KQNSO5AEMVJF66Q9ASUAAJG Content-type: application/x-amz-json-1.1 x-amz-target: StorageGateway_20120630.DeleteChapCredentials { TargetARN: "arn:aws:storagegateway:us-east-1:111122223333:gatewaymygatewaytarget/iqn.1997-05.com.amazon:myvolume", InitiatorName: "iqn.1991-05.com.microsoft:computername.domain.example.com" } HTTP/1.1 200 OK x-amzn-RequestId: CSOC7TJPLR0OOKIRLGOHVAICUFVV4KQNSO5AEMVJF66Q9ASUAAJG Date: Wed, 25 Apr 2012 12:00:02 GMT Content-type: application/x-amz-json-1.1 Content-length: 161 { TargetARN: "arn:aws:storagegateway:us-east-1:111122223333:gatewaymygatewaytarget/iqn.1997-05.com.amazon:myvolume", InitiatorName: "iqn.1991-05.com.microsoft:computername.domain.example.com" }.

Request

data DeleteChapCredentials

A JSON object containing one or more of the following fields: DeleteChapCredentialsInput$InitiatorName DeleteChapCredentialsInput$TargetARN.

Request constructor

mkDeleteChapCredentials

Arguments

:: Text
dccTargetARN

-> Text
dccInitiatorName

-> DeleteChapCredentials

Smart constructor for the minimum required parameters to construct a valid DeleteChapCredentials request.

Request lenses

dccTargetARN :: Lens' DeleteChapCredentials Text

The Amazon Resource Name (ARN) of the iSCSI volume target. Use the DescribeStorediSCSIVolumes operation to return to retrieve the TargetARN for specified VolumeARN.

dccInitiatorName :: Lens' DeleteChapCredentials Text

The iSCSI initiator that connects to the target.

Response

data DeleteChapCredentialsResponse

A JSON object containing the following fields:.

Response lenses

dccrsTargetARN :: Lens' DeleteChapCredentialsResponse (Maybe Text)

The Amazon Resource Name (ARN) of the target.

dccrsInitiatorName :: Lens' DeleteChapCredentialsResponse (Maybe Text)

The iSCSI initiator that connects to the target.