Network.AWS.StorageGateway.V2013_06_30.DescribeCache

Description

This operation returns information about the cache of a gateway. This operation is supported only for the gateway-cached volume architecture. The response includes disk IDs that are configured as cache, and it includes the amount of cache allocated and used. Example Request The following example shows a request to obtain a description of a gateway's working storage. POST HTTP1.1 Host: storagegateway.us-east-1.amazonaws.com Content-Type: application/x-amz-json-1.1 Authorization: AWS4-HMAC-SHA256 Credential=AKIAIOSFODNN7EXAMPLE20120425us-east-1storagegatewayaws4_request, SignedHeaders=content-type;host;x-amz-date;x-amz-target, Signature=9cd5a3584d1d67d57e61f120f35102d6b3649066abdd4bf4bbcf05bd9f2f8fe2 x-amz-date: 20120912T120000Z x-amz-target: StorageGateway_20120630.DescribeCache { GatewayARN:"arn:aws:storagegateway:us-east-1:111122223333:gateway/mygateway" } HTTP/1.1 200 OK x-amzn-RequestId: gur28r2rqlgb8vvs0mq17hlgij1q8glle1qeu3kpgg6f0kstauu0 Date: Wed, 12 Sep 2012 12:00:02 GMT Content-Type: application/x-amz-json-1.1 Content-length: 271 { CacheAllocationInBytes: 2199023255552, CacheDirtyPercentage: 0.07, CacheHitPercentage: 99.68, CacheMissPercentage: 0.32, CacheUsedPercentage: 0.07, DiskIds: [ "pci-0000:03:00.0-scsi-0:0:0:0", "pci-0000:04:00.0-scsi-0:1:0:0" ], GatewayARN: "arn:aws:storagegateway:us-east-1:111122223333:gateway/mygateway" }.

Documentation

data DescribeCache

Constructors

DescribeCache

Fields

_dciGatewayARN :: Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

dciGatewayARN :: Iso' DescribeCache Text

data DescribeCacheResponse

Constructors

DescribeCacheResponse

Fields

_dcoDiskIds :: [Text]

_dcoCacheUsedPercentage :: Maybe Double

_dcoCacheHitPercentage :: Maybe Double

_dcoCacheMissPercentage :: Maybe Double

_dcoCacheDirtyPercentage :: Maybe Double

_dcoGatewayARN :: Maybe Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

_dcoCacheAllocatedInBytes :: Maybe Integer

dcoGatewayARN :: Lens' DescribeCacheResponse (Maybe Text)

dcoDiskIds :: Lens' DescribeCacheResponse [Text]

dcoCacheUsedPercentage :: Lens' DescribeCacheResponse (Maybe Double)

dcoCacheMissPercentage :: Lens' DescribeCacheResponse (Maybe Double)

dcoCacheHitPercentage :: Lens' DescribeCacheResponse (Maybe Double)

dcoCacheDirtyPercentage :: Lens' DescribeCacheResponse (Maybe Double)

dcoCacheAllocatedInBytes :: Lens' DescribeCacheResponse (Maybe Integer)