Network.AWS.StorageGateway.V2013_06_30.DescribeSnapshotSchedule

Description

This operation describes the snapshot schedule for the specified gateway volume. The snapshot schedule information includes intervals at which snapshots are automatically initiated on the volume. Example Request The following example shows a request that retrieves the snapshot schedule for a volume. POST HTTP1.1 Host: storagegateway.us-east-1.amazonaws.com x-amz-Date: 20120425T120000Z Authorization: CSOC7TJPLR0OOKIRLGOHVAICUFVV4KQNSO5AEMVJF66Q9ASUAAJG Content-type: application/x-amz-json-1.1 x-amz-target: StorageGateway_20120630.DescribeSnapshotSchedule { VolumeARN: "arn:aws:storagegateway:us-east-1:111122223333:gatewaymygatewayvolume/vol-1122AABB" } HTTP/1.1 200 OK x-amzn-RequestId: CSOC7TJPLR0OOKIRLGOHVAICUFVV4KQNSO5AEMVJF66Q9ASUAAJG Date: Wed, 25 Apr 2012 12:00:02 GMT Content-type: application/x-amz-json-1.1 Content-length: 211 { VolumeARN: "arn:aws:storagegateway:us-east-1:111122223333:gatewaymygatewayvolume/vol-1122AABB", StartAt: 6, RecurrenceInHours: 24, Description: "sgw-AABB1122:vol-AABB1122:Schedule", Timezone: "GMT+7:00" }.

Documentation

data DescribeSnapshotSchedule

Constructors

DescribeSnapshotSchedule

Fields

_dssiVolumeARN :: Text

The Amazon Resource Name (ARN) of the volume. Use the ListVolumes operation to return a list of gateway volumes.

dssiVolumeARN :: Iso' DescribeSnapshotSchedule Text

data DescribeSnapshotScheduleResponse

Constructors

DescribeSnapshotScheduleResponse

Fields

_dssoDescription :: Maybe Text

_dssoTimezone :: Maybe Text

_dssoStartAt :: Maybe Integer

_dssoRecurrenceInHours :: Maybe Integer

_dssoVolumeARN :: Maybe Text

dssoVolumeARN :: Lens' DescribeSnapshotScheduleResponse (Maybe Text)

dssoTimezone :: Lens' DescribeSnapshotScheduleResponse (Maybe Text)

dssoStartAt :: Lens' DescribeSnapshotScheduleResponse (Maybe Integer)

dssoRecurrenceInHours :: Lens' DescribeSnapshotScheduleResponse (Maybe Integer)

dssoDescription :: Lens' DescribeSnapshotScheduleResponse (Maybe Text)