Network.AWS.StorageGateway.V2013_06_30.DescribeTapeArchives

Request

data DescribeTapeArchives

Request constructor

mkDescribeTapeArchives :: DescribeTapeArchives

Smart constructor for the minimum required parameters to construct a valid DescribeTapeArchives request.

Request lenses

dta1TapeARNs :: Lens' DescribeTapeArchives [Text]

dta1Marker :: Lens' DescribeTapeArchives (Maybe Text)

dta1Limit :: Lens' DescribeTapeArchives (Maybe Integer)

Response

data DescribeTapeArchivesResponse

Response constructor

mkDescribeTapeArchivesResponse :: DescribeTapeArchivesResponse

Smart constructor for the minimum required parameters to construct a valid DescribeTapeArchivesResponse response.

This constructor is provided for convenience and testing purposes.

Response lenses

dtarrTapeArchives :: Lens' DescribeTapeArchivesResponse [TapeArchive]

dtarrMarker :: Lens' DescribeTapeArchivesResponse (Maybe Text)