Network.AWS.StorageGateway.V2013_06_30.DescribeTapeRecoveryPoints

Documentation

describeTapeRecoveryPoints

Arguments

:: Text
_dtrpiGatewayARN

-> DescribeTapeRecoveryPoints

Minimum specification for a DescribeTapeRecoveryPoints request.

data DescribeTapeRecoveryPoints

Constructors

DescribeTapeRecoveryPoints

Fields

_dtrpiGatewayARN :: Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

_dtrpiMarker :: Maybe Text

_dtrpiLimit :: Maybe Integer

dtrpiMarker :: Lens' DescribeTapeRecoveryPoints (Maybe Text)

dtrpiLimit :: Lens' DescribeTapeRecoveryPoints (Maybe Integer)

dtrpiGatewayARN :: Lens' DescribeTapeRecoveryPoints Text

data DescribeTapeRecoveryPointsResponse

Constructors

DescribeTapeRecoveryPointsResponse

Fields

_dtrpoGatewayARN :: Maybe Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

_dtrpoMarker :: Maybe Text

_dtrpoTapeRecoveryPointInfos :: [TapeRecoveryPointInfo]

dtrpoTapeRecoveryPointInfos :: Lens' DescribeTapeRecoveryPointsResponse [TapeRecoveryPointInfo]

dtrpoMarker :: Lens' DescribeTapeRecoveryPointsResponse (Maybe Text)

dtrpoGatewayARN :: Lens' DescribeTapeRecoveryPointsResponse (Maybe Text)