Network.AWS.StorageGateway.V2013_06_30.DescribeUploadBuffer

Description

This operation returns information about the upload buffer of a gateway. This operation is supported for both the gateway-stored and gateway-cached volume architectures. The response includes disk IDs that are configured as upload buffer space, and it includes the amount of upload buffer space allocated and used. Example Request The following example shows a request to obtain a description of a gateway's working storage. POST HTTP1.1 Host: storagegateway.us-east-1.amazonaws.com Content-Type: application/x-amz-json-1.1 Authorization: AWS4-HMAC-SHA256 Credential=AKIAIOSFODNN7EXAMPLE20120425us-east-1storagegatewayaws4_request, SignedHeaders=content-type;host;x-amz-date;x-amz-target, Signature=9cd5a3584d1d67d57e61f120f35102d6b3649066abdd4bf4bbcf05bd9f2f8fe2 x-amz-date: 20120912T120000Z x-amz-target: StorageGateway_20120630.DescribeUploadBuffer { GatewayARN:"arn:aws:storagegateway:us-east-1:111122223333:gateway/mygateway" } HTTP/1.1 200 OK x-amzn-RequestId: gur28r2rqlgb8vvs0mq17hlgij1q8glle1qeu3kpgg6f0kstauu0 Date: Wed, 12 Sep 2012 12:00:02 GMT Content-Type: application/x-amz-json-1.1 Content-length: 271 { DiskIds: [ "pci-0000:03:00.0-scsi-0:0:0:0", "pci-0000:04:00.0-scsi-0:1:0:0" ], GatewayARN: "arn:aws:storagegateway:us-east-1:111122223333:gateway/mygateway", UploadBufferAllocatedInBytes: 161061273600, UploadBufferUsedInBytes: 0 }.

Request

data DescribeUploadBuffer

Request constructor

mkDescribeUploadBuffer

Arguments

:: Text
dubGatewayARN

-> DescribeUploadBuffer

Smart constructor for the minimum required parameters to construct a valid DescribeUploadBuffer request.

Request lenses

dubGatewayARN :: Lens' DescribeUploadBuffer Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

Response

data DescribeUploadBufferResponse

Response constructor

mkDescribeUploadBufferResponse :: DescribeUploadBufferResponse

Smart constructor for the minimum required parameters to construct a valid DescribeUploadBufferResponse response.

This constructor is provided for convenience and testing purposes.

Response lenses

dubrGatewayARN :: Lens' DescribeUploadBufferResponse (Maybe Text)

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

dubrDiskIds :: Lens' DescribeUploadBufferResponse [Text]

dubrUploadBufferUsedInBytes :: Lens' DescribeUploadBufferResponse (Maybe Integer)

dubrUploadBufferAllocatedInBytes :: Lens' DescribeUploadBufferResponse (Maybe Integer)