Network.AWS.StorageGateway.V2013_06_30.ShutdownGateway

Description

This operation shuts down a gateway. To specify which gateway to shut down, use the Amazon Resource Name (ARN) of the gateway in the body of your request. The operation shuts down the gateway service component running in the storage gateway's virtual machine (VM) and not the VM. If you want to shut down the VM, it is recommended that you first shut down the gateway component in the VM to avoid unpredictable conditions. After the gateway is shutdown, you cannot call any other API except StartGateway, DescribeGatewayInformation, and ListGateways. For more information, see ActivateGateway. Your applications cannot read from or write to the gateway's storage volumes, and there are no snapshots taken. When you make a shutdown request, you will get a 200 OK success response immediately. However, it might take some time for the gateway to shut down. You can call the DescribeGatewayInformation API to check the status. For more information, see ActivateGateway. If do not intend to use the gateway again, you must delete the gateway (using DeleteGateway) to no longer pay software charges associated with the gateway. Example Request The following example shows a request that shuts down a gateway. POST HTTP1.1 Host: storagegateway.us-east-1.amazonaws.com x-amz-Date: 20120425T120000Z Authorization: CSOC7TJPLR0OOKIRLGOHVAICUFVV4KQNSO5AEMVJF66Q9ASUAAJG Content-type: application/x-amz-json-1.1 x-amz-target: StorageGateway_20120630.ShutdownGateway { GatewayARN: "arn:aws:storagegateway:us-east-1:111122223333:gateway/mygateway" } HTTP/1.1 200 OK x-amzn-RequestId: CSOC7TJPLR0OOKIRLGOHVAICUFVV4KQNSO5AEMVJF66Q9ASUAAJG Date: Wed, 25 Apr 2012 12:00:02 GMT Content-type: application/x-amz-json-1.1 Content-length: 80 { GatewayARN: "arn:aws:storagegateway:us-east-1:111122223333:gateway/mygateway" }.

Request

data ShutdownGateway

A JSON object containing the of the gateway to shut down.

Request constructor

mkShutdownGateway

Arguments

:: Text
sgGatewayARN

-> ShutdownGateway

Smart constructor for the minimum required parameters to construct a valid ShutdownGateway request.

The fields accessible through corresponding lenses are:

GatewayARN :: Text

Request lenses

sgGatewayARN :: Lens' ShutdownGateway Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

Response

data ShutdownGatewayResponse

A JSON object containing the of the gateway that was shut down.

Response constructor

mkShutdownGatewayResponse :: ShutdownGatewayResponse

Smart constructor for the minimum required parameters to construct a valid ShutdownGatewayResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

GatewayARN :: Maybe Text

Response lenses

sgrGatewayARN :: Lens' ShutdownGatewayResponse (Maybe Text)

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.