Network.AWS.StorageGateway.V2013_06_30.StartGateway

Description

This operation starts a gateway that you previously shut down (see ShutdownGateway). After the gateway starts, you can then make other API calls, your applications can read from or write to the gateway's storage volumes and you will be able to take snapshot backups. When you make a request, you will get a 200 OK success response immediately. However, it might take some time for the gateway to be ready. You should call DescribeGatewayInformation and check the status before making any additional API calls. For more information, see ActivateGateway. To specify which gateway to start, use the Amazon Resource Name (ARN) of the gateway in your request. Example Request The following example shows a request that starts a gateway. POST HTTP1.1 Host: storagegateway.us-east-1.amazonaws.com x-amz-Date: 20120425T120000Z Authorization: CSOC7TJPLR0OOKIRLGOHVAICUFVV4KQNSO5AEMVJF66Q9ASUAAJG Content-type: application/x-amz-json-1.1 x-amz-target: StorageGateway_20120630.StartGateway { GatewayARN: "arn:aws:storagegateway:us-east-1:111122223333:gateway/mygateway" } HTTP/1.1 200 OK x-amzn-RequestId: CSOC7TJPLR0OOKIRLGOHVAICUFVV4KQNSO5AEMVJF66Q9ASUAAJG Date: Wed, 25 Apr 2012 12:00:02 GMT Content-type: application/x-amz-json-1.1 Content-length: 80 { GatewayARN: "arn:aws:storagegateway:us-east-1:111122223333:gateway/mygateway" }.

Documentation

data StartGateway

Constructors

StartGateway

Fields

_sgiGatewayARN :: Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

sgiGatewayARN :: Iso' StartGateway Text

data StartGatewayResponse

Constructors

StartGatewayResponse

Fields

_sgoGatewayARN :: Maybe Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

sgoGatewayARN :: Iso' StartGatewayResponse (Maybe Text)