Network.AWS.StorageGateway.V2013_06_30.Types

Description

AWS Storage Gateway is a service that connects an on-premises software appliance with cloud-based storage to provide seamless and secure integration between your on-premises IT environment and AWS's storage infrastructure.

Service

data StorageGateway

Supported version (2013-06-30) of the AWS Storage Gateway service.

ErrorCode

data ErrorCode

Constructors

ErrorCodeActivationKeyExpired
ActivationKeyExpired

ErrorCodeActivationKeyInvalid
ActivationKeyInvalid

ErrorCodeActivationKeyNotFound
ActivationKeyNotFound

ErrorCodeAuthenticationFailure
AuthenticationFailure

ErrorCodeBandwidthThrottleScheduleNotFound
BandwidthThrottleScheduleNotFound

ErrorCodeBlocked
Blocked

ErrorCodeCannotExportSnapshot
CannotExportSnapshot

ErrorCodeChapCredentialNotFound
ChapCredentialNotFound

ErrorCodeDiskAlreadyAllocated
DiskAlreadyAllocated

ErrorCodeDiskDoesNotExist
DiskDoesNotExist

ErrorCodeDiskSizeGreaterThanVolumeMaxSize
DiskSizeGreaterThanVolumeMaxSize

ErrorCodeDiskSizeLessThanVolumeSize
DiskSizeLessThanVolumeSize

ErrorCodeDiskSizeNotGigAligned
DiskSizeNotGigAligned

ErrorCodeDuplicateCertificateInfo
DuplicateCertificateInfo

ErrorCodeDuplicateSchedule
DuplicateSchedule

ErrorCodeEndpointNotFound
EndpointNotFound

ErrorCodeGatewayInternalError
GatewayInternalError

ErrorCodeGatewayNotConnected
GatewayNotConnected

ErrorCodeGatewayNotFound
GatewayNotFound

ErrorCodeGatewayProxyNetworkConnectionBusy
GatewayProxyNetworkConnectionBusy

ErrorCodeIAMNotSupported
IAMNotSupported

ErrorCodeInitiatorInvalid
InitiatorInvalid

ErrorCodeInitiatorNotFound
InitiatorNotFound

ErrorCodeInternalError
InternalError

ErrorCodeInvalidEndpoint
InvalidEndpoint

ErrorCodeInvalidGateway
InvalidGateway

ErrorCodeInvalidParameters
InvalidParameters

ErrorCodeInvalidSchedule
InvalidSchedule

ErrorCodeLocalStorageLimitExceeded
LocalStorageLimitExceeded

ErrorCodeLunAlreadyAllocated
LunAlreadyAllocated

ErrorCodeLunInvalid
LunInvalid

ErrorCodeMaximumContentLengthExceeded
MaximumContentLengthExceeded

ErrorCodeMaximumTapeCartridgeCountExceeded
MaximumTapeCartridgeCountExceeded

ErrorCodeMaximumVolumeCountExceeded
MaximumVolumeCountExceeded

ErrorCodeNetworkConfigurationChanged
NetworkConfigurationChanged

ErrorCodeNoDisksAvailable
NoDisksAvailable

ErrorCodeNotImplemented
NotImplemented

ErrorCodeNotSupported
NotSupported

ErrorCodeOperationAborted
OperationAborted

ErrorCodeOutdatedGateway
OutdatedGateway

ErrorCodeParametersNotImplemented
ParametersNotImplemented

ErrorCodeRegionInvalid
RegionInvalid

ErrorCodeRequestTimeout
RequestTimeout

ErrorCodeServiceUnavailable
ServiceUnavailable

ErrorCodeSnapshotDeleted
SnapshotDeleted

ErrorCodeSnapshotIdInvalid
SnapshotIdInvalid

ErrorCodeSnapshotInProgress
SnapshotInProgress

ErrorCodeSnapshotNotFound
SnapshotNotFound

ErrorCodeSnapshotScheduleNotFound
SnapshotScheduleNotFound

ErrorCodeStagingAreaFull
StagingAreaFull

ErrorCodeStorageFailure
StorageFailure

ErrorCodeTapeCartridgeNotFound
TapeCartridgeNotFound

ErrorCodeTargetAlreadyExists
TargetAlreadyExists

ErrorCodeTargetInvalid
TargetInvalid

ErrorCodeTargetNotFound
TargetNotFound

ErrorCodeUnauthorizedOperation
UnauthorizedOperation

ErrorCodeVolumeAlreadyExists
VolumeAlreadyExists

ErrorCodeVolumeIdInvalid
VolumeIdInvalid

ErrorCodeVolumeInUse
VolumeInUse

ErrorCodeVolumeNotFound
VolumeNotFound

ErrorCodeVolumeNotReady
VolumeNotReady

CachediSCSIVolumeInformation

data CachediSCSIVolumeInformation

mkCachediSCSIVolumeInformation :: CachediSCSIVolumeInformation

Smart constructor for the minimum required fields to construct a valid CachediSCSIVolumeInformation data type to populate a request.

cscsiviVolumeARN :: Lens' CachediSCSIVolumeInformation (Maybe Text)

cscsiviVolumeId :: Lens' CachediSCSIVolumeInformation (Maybe Text)

cscsiviVolumeType :: Lens' CachediSCSIVolumeInformation (Maybe Text)

cscsiviVolumeStatus :: Lens' CachediSCSIVolumeInformation (Maybe Text)

cscsiviVolumeSizeInBytes :: Lens' CachediSCSIVolumeInformation (Maybe Integer)

cscsiviVolumeProgress :: Lens' CachediSCSIVolumeInformation (Maybe Double)

cscsiviSourceSnapshotId :: Lens' CachediSCSIVolumeInformation (Maybe Text)

cscsiviVolumeiSCSIAttributes :: Lens' CachediSCSIVolumeInformation (Maybe VolumeiSCSIAttributes)

Lists iSCSI information about a volume.

ChapInfo

data ChapInfo

Describes Challenge-Handshake Authentication Protocol (CHAP) information that supports authentication between your gateway and iSCSI initiators.

mkChapInfo :: ChapInfo

Smart constructor for the minimum required fields to construct a valid ChapInfo data type to populate a request.

ciTargetARN :: Lens' ChapInfo (Maybe Text)

The Amazon Resource Name (ARN) of the volume. Valid Values: 50 to 500 lowercase letters, numbers, periods (.), and hyphens (-).

ciSecretToAuthenticateInitiator :: Lens' ChapInfo (Maybe Text)

The secret key that the initiator (e.g. Windows client) must provide to participate in mutual CHAP with the target.

ciInitiatorName :: Lens' ChapInfo (Maybe Text)

The iSCSI initiator that connects to the target.

ciSecretToAuthenticateTarget :: Lens' ChapInfo (Maybe Text)

The secret key that the target must provide to participate in mutual CHAP with the initiator (e.g. Windows client).

DeviceiSCSIAttributes

data DeviceiSCSIAttributes

mkDeviceiSCSIAttributes :: DeviceiSCSIAttributes

Smart constructor for the minimum required fields to construct a valid DeviceiSCSIAttributes data type to populate a request.

dscsiaTargetARN :: Lens' DeviceiSCSIAttributes (Maybe Text)

dscsiaNetworkInterfaceId :: Lens' DeviceiSCSIAttributes (Maybe Text)

dscsiaNetworkInterfacePort :: Lens' DeviceiSCSIAttributes (Maybe Integer)

dscsiaChapEnabled :: Lens' DeviceiSCSIAttributes (Maybe Bool)

DiskInformation

data DiskInformation

mkDiskInformation :: DiskInformation

Smart constructor for the minimum required fields to construct a valid DiskInformation data type to populate a request.

diDiskId :: Lens' DiskInformation (Maybe Text)

diDiskPath :: Lens' DiskInformation (Maybe Text)

diDiskNode :: Lens' DiskInformation (Maybe Text)

diDiskSizeInBytes :: Lens' DiskInformation (Maybe Integer)

diDiskAllocationType :: Lens' DiskInformation (Maybe Text)

diDiskAllocationResource :: Lens' DiskInformation (Maybe Text)

GatewayInformation

data GatewayInformation

mkGatewayInformation :: GatewayInformation

Smart constructor for the minimum required fields to construct a valid GatewayInformation data type to populate a request.

giGatewayARN :: Lens' GatewayInformation (Maybe Text)

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

giGatewayType :: Lens' GatewayInformation (Maybe Text)

NetworkInterface

data NetworkInterface

Describes a gateway's network interface.

mkNetworkInterface :: NetworkInterface

Smart constructor for the minimum required fields to construct a valid NetworkInterface data type to populate a request.

niIpv4Address :: Lens' NetworkInterface (Maybe Text)

The Internet Protocol version 4 (IPv4) address of the interface.

niMacAddress :: Lens' NetworkInterface (Maybe Text)

The Media Access Control (MAC) address of the interface. This is currently unsupported and will not be returned in output.

niIpv6Address :: Lens' NetworkInterface (Maybe Text)

The Internet Protocol version 6 (IPv6) address of the interface. Currently not supported.

StorageGatewayError

data StorageGatewayError

A StorageGatewayError that provides more detail about the cause of the error.

mkStorageGatewayError :: StorageGatewayError

Smart constructor for the minimum required fields to construct a valid StorageGatewayError data type to populate a request.

sgeErrorCode :: Lens' StorageGatewayError (Maybe ErrorCode)

Additional information about the error.

sgeErrorDetails :: Lens' StorageGatewayError (Map Text Text)

Human-readable text that provides detail about the error that occured.

StorediSCSIVolumeInformation

data StorediSCSIVolumeInformation

mkStorediSCSIVolumeInformation :: StorediSCSIVolumeInformation

Smart constructor for the minimum required fields to construct a valid StorediSCSIVolumeInformation data type to populate a request.

sscsiviVolumeARN :: Lens' StorediSCSIVolumeInformation (Maybe Text)

sscsiviVolumeId :: Lens' StorediSCSIVolumeInformation (Maybe Text)

sscsiviVolumeType :: Lens' StorediSCSIVolumeInformation (Maybe Text)

sscsiviVolumeStatus :: Lens' StorediSCSIVolumeInformation (Maybe Text)

sscsiviVolumeSizeInBytes :: Lens' StorediSCSIVolumeInformation (Maybe Integer)

sscsiviVolumeProgress :: Lens' StorediSCSIVolumeInformation (Maybe Double)

sscsiviVolumeDiskId :: Lens' StorediSCSIVolumeInformation (Maybe Text)

sscsiviSourceSnapshotId :: Lens' StorediSCSIVolumeInformation (Maybe Text)

sscsiviPreservedExistingData :: Lens' StorediSCSIVolumeInformation (Maybe Bool)

sscsiviVolumeiSCSIAttributes :: Lens' StorediSCSIVolumeInformation (Maybe VolumeiSCSIAttributes)

Lists iSCSI information about a volume.

Tape

data Tape

mkTape :: Tape

Smart constructor for the minimum required fields to construct a valid Tape data type to populate a request.

tTapeARN :: Lens' Tape (Maybe Text)

tTapeBarcode :: Lens' Tape (Maybe Text)

tTapeSizeInBytes :: Lens' Tape (Maybe Integer)

tTapeStatus :: Lens' Tape (Maybe Text)

tVTLDevice :: Lens' Tape (Maybe Text)

tProgress :: Lens' Tape (Maybe Double)

TapeArchive

data TapeArchive

mkTapeArchive :: TapeArchive

Smart constructor for the minimum required fields to construct a valid TapeArchive data type to populate a request.

taTapeARN :: Lens' TapeArchive (Maybe Text)

taTapeBarcode :: Lens' TapeArchive (Maybe Text)

taTapeSizeInBytes :: Lens' TapeArchive (Maybe Integer)

taCompletionTime :: Lens' TapeArchive (Maybe ISO8601)

taRetrievedTo :: Lens' TapeArchive (Maybe Text)

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

taTapeStatus :: Lens' TapeArchive (Maybe Text)

TapeRecoveryPointInfo

data TapeRecoveryPointInfo

mkTapeRecoveryPointInfo :: TapeRecoveryPointInfo

Smart constructor for the minimum required fields to construct a valid TapeRecoveryPointInfo data type to populate a request.

trpiTapeARN :: Lens' TapeRecoveryPointInfo (Maybe Text)

trpiTapeRecoveryPointTime :: Lens' TapeRecoveryPointInfo (Maybe ISO8601)

trpiTapeSizeInBytes :: Lens' TapeRecoveryPointInfo (Maybe Integer)

VTLDevice

data VTLDevice

mkVTLDevice :: VTLDevice

Smart constructor for the minimum required fields to construct a valid VTLDevice data type to populate a request.

vtldVTLDeviceARN :: Lens' VTLDevice (Maybe Text)

vtldDeviceiSCSIAttributes :: Lens' VTLDevice (Maybe DeviceiSCSIAttributes)

VolumeInformation

data VolumeInformation

mkVolumeInformation :: VolumeInformation

Smart constructor for the minimum required fields to construct a valid VolumeInformation data type to populate a request.

viVolumeARN :: Lens' VolumeInformation (Maybe Text)

viVolumeType :: Lens' VolumeInformation (Maybe Text)

VolumeRecoveryPointInfo

data VolumeRecoveryPointInfo

mkVolumeRecoveryPointInfo :: VolumeRecoveryPointInfo

Smart constructor for the minimum required fields to construct a valid VolumeRecoveryPointInfo data type to populate a request.

vrpiVolumeARN :: Lens' VolumeRecoveryPointInfo (Maybe Text)

vrpiVolumeSizeInBytes :: Lens' VolumeRecoveryPointInfo (Maybe Integer)

vrpiVolumeUsageInBytes :: Lens' VolumeRecoveryPointInfo (Maybe Integer)

vrpiVolumeRecoveryPointTime :: Lens' VolumeRecoveryPointInfo (Maybe Text)

VolumeiSCSIAttributes

data VolumeiSCSIAttributes

Lists iSCSI information about a volume.

mkVolumeiSCSIAttributes :: VolumeiSCSIAttributes

Smart constructor for the minimum required fields to construct a valid VolumeiSCSIAttributes data type to populate a request.

vscsiaTargetARN :: Lens' VolumeiSCSIAttributes (Maybe Text)

The Amazon Resource Name (ARN) of the volume target.

vscsiaNetworkInterfaceId :: Lens' VolumeiSCSIAttributes (Maybe Text)

The network interface identifier.

vscsiaNetworkInterfacePort :: Lens' VolumeiSCSIAttributes (Maybe Integer)

The port used to communicate with iSCSI targets.

vscsiaLunNumber :: Lens' VolumeiSCSIAttributes (Maybe Integer)

The logical disk number.

vscsiaChapEnabled :: Lens' VolumeiSCSIAttributes (Maybe Bool)

Indicates whether mutual CHAP is enabled for the iSCSI target.