Network.AWS.StorageGateway.V2013_06_30.UpdateChapCredentials

Description

This operation updates the Challenge-Handshake Authentication Protocol (CHAP) credentials for a specified iSCSI target. By default, a gateway does not have CHAP enabled; however, for added security, you might use it. When you update CHAP credentials, all existing connections on the target are closed and initiators must reconnect with the new credentials. Example Request The following example shows a request that updates CHAP credentials for an iSCSI target. POST HTTP1.1 Host: storagegateway.us-east-1.amazonaws.com x-amz-Date: 20120425T120000Z Authorization: CSOC7TJPLR0OOKIRLGOHVAICUFVV4KQNSO5AEMVJF66Q9ASUAAJG Content-type: application/x-amz-json-1.1 x-amz-target: StorageGateway_20120630.UpdateChapCredentials { TargetARN: "arn:aws:storagegateway:us-east-1:111122223333:gatewaymygatewaytarget/iqn.1997-05.com.amazon:myvolume", SecretToAuthenticateInitiator: "111111111111", InitiatorName: "iqn.1991-05.com.microsoft:computername.domain.example.com", SecretToAuthenticateTarget: "222222222222" } HTTP/1.1 200 OK x-amzn-RequestId: CSOC7TJPLR0OOKIRLGOHVAICUFVV4KQNSO5AEMVJF66Q9ASUAAJG Date: Wed, 25 Apr 2012 12:00:02 GMT Content-type: application/x-amz-json-1.1 Content-length: 161 { TargetARN: "arn:aws:storagegateway:us-east-1:111122223333:gatewaymygatewaytarget/iqn.1997-05.com.amazon:myvolume", InitiatorName: "iqn.1991-05.com.microsoft:computername.domain.example.com" }.

Request

data UpdateChapCredentials

A JSON object containing one or more of the following fields: UpdateChapCredentialsInput$InitiatorName UpdateChapCredentialsInput$SecretToAuthenticateInitiator UpdateChapCredentialsInput$SecretToAuthenticateTarget UpdateChapCredentialsInput$TargetARN.

Request constructor

mkUpdateChapCredentials

Arguments

:: Text
uccTargetARN

-> Text
uccSecretToAuthenticateInitiator

-> Text
uccInitiatorName

-> UpdateChapCredentials

Smart constructor for the minimum required parameters to construct a valid UpdateChapCredentials request.

The fields accessible through corresponding lenses are:

TargetARN :: Text

SecretToAuthenticateInitiator :: Text

InitiatorName :: Text

SecretToAuthenticateTarget :: Maybe Text

Request lenses

uccTargetARN :: Lens' UpdateChapCredentials Text

The Amazon Resource Name (ARN) of the iSCSI volume target. Use the DescribeStorediSCSIVolumes operation to return to retrieve the TargetARN for specified VolumeARN.

uccSecretToAuthenticateInitiator :: Lens' UpdateChapCredentials Text

The secret key that the initiator (e.g. Windows client) must provide to participate in mutual CHAP with the target.

uccInitiatorName :: Lens' UpdateChapCredentials Text

The iSCSI initiator that connects to the target.

uccSecretToAuthenticateTarget :: Lens' UpdateChapCredentials (Maybe Text)

The secret key that the target must provide to participate in mutual CHAP with the initiator (e.g. Windows client).

Response

data UpdateChapCredentialsResponse

A JSON object containing the following fields:.

Response constructor

mkUpdateChapCredentialsResponse :: UpdateChapCredentialsResponse

Smart constructor for the minimum required parameters to construct a valid UpdateChapCredentialsResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

TargetARN :: Maybe Text

InitiatorName :: Maybe Text

Response lenses

uccrTargetARN :: Lens' UpdateChapCredentialsResponse (Maybe Text)

The Amazon Resource Name (ARN) of the target. This is the same target specified in the request.

uccrInitiatorName :: Lens' UpdateChapCredentialsResponse (Maybe Text)

The iSCSI initiator that connects to the target. This is the same initiator name specified in the request.