Network.AWS.StorageGateway.V2013_06_30.UpdateGatewayInformation

Description

This operation updates a gateway's metadata, which includes the gateway's name and time zone. To specify which gateway to update, use the Amazon Resource Name (ARN) of the gateway in your request. Example Request The following example shows a request that updates the name of a gateway. POST HTTP1.1 Host: storagegateway.us-east-1.amazonaws.com x-amz-Date: 20120425T120000Z Authorization: CSOC7TJPLR0OOKIRLGOHVAICUFVV4KQNSO5AEMVJF66Q9ASUAAJG Content-type: application/x-amz-json-1.1 x-amz-target: StorageGateway_20120630.UpdateGatewayInformation { GatewayARN: "arn:aws:storagegateway:us-east-1:111122223333:gateway/mygateway", GatewayName "mygateway2" } HTTP/1.1 200 OK x-amzn-RequestId: CSOC7TJPLR0OOKIRLGOHVAICUFVV4KQNSO5AEMVJF66Q9ASUAAJG Date: Wed, 25 Apr 2012 12:00:02 GMT Content-type: application/x-amz-json-1.1 Content-length: 81 { GatewayARN: "arn:aws:storagegateway:us-east-1:111122223333:gateway/mygateway2" }.

Documentation

updateGatewayInformation

Arguments

:: Text
_ugiiGatewayARN

-> UpdateGatewayInformation

Minimum specification for a UpdateGatewayInformation request.

data UpdateGatewayInformation

Constructors

UpdateGatewayInformation

Fields

_ugiiGatewayARN :: Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

_ugiiGatewayName :: Maybe Text

A unique identifier for your gateway. This name becomes part of the gateway Amazon Resources Name (ARN) which is what you use as an input to other operations.

_ugiiGatewayTimezone :: Maybe Text

ugiiGatewayTimezone :: Lens' UpdateGatewayInformation (Maybe Text)

ugiiGatewayName :: Lens' UpdateGatewayInformation (Maybe Text)

ugiiGatewayARN :: Lens' UpdateGatewayInformation Text

data UpdateGatewayInformationResponse

Constructors

UpdateGatewayInformationResponse

Fields

_ugioGatewayARN :: Maybe Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

ugioGatewayARN :: Iso' UpdateGatewayInformationResponse (Maybe Text)