Network.AWS.StorageGateway.V2013_06_30.UpdateSnapshotSchedule

Description

This operation updates a snapshot schedule configured for a gateway volume. The default snapshot schedule for volume is once every 24 hours, starting at the creation time of the volume. You can use this API to change the shapshot schedule configured for the volume. In the request you must identify the gateway volume whose snapshot schedule you want to update, and the schedule information, including when you want the snapshot to begin on a day and the frequency (in hours) of snapshots. Example Request The following example shows a request that updates a snapshot schedule. POST / HTTP/1.1 Host: storagegateway.us-east-1.amazonaws.com x-amz-Date: 20120425T120000Z Authorization: CSOC7TJPLR0OOKIRLGOHVAICUFVV4KQNSO5AEMVJF66Q9ASUAAJG Content-type: application/x-amz-json-1.1 x-amz-target: StorageGateway_20120630.UpdateSnapshotSchedule { VolumeARN: "arn:aws:storagegateway:us-east-1:111122223333:gatewaymygatewayvolume/vol-1122AABB", StartAt: 0, RecurrenceInHours: 1, Description: "hourly snapshot" } HTTP/1.1 200 OK x-amzn-RequestId: CSOC7TJPLR0OOKIRLGOHVAICUFVV4KQNSO5AEMVJF66Q9ASUAAJG Date: Wed, 25 Apr 2012 12:00:02 GMT Content-type: application/x-amz-json-1.1 Content-length: 99 { VolumeARN: "arn:aws:storagegateway:us-east-1:111122223333:gatewaymygatewayvolume/vol-1122AABB" }.

Request

data UpdateSnapshotSchedule

A JSON object containing one or more of the following fields: UpdateSnapshotScheduleInput$Description UpdateSnapshotScheduleInput$RecurrenceInHours UpdateSnapshotScheduleInput$StartAt UpdateSnapshotScheduleInput$VolumeARN.

Request constructor

mkUpdateSnapshotSchedule

Arguments

:: Text
ussVolumeARN

-> Integer
ussStartAt

-> Integer
ussRecurrenceInHours

-> UpdateSnapshotSchedule

Smart constructor for the minimum required parameters to construct a valid UpdateSnapshotSchedule request.

Request lenses

ussVolumeARN :: Lens' UpdateSnapshotSchedule Text

The Amazon Resource Name (ARN) of the volume. Use the ListVolumes operation to return a list of gateway volumes.

ussStartAt :: Lens' UpdateSnapshotSchedule Integer

The hour of the day at which the snapshot schedule begins represented as hh, where hh is the hour (0 to 23). The hour of the day is in the time zone of the gateway.

ussRecurrenceInHours :: Lens' UpdateSnapshotSchedule Integer

Frequency of snapshots. Specify the number of hours between snapshots.

ussDescription :: Lens' UpdateSnapshotSchedule (Maybe Text)

Optional description of the snapshot that overwrites the existing description.

Response

data UpdateSnapshotScheduleResponse

A JSON object containing the of the updated storage volume.

Response lenses

ussrsVolumeARN :: Lens' UpdateSnapshotScheduleResponse (Maybe Text)