Network.AWS.StorageGateway

Description

AWS Storage Gateway Service

AWS Storage Gateway is the service that connects an on-premises software appliance with cloud-based storage to provide seamless and secure integration between an organization's on-premises IT environment and AWS's storage infrastructure. The service enables you to securely upload data to the AWS cloud for cost effective backup and rapid disaster recovery.

Use the following links to get started using the AWS Storage Gateway Service API Reference :

AWS Storage Gateway Required Request Headers : Describes the required headers that you must send with every POST request to AWS Storage Gateway.

Signing Requests : AWS Storage Gateway requires that you authenticate every request you send; this topic describes how sign such a request.

Error Responses : Provides reference information about AWS Storage Gateway errors.

Operations in AWS Storage Gateway : Contains detailed descriptions of all AWS Storage Gateway operations, their request parameters, response elements, possible errors, and examples of requests and responses.

AWS Storage Gateway Regions and Endpoints: Provides a list of each region and endpoints available for use with AWS Storage Gateway.

Important: IDs for Storage Gateway volumes and Amazon EBS snapshots created from gateway volumes are changing to a longer format. Starting in December 2016, all new volumes and snapshots will be created with a 17-character string. Starting in April 2016, you will be able to use these longer IDs so you can test your systems with the new format. For more information, see Longer EC2 and EBS Resource IDs .

For example, a volume Amazon Resource Name (ARN) with the longer volume ID format looks like the following:

arn:aws:storagegateway:us-west-2:111122223333:gatewaysgw-12A3456Bvolume/vol-1122AABBCCDDEEFFG .

A snapshot ID with the longer ID format looks like the following: snap-78e226633445566ee .

For more information, see Announcement: Heads-up – Longer AWS Storage Gateway volume and snapshot IDs coming in 2016 .

Service Configuration

storageGateway :: Service

API version 2013-06-30 of the Amazon Storage Gateway SDK configuration.

Errors

Error matchers are designed for use with the functions provided by Control.Exception.Lens. This allows catching (and rethrowing) service specific errors returned by StorageGateway.

InvalidGatewayRequestException

_InvalidGatewayRequestException :: AsError a => Getting (First ServiceError) a ServiceError

An exception occurred because an invalid gateway request was issued to the service. For more information, see the error and message fields.

ServiceUnavailableError

_ServiceUnavailableError :: AsError a => Getting (First ServiceError) a ServiceError

An internal server error has occurred because the service is unavailable. For more information, see the error and message fields.

InternalServerError

_InternalServerError :: AsError a => Getting (First ServiceError) a ServiceError

An internal server error has occurred during the request. For more information, see the error and message fields.

Waiters

Waiters poll by repeatedly sending a request until some remote success condition configured by the Wait specification is fulfilled. The Wait specification determines how many attempts should be made, in addition to delay and retry strategies.

Operations

Some AWS operations return results that are incomplete and require subsequent requests in order to obtain the entire result set. The process of sending subsequent requests to continue where a previous request left off is called pagination. For example, the ListObjects operation of Amazon S3 returns up to 1000 objects at a time, and you must send subsequent requests with the appropriate Marker in order to retrieve the next page of results.

Operations that have an AWSPager instance can transparently perform subsequent requests, correctly setting Markers and other request facets to iterate through the entire result set of a truncated API operation. Operations which support this have an additional note in the documentation.

Many operations have the ability to filter results on the server side. See the individual operation parameters for details.

CancelArchival

module Network.AWS.StorageGateway.CancelArchival

CreateStorediSCSIVolume

module Network.AWS.StorageGateway.CreateStorediSCSIVolume

CreateNFSFileShare

module Network.AWS.StorageGateway.CreateNFSFileShare

DescribeChapCredentials

module Network.AWS.StorageGateway.DescribeChapCredentials

SetLocalConsolePassword

module Network.AWS.StorageGateway.SetLocalConsolePassword

CreateTapes

module Network.AWS.StorageGateway.CreateTapes

UpdateVTLDeviceType

module Network.AWS.StorageGateway.UpdateVTLDeviceType

CreateCachediSCSIVolume

module Network.AWS.StorageGateway.CreateCachediSCSIVolume

ListFileShares

module Network.AWS.StorageGateway.ListFileShares

DeleteFileShare

module Network.AWS.StorageGateway.DeleteFileShare

ListVolumeInitiators

module Network.AWS.StorageGateway.ListVolumeInitiators

AddUploadBuffer

module Network.AWS.StorageGateway.AddUploadBuffer

ListTagsForResource

module Network.AWS.StorageGateway.ListTagsForResource

UpdateGatewayInformation

module Network.AWS.StorageGateway.UpdateGatewayInformation

DescribeMaintenanceStartTime

module Network.AWS.StorageGateway.DescribeMaintenanceStartTime

DescribeWorkingStorage

module Network.AWS.StorageGateway.DescribeWorkingStorage

DescribeCachediSCSIVolumes

module Network.AWS.StorageGateway.DescribeCachediSCSIVolumes

AddCache

module Network.AWS.StorageGateway.AddCache

StartGateway

module Network.AWS.StorageGateway.StartGateway

ShutdownGateway

module Network.AWS.StorageGateway.ShutdownGateway

UpdateGatewaySoftwareNow

module Network.AWS.StorageGateway.UpdateGatewaySoftwareNow

RemoveTagsFromResource

module Network.AWS.StorageGateway.RemoveTagsFromResource

DeleteChapCredentials

module Network.AWS.StorageGateway.DeleteChapCredentials

UpdateChapCredentials

module Network.AWS.StorageGateway.UpdateChapCredentials

DescribeUploadBuffer

module Network.AWS.StorageGateway.DescribeUploadBuffer

DescribeTapes (Paginated)

module Network.AWS.StorageGateway.DescribeTapes

DescribeStorediSCSIVolumes

module Network.AWS.StorageGateway.DescribeStorediSCSIVolumes

CreateSnapshotFromVolumeRecoveryPoint

module Network.AWS.StorageGateway.CreateSnapshotFromVolumeRecoveryPoint

RetrieveTapeRecoveryPoint

module Network.AWS.StorageGateway.RetrieveTapeRecoveryPoint

AddTagsToResource

module Network.AWS.StorageGateway.AddTagsToResource

DeleteGateway

module Network.AWS.StorageGateway.DeleteGateway

UpdateMaintenanceStartTime

module Network.AWS.StorageGateway.UpdateMaintenanceStartTime

DescribeGatewayInformation

module Network.AWS.StorageGateway.DescribeGatewayInformation

RefreshCache

module Network.AWS.StorageGateway.RefreshCache

UpdateNFSFileShare

module Network.AWS.StorageGateway.UpdateNFSFileShare

RetrieveTapeArchive

module Network.AWS.StorageGateway.RetrieveTapeArchive

DescribeTapeArchives (Paginated)

module Network.AWS.StorageGateway.DescribeTapeArchives

DisableGateway

module Network.AWS.StorageGateway.DisableGateway

DescribeSnapshotSchedule

module Network.AWS.StorageGateway.DescribeSnapshotSchedule

CreateTapeWithBarcode

module Network.AWS.StorageGateway.CreateTapeWithBarcode

DescribeBandwidthRateLimit

module Network.AWS.StorageGateway.DescribeBandwidthRateLimit

DeleteSnapshotSchedule

module Network.AWS.StorageGateway.DeleteSnapshotSchedule

UpdateSnapshotSchedule

module Network.AWS.StorageGateway.UpdateSnapshotSchedule

CreateSnapshot

module Network.AWS.StorageGateway.CreateSnapshot

CancelRetrieval

module Network.AWS.StorageGateway.CancelRetrieval

DescribeVTLDevices (Paginated)

module Network.AWS.StorageGateway.DescribeVTLDevices

DeleteTapeArchive

module Network.AWS.StorageGateway.DeleteTapeArchive

DescribeNFSFileShares

module Network.AWS.StorageGateway.DescribeNFSFileShares

ListVolumeRecoveryPoints

module Network.AWS.StorageGateway.ListVolumeRecoveryPoints

ListTapes

module Network.AWS.StorageGateway.ListTapes

ResetCache

module Network.AWS.StorageGateway.ResetCache

ListGateways (Paginated)

module Network.AWS.StorageGateway.ListGateways

DeleteTape

module Network.AWS.StorageGateway.DeleteTape

ListLocalDisks

module Network.AWS.StorageGateway.ListLocalDisks

ListVolumes (Paginated)

module Network.AWS.StorageGateway.ListVolumes

UpdateBandwidthRateLimit

module Network.AWS.StorageGateway.UpdateBandwidthRateLimit

AddWorkingStorage

module Network.AWS.StorageGateway.AddWorkingStorage

DescribeTapeRecoveryPoints (Paginated)

module Network.AWS.StorageGateway.DescribeTapeRecoveryPoints

DeleteBandwidthRateLimit

module Network.AWS.StorageGateway.DeleteBandwidthRateLimit

ActivateGateway

module Network.AWS.StorageGateway.ActivateGateway

DescribeCache

module Network.AWS.StorageGateway.DescribeCache

DeleteVolume

module Network.AWS.StorageGateway.DeleteVolume

Types

CachediSCSIVolume

data CachediSCSIVolume

Describes an iSCSI cached volume.

See: cachediSCSIVolume smart constructor.

cachediSCSIVolume :: CachediSCSIVolume

Creates a value of CachediSCSIVolume with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cscsivVolumeiSCSIAttributes - An VolumeiSCSIAttributes object that represents a collection of iSCSI attributes for one stored volume.

cscsivVolumeStatus - One of the VolumeStatus values that indicates the state of the storage volume.

cscsivSourceSnapshotId - If the cached volume was created from a snapshot, this field contains the snapshot ID used, e.g. snap-78e22663. Otherwise, this field is not included.

cscsivVolumeARN - The Amazon Resource Name (ARN) of the storage volume.

cscsivVolumeProgress - Represents the percentage complete if the volume is restoring or bootstrapping that represents the percent of data transferred. This field does not appear in the response if the cached volume is not restoring or bootstrapping.

cscsivVolumeSizeInBytes - The size, in bytes, of the volume capacity.

cscsivCreatedDate - The date the volume was created. Volumes created prior to March 28, 2017 don’t have this time stamp.

cscsivVolumeId - The unique identifier of the volume, e.g. vol-AE4B946D.

cscsivVolumeType - One of the VolumeType enumeration values that describes the type of the volume.

cscsivVolumeiSCSIAttributes :: Lens' CachediSCSIVolume (Maybe VolumeiSCSIAttributes)

An VolumeiSCSIAttributes object that represents a collection of iSCSI attributes for one stored volume.

cscsivVolumeStatus :: Lens' CachediSCSIVolume (Maybe Text)

One of the VolumeStatus values that indicates the state of the storage volume.

cscsivSourceSnapshotId :: Lens' CachediSCSIVolume (Maybe Text)

If the cached volume was created from a snapshot, this field contains the snapshot ID used, e.g. snap-78e22663. Otherwise, this field is not included.

cscsivVolumeARN :: Lens' CachediSCSIVolume (Maybe Text)

The Amazon Resource Name (ARN) of the storage volume.

cscsivVolumeProgress :: Lens' CachediSCSIVolume (Maybe Double)

Represents the percentage complete if the volume is restoring or bootstrapping that represents the percent of data transferred. This field does not appear in the response if the cached volume is not restoring or bootstrapping.

cscsivVolumeSizeInBytes :: Lens' CachediSCSIVolume (Maybe Integer)

The size, in bytes, of the volume capacity.

cscsivCreatedDate :: Lens' CachediSCSIVolume (Maybe UTCTime)

The date the volume was created. Volumes created prior to March 28, 2017 don’t have this time stamp.

cscsivVolumeId :: Lens' CachediSCSIVolume (Maybe Text)

The unique identifier of the volume, e.g. vol-AE4B946D.

cscsivVolumeType :: Lens' CachediSCSIVolume (Maybe Text)

One of the VolumeType enumeration values that describes the type of the volume.

ChapInfo

data ChapInfo

Describes Challenge-Handshake Authentication Protocol (CHAP) information that supports authentication between your gateway and iSCSI initiators.

See: chapInfo smart constructor.

chapInfo :: ChapInfo

Creates a value of ChapInfo with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ciTargetARN - The Amazon Resource Name (ARN) of the volume. Valid Values: 50 to 500 lowercase letters, numbers, periods (.), and hyphens (-).

ciSecretToAuthenticateInitiator - The secret key that the initiator (for example, the Windows client) must provide to participate in mutual CHAP with the target.

ciInitiatorName - The iSCSI initiator that connects to the target.

ciSecretToAuthenticateTarget - The secret key that the target must provide to participate in mutual CHAP with the initiator (e.g. Windows client).

ciTargetARN :: Lens' ChapInfo (Maybe Text)

The Amazon Resource Name (ARN) of the volume. Valid Values: 50 to 500 lowercase letters, numbers, periods (.), and hyphens (-).

ciSecretToAuthenticateInitiator :: Lens' ChapInfo (Maybe Text)

The secret key that the initiator (for example, the Windows client) must provide to participate in mutual CHAP with the target.

ciInitiatorName :: Lens' ChapInfo (Maybe Text)

The iSCSI initiator that connects to the target.

ciSecretToAuthenticateTarget :: Lens' ChapInfo (Maybe Text)

The secret key that the target must provide to participate in mutual CHAP with the initiator (e.g. Windows client).

DeviceiSCSIAttributes

data DeviceiSCSIAttributes

Lists iSCSI information about a VTL device.

See: deviceiSCSIAttributes smart constructor.

deviceiSCSIAttributes :: DeviceiSCSIAttributes

Creates a value of DeviceiSCSIAttributes with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dscsiaTargetARN - Specifies the unique Amazon Resource Name(ARN) that encodes the iSCSI qualified name(iqn) of a tape drive or media changer target.

dscsiaChapEnabled - Indicates whether mutual CHAP is enabled for the iSCSI target.

dscsiaNetworkInterfaceId - The network interface identifier of the VTL device.

dscsiaNetworkInterfacePort - The port used to communicate with iSCSI VTL device targets.

dscsiaTargetARN :: Lens' DeviceiSCSIAttributes (Maybe Text)

Specifies the unique Amazon Resource Name(ARN) that encodes the iSCSI qualified name(iqn) of a tape drive or media changer target.

dscsiaChapEnabled :: Lens' DeviceiSCSIAttributes (Maybe Bool)

Indicates whether mutual CHAP is enabled for the iSCSI target.

dscsiaNetworkInterfaceId :: Lens' DeviceiSCSIAttributes (Maybe Text)

The network interface identifier of the VTL device.

dscsiaNetworkInterfacePort :: Lens' DeviceiSCSIAttributes (Maybe Int)

The port used to communicate with iSCSI VTL device targets.

Disk

data Disk

See: disk smart constructor.

disk :: Disk

Creates a value of Disk with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dDiskAllocationResource - Undocumented member.

dDiskAllocationType - Undocumented member.

dDiskNode - Undocumented member.

dDiskPath - Undocumented member.

dDiskSizeInBytes - Undocumented member.

dDiskStatus - Undocumented member.

dDiskId - Undocumented member.

dDiskAllocationResource :: Lens' Disk (Maybe Text)

Undocumented member.

dDiskAllocationType :: Lens' Disk (Maybe Text)

Undocumented member.

dDiskNode :: Lens' Disk (Maybe Text)

Undocumented member.

dDiskPath :: Lens' Disk (Maybe Text)

Undocumented member.

dDiskSizeInBytes :: Lens' Disk (Maybe Integer)

Undocumented member.

dDiskStatus :: Lens' Disk (Maybe Text)

Undocumented member.

dDiskId :: Lens' Disk (Maybe Text)

Undocumented member.

FileShareInfo

data FileShareInfo

Describes a file share.

See: fileShareInfo smart constructor.

fileShareInfo :: FileShareInfo

Creates a value of FileShareInfo with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

fsiFileShareStatus - Undocumented member.

fsiGatewayARN - Undocumented member.

fsiFileShareId - Undocumented member.

fsiFileShareARN - Undocumented member.

fsiFileShareStatus :: Lens' FileShareInfo (Maybe Text)

Undocumented member.

fsiGatewayARN :: Lens' FileShareInfo (Maybe Text)

Undocumented member.

fsiFileShareId :: Lens' FileShareInfo (Maybe Text)

Undocumented member.

fsiFileShareARN :: Lens' FileShareInfo (Maybe Text)

Undocumented member.

GatewayInfo

data GatewayInfo

Describes a gateway object.

See: gatewayInfo smart constructor.

gatewayInfo :: GatewayInfo

Creates a value of GatewayInfo with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

giGatewayARN - The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

giGatewayOperationalState - The state of the gateway. Valid Values: DISABLED or ACTIVE

giGatewayName - The name of the gateway.

giGatewayId - The unique identifier assigned to your gateway during activation. This ID becomes part of the gateway Amazon Resource Name (ARN), which you use as input for other operations.

giGatewayType - The type of the gateway.

giGatewayARN :: Lens' GatewayInfo (Maybe Text)

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

giGatewayOperationalState :: Lens' GatewayInfo (Maybe Text)

The state of the gateway. Valid Values: DISABLED or ACTIVE

giGatewayName :: Lens' GatewayInfo (Maybe Text)

The name of the gateway.

giGatewayId :: Lens' GatewayInfo (Maybe Text)

The unique identifier assigned to your gateway during activation. This ID becomes part of the gateway Amazon Resource Name (ARN), which you use as input for other operations.

giGatewayType :: Lens' GatewayInfo (Maybe Text)

The type of the gateway.

NFSFileShareDefaults

data NFSFileShareDefaults

Describes file share default values. Files and folders stored as Amazon S3 objects in S3 buckets don't, by default, have Unix file permissions assigned to them. Upon discovery in an S3 bucket by Storage Gateway, the S3 objects that represent files and folders are assigned these default Unix permissions. This operation is only supported in the file gateway architecture.

See: nFSFileShareDefaults smart constructor.

nFSFileShareDefaults :: NFSFileShareDefaults

Creates a value of NFSFileShareDefaults with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

nfsfsdFileMode - The Unix file mode in the form "nnnn". For example, "0666" represents the default file mode inside the file share. The default value is 0666.

nfsfsdOwnerId - The default owner ID for files in the file share (unless the files have another owner ID specified). The default value is nfsnobody.

nfsfsdDirectoryMode - The Unix directory mode in the form "nnnn". For example, "0666" represents the default access mode for all directories inside the file share. The default value is 0777.

nfsfsdGroupId - The default group ID for the file share (unless the files have another group ID specified). The default value is nfsnobody.

nfsfsdFileMode :: Lens' NFSFileShareDefaults (Maybe Text)

The Unix file mode in the form "nnnn". For example, "0666" represents the default file mode inside the file share. The default value is 0666.

nfsfsdOwnerId :: Lens' NFSFileShareDefaults (Maybe Natural)

The default owner ID for files in the file share (unless the files have another owner ID specified). The default value is nfsnobody.

nfsfsdDirectoryMode :: Lens' NFSFileShareDefaults (Maybe Text)

The Unix directory mode in the form "nnnn". For example, "0666" represents the default access mode for all directories inside the file share. The default value is 0777.

nfsfsdGroupId :: Lens' NFSFileShareDefaults (Maybe Natural)

The default group ID for the file share (unless the files have another group ID specified). The default value is nfsnobody.

NFSFileShareInfo

data NFSFileShareInfo

The Unix file permissions and ownership information assigned, by default, to native S3 objects when file gateway discovers them in S3 buckets. This operation is only supported in file gateways.

See: nFSFileShareInfo smart constructor.

nFSFileShareInfo :: NFSFileShareInfo

Creates a value of NFSFileShareInfo with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

nfsfsiFileShareStatus - Undocumented member.

nfsfsiKMSKey - Undocumented member.

nfsfsiGatewayARN - Undocumented member.

nfsfsiPath - Undocumented member.

nfsfsiKMSEncrypted - True to use Amazon S3 server side encryption with your own KMS key, or false to use a key managed by Amazon S3. Optional.

nfsfsiFileShareId - Undocumented member.

nfsfsiFileShareARN - Undocumented member.

nfsfsiDefaultStorageClass - The default storage class for objects put into an Amazon S3 bucket by file gateway. Possible values are S3_STANDARD or S3_STANDARD_IA. If this field is not populated, the default value S3_STANDARD is used. Optional.

nfsfsiRole - Undocumented member.

nfsfsiSquash - Undocumented member.

nfsfsiNFSFileShareDefaults - Undocumented member.

nfsfsiLocationARN - Undocumented member.

nfsfsiClientList - Undocumented member.

nfsfsiReadOnly - Undocumented member.

nfsfsiFileShareStatus :: Lens' NFSFileShareInfo (Maybe Text)

Undocumented member.

nfsfsiKMSKey :: Lens' NFSFileShareInfo (Maybe Text)

Undocumented member.

nfsfsiGatewayARN :: Lens' NFSFileShareInfo (Maybe Text)

Undocumented member.

nfsfsiPath :: Lens' NFSFileShareInfo (Maybe Text)

Undocumented member.

nfsfsiKMSEncrypted :: Lens' NFSFileShareInfo (Maybe Bool)

True to use Amazon S3 server side encryption with your own KMS key, or false to use a key managed by Amazon S3. Optional.

nfsfsiFileShareId :: Lens' NFSFileShareInfo (Maybe Text)

Undocumented member.

nfsfsiFileShareARN :: Lens' NFSFileShareInfo (Maybe Text)

Undocumented member.

nfsfsiDefaultStorageClass :: Lens' NFSFileShareInfo (Maybe Text)

The default storage class for objects put into an Amazon S3 bucket by file gateway. Possible values are S3_STANDARD or S3_STANDARD_IA. If this field is not populated, the default value S3_STANDARD is used. Optional.

nfsfsiRole :: Lens' NFSFileShareInfo (Maybe Text)

Undocumented member.

nfsfsiSquash :: Lens' NFSFileShareInfo (Maybe Text)

Undocumented member.

nfsfsiNFSFileShareDefaults :: Lens' NFSFileShareInfo (Maybe NFSFileShareDefaults)

Undocumented member.

nfsfsiLocationARN :: Lens' NFSFileShareInfo (Maybe Text)

Undocumented member.

nfsfsiClientList :: Lens' NFSFileShareInfo (Maybe (NonEmpty Text))

Undocumented member.

nfsfsiReadOnly :: Lens' NFSFileShareInfo (Maybe Bool)

Undocumented member.

NetworkInterface

data NetworkInterface

Describes a gateway's network interface.

See: networkInterface smart constructor.

networkInterface :: NetworkInterface

Creates a value of NetworkInterface with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

niIPv6Address - The Internet Protocol version 6 (IPv6) address of the interface. Currently not supported .

niMACAddress - The Media Access Control (MAC) address of the interface.

niIPv4Address - The Internet Protocol version 4 (IPv4) address of the interface.

niIPv6Address :: Lens' NetworkInterface (Maybe Text)

The Internet Protocol version 6 (IPv6) address of the interface. Currently not supported .

niMACAddress :: Lens' NetworkInterface (Maybe Text)

The Media Access Control (MAC) address of the interface.

niIPv4Address :: Lens' NetworkInterface (Maybe Text)

The Internet Protocol version 4 (IPv4) address of the interface.

StorediSCSIVolume

data StorediSCSIVolume

Describes an iSCSI stored volume.

See: storediSCSIVolume smart constructor.

storediSCSIVolume :: StorediSCSIVolume

Creates a value of StorediSCSIVolume with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sscsivVolumeiSCSIAttributes - An VolumeiSCSIAttributes object that represents a collection of iSCSI attributes for one stored volume.

sscsivVolumeStatus - One of the VolumeStatus values that indicates the state of the storage volume.

sscsivSourceSnapshotId - If the stored volume was created from a snapshot, this field contains the snapshot ID used, e.g. snap-78e22663. Otherwise, this field is not included.

sscsivPreservedExistingData - Indicates if when the stored volume was created, existing data on the underlying local disk was preserved. Valid Values: true, false

sscsivVolumeARN - The Amazon Resource Name (ARN) of the storage volume.

sscsivVolumeProgress - Represents the percentage complete if the volume is restoring or bootstrapping that represents the percent of data transferred. This field does not appear in the response if the stored volume is not restoring or bootstrapping.

sscsivVolumeSizeInBytes - The size of the volume in bytes.

sscsivCreatedDate - The date the volume was created. Volumes created prior to March 28, 2017 don’t have this time stamp.

sscsivVolumeId - The unique identifier of the volume, e.g. vol-AE4B946D.

sscsivVolumeDiskId - The ID of the local disk that was specified in the CreateStorediSCSIVolume operation.

sscsivVolumeType - One of the VolumeType enumeration values describing the type of the volume.

sscsivVolumeiSCSIAttributes :: Lens' StorediSCSIVolume (Maybe VolumeiSCSIAttributes)

An VolumeiSCSIAttributes object that represents a collection of iSCSI attributes for one stored volume.

sscsivVolumeStatus :: Lens' StorediSCSIVolume (Maybe Text)

One of the VolumeStatus values that indicates the state of the storage volume.

sscsivSourceSnapshotId :: Lens' StorediSCSIVolume (Maybe Text)

If the stored volume was created from a snapshot, this field contains the snapshot ID used, e.g. snap-78e22663. Otherwise, this field is not included.

sscsivPreservedExistingData :: Lens' StorediSCSIVolume (Maybe Bool)

Indicates if when the stored volume was created, existing data on the underlying local disk was preserved. Valid Values: true, false

sscsivVolumeARN :: Lens' StorediSCSIVolume (Maybe Text)

The Amazon Resource Name (ARN) of the storage volume.

sscsivVolumeProgress :: Lens' StorediSCSIVolume (Maybe Double)

Represents the percentage complete if the volume is restoring or bootstrapping that represents the percent of data transferred. This field does not appear in the response if the stored volume is not restoring or bootstrapping.

sscsivVolumeSizeInBytes :: Lens' StorediSCSIVolume (Maybe Integer)

The size of the volume in bytes.

sscsivCreatedDate :: Lens' StorediSCSIVolume (Maybe UTCTime)

The date the volume was created. Volumes created prior to March 28, 2017 don’t have this time stamp.

sscsivVolumeId :: Lens' StorediSCSIVolume (Maybe Text)

The unique identifier of the volume, e.g. vol-AE4B946D.

sscsivVolumeDiskId :: Lens' StorediSCSIVolume (Maybe Text)

The ID of the local disk that was specified in the CreateStorediSCSIVolume operation.

sscsivVolumeType :: Lens' StorediSCSIVolume (Maybe Text)

One of the VolumeType enumeration values describing the type of the volume.

Tag

data Tag

See: tag smart constructor.

tag

Arguments

:: Text
tagKey

-> Text
tagValue

-> Tag

Creates a value of Tag with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tagKey - Undocumented member.

tagValue - Undocumented member.

tagKey :: Lens' Tag Text

Undocumented member.

tagValue :: Lens' Tag Text

Undocumented member.

Tape

data Tape

Describes a virtual tape object.

See: tape smart constructor.

tape :: Tape

Creates a value of Tape with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tTapeBarcode - The barcode that identifies a specific virtual tape.

tTapeStatus - The current state of the virtual tape.

tTapeARN - The Amazon Resource Name (ARN) of the virtual tape.

tProgress - For archiving virtual tapes, indicates how much data remains to be uploaded before archiving is complete. Range: 0 (not started) to 100 (complete).

tTapeSizeInBytes - The size, in bytes, of the virtual tape capacity.

tVTLDevice - The virtual tape library (VTL) device that the virtual tape is associated with.

tTapeUsedInBytes - The size, in bytes, of data written to the virtual tape.

tTapeCreatedDate - The date the virtual tape was created.

tTapeBarcode :: Lens' Tape (Maybe Text)

The barcode that identifies a specific virtual tape.

tTapeStatus :: Lens' Tape (Maybe Text)

The current state of the virtual tape.

tTapeARN :: Lens' Tape (Maybe Text)

The Amazon Resource Name (ARN) of the virtual tape.

tProgress :: Lens' Tape (Maybe Double)

For archiving virtual tapes, indicates how much data remains to be uploaded before archiving is complete. Range: 0 (not started) to 100 (complete).

tTapeSizeInBytes :: Lens' Tape (Maybe Integer)

The size, in bytes, of the virtual tape capacity.

tVTLDevice :: Lens' Tape (Maybe Text)

The virtual tape library (VTL) device that the virtual tape is associated with.

tTapeUsedInBytes :: Lens' Tape (Maybe Integer)

The size, in bytes, of data written to the virtual tape.

tTapeCreatedDate :: Lens' Tape (Maybe UTCTime)

The date the virtual tape was created.

TapeArchive

data TapeArchive

Represents a virtual tape that is archived in the virtual tape shelf (VTS).

See: tapeArchive smart constructor.

tapeArchive :: TapeArchive

Creates a value of TapeArchive with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

taTapeBarcode - The barcode that identifies the archived virtual tape.

taTapeStatus - The current state of the archived virtual tape.

taTapeARN - The Amazon Resource Name (ARN) of an archived virtual tape.

taTapeSizeInBytes - The size, in bytes, of the archived virtual tape.

taCompletionTime - The time that the archiving of the virtual tape was completed. The string format of the completion time is in the ISO8601 extended YYYY-MM-DD'T'HH:MM:SS'Z' format.

taTapeUsedInBytes - The size, in bytes, of data written to the virtual tape.

taTapeCreatedDate - Undocumented member.

taRetrievedTo - The Amazon Resource Name (ARN) of the tape gateway that the virtual tape is being retrieved to. The virtual tape is retrieved from the virtual tape shelf (VTS).

taTapeBarcode :: Lens' TapeArchive (Maybe Text)

The barcode that identifies the archived virtual tape.

taTapeStatus :: Lens' TapeArchive (Maybe Text)

The current state of the archived virtual tape.

taTapeARN :: Lens' TapeArchive (Maybe Text)

The Amazon Resource Name (ARN) of an archived virtual tape.

taTapeSizeInBytes :: Lens' TapeArchive (Maybe Integer)

The size, in bytes, of the archived virtual tape.

taCompletionTime :: Lens' TapeArchive (Maybe UTCTime)

The time that the archiving of the virtual tape was completed. The string format of the completion time is in the ISO8601 extended YYYY-MM-DD'T'HH:MM:SS'Z' format.

taTapeUsedInBytes :: Lens' TapeArchive (Maybe Integer)

The size, in bytes, of data written to the virtual tape.

taTapeCreatedDate :: Lens' TapeArchive (Maybe UTCTime)

Undocumented member.

taRetrievedTo :: Lens' TapeArchive (Maybe Text)

The Amazon Resource Name (ARN) of the tape gateway that the virtual tape is being retrieved to. The virtual tape is retrieved from the virtual tape shelf (VTS).

TapeInfo

data TapeInfo

Describes a virtual tape.

See: tapeInfo smart constructor.

tapeInfo :: TapeInfo

Creates a value of TapeInfo with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tiTapeBarcode - The barcode that identifies a specific virtual tape.

tiTapeStatus - The status of the tape.

tiTapeARN - The Amazon Resource Name (ARN) of a virtual tape.

tiGatewayARN - The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

tiTapeSizeInBytes - The size, in bytes, of a virtual tape.

tiTapeBarcode :: Lens' TapeInfo (Maybe Text)

The barcode that identifies a specific virtual tape.

tiTapeStatus :: Lens' TapeInfo (Maybe Text)

The status of the tape.

tiTapeARN :: Lens' TapeInfo (Maybe Text)

The Amazon Resource Name (ARN) of a virtual tape.

tiGatewayARN :: Lens' TapeInfo (Maybe Text)

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

tiTapeSizeInBytes :: Lens' TapeInfo (Maybe Integer)

The size, in bytes, of a virtual tape.

TapeRecoveryPointInfo

data TapeRecoveryPointInfo

Describes a recovery point.

See: tapeRecoveryPointInfo smart constructor.

tapeRecoveryPointInfo :: TapeRecoveryPointInfo

Creates a value of TapeRecoveryPointInfo with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

trpiTapeStatus - Undocumented member.

trpiTapeRecoveryPointTime - The time when the point-in-time view of the virtual tape was replicated for later recovery. The string format of the tape recovery point time is in the ISO8601 extended YYYY-MM-DD'T'HH:MM:SS'Z' format.

trpiTapeARN - The Amazon Resource Name (ARN) of the virtual tape.

trpiTapeSizeInBytes - The size, in bytes, of the virtual tapes to recover.

trpiTapeStatus :: Lens' TapeRecoveryPointInfo (Maybe Text)

Undocumented member.

trpiTapeRecoveryPointTime :: Lens' TapeRecoveryPointInfo (Maybe UTCTime)

The time when the point-in-time view of the virtual tape was replicated for later recovery. The string format of the tape recovery point time is in the ISO8601 extended YYYY-MM-DD'T'HH:MM:SS'Z' format.

trpiTapeARN :: Lens' TapeRecoveryPointInfo (Maybe Text)

The Amazon Resource Name (ARN) of the virtual tape.

trpiTapeSizeInBytes :: Lens' TapeRecoveryPointInfo (Maybe Integer)

The size, in bytes, of the virtual tapes to recover.

VTLDevice

data VTLDevice

Represents a device object associated with a tape gateway.

See: vTLDevice smart constructor.

vTLDevice :: VTLDevice

Creates a value of VTLDevice with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

vtldDeviceiSCSIAttributes - A list of iSCSI information about a VTL device.

vtldVTLDeviceVendor - Undocumented member.

vtldVTLDeviceARN - Specifies the unique Amazon Resource Name (ARN) of the device (tape drive or media changer).

vtldVTLDeviceType - Undocumented member.

vtldVTLDeviceProductIdentifier - Undocumented member.

vtldDeviceiSCSIAttributes :: Lens' VTLDevice (Maybe DeviceiSCSIAttributes)

A list of iSCSI information about a VTL device.

vtldVTLDeviceVendor :: Lens' VTLDevice (Maybe Text)

Undocumented member.

vtldVTLDeviceARN :: Lens' VTLDevice (Maybe Text)

Specifies the unique Amazon Resource Name (ARN) of the device (tape drive or media changer).

vtldVTLDeviceType :: Lens' VTLDevice (Maybe Text)

Undocumented member.

vtldVTLDeviceProductIdentifier :: Lens' VTLDevice (Maybe Text)

Undocumented member.

VolumeInfo

data VolumeInfo

Describes a storage volume object.

See: volumeInfo smart constructor.

volumeInfo :: VolumeInfo

Creates a value of VolumeInfo with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

viGatewayARN - Undocumented member.

viVolumeARN - The Amazon Resource Name (ARN) for the storage volume. For example, the following is a valid ARN: arn:aws:storagegateway:us-east-2:111122223333:gatewaysgw-12A3456Bvolume/vol-1122AABB Valid Values: 50 to 500 lowercase letters, numbers, periods (.), and hyphens (-).

viVolumeSizeInBytes - The size of the volume in bytes. Valid Values: 50 to 500 lowercase letters, numbers, periods (.), and hyphens (-).

viVolumeId - The unique identifier assigned to the volume. This ID becomes part of the volume Amazon Resource Name (ARN), which you use as input for other operations. Valid Values: 50 to 500 lowercase letters, numbers, periods (.), and hyphens (-).

viGatewayId - The unique identifier assigned to your gateway during activation. This ID becomes part of the gateway Amazon Resource Name (ARN), which you use as input for other operations. Valid Values: 50 to 500 lowercase letters, numbers, periods (.), and hyphens (-).

viVolumeType - Undocumented member.

viGatewayARN :: Lens' VolumeInfo (Maybe Text)

Undocumented member.

viVolumeARN :: Lens' VolumeInfo (Maybe Text)

The Amazon Resource Name (ARN) for the storage volume. For example, the following is a valid ARN: arn:aws:storagegateway:us-east-2:111122223333:gatewaysgw-12A3456Bvolume/vol-1122AABB Valid Values: 50 to 500 lowercase letters, numbers, periods (.), and hyphens (-).

viVolumeSizeInBytes :: Lens' VolumeInfo (Maybe Integer)

The size of the volume in bytes. Valid Values: 50 to 500 lowercase letters, numbers, periods (.), and hyphens (-).

viVolumeId :: Lens' VolumeInfo (Maybe Text)

The unique identifier assigned to the volume. This ID becomes part of the volume Amazon Resource Name (ARN), which you use as input for other operations. Valid Values: 50 to 500 lowercase letters, numbers, periods (.), and hyphens (-).

viGatewayId :: Lens' VolumeInfo (Maybe Text)

The unique identifier assigned to your gateway during activation. This ID becomes part of the gateway Amazon Resource Name (ARN), which you use as input for other operations. Valid Values: 50 to 500 lowercase letters, numbers, periods (.), and hyphens (-).

viVolumeType :: Lens' VolumeInfo (Maybe Text)

Undocumented member.

VolumeRecoveryPointInfo

data VolumeRecoveryPointInfo

See: volumeRecoveryPointInfo smart constructor.

volumeRecoveryPointInfo :: VolumeRecoveryPointInfo

Creates a value of VolumeRecoveryPointInfo with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

vrpiVolumeRecoveryPointTime - Undocumented member.

vrpiVolumeARN - Undocumented member.

vrpiVolumeSizeInBytes - Undocumented member.

vrpiVolumeUsageInBytes - Undocumented member.

vrpiVolumeRecoveryPointTime :: Lens' VolumeRecoveryPointInfo (Maybe Text)

Undocumented member.

vrpiVolumeARN :: Lens' VolumeRecoveryPointInfo (Maybe Text)

Undocumented member.

vrpiVolumeSizeInBytes :: Lens' VolumeRecoveryPointInfo (Maybe Integer)

Undocumented member.

vrpiVolumeUsageInBytes :: Lens' VolumeRecoveryPointInfo (Maybe Integer)

Undocumented member.

VolumeiSCSIAttributes

data VolumeiSCSIAttributes

Lists iSCSI information about a volume.

See: volumeiSCSIAttributes smart constructor.

volumeiSCSIAttributes :: VolumeiSCSIAttributes

Creates a value of VolumeiSCSIAttributes with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

vscsiaLunNumber - The logical disk number.

vscsiaTargetARN - The Amazon Resource Name (ARN) of the volume target.

vscsiaChapEnabled - Indicates whether mutual CHAP is enabled for the iSCSI target.

vscsiaNetworkInterfaceId - The network interface identifier.

vscsiaNetworkInterfacePort - The port used to communicate with iSCSI targets.

vscsiaLunNumber :: Lens' VolumeiSCSIAttributes (Maybe Natural)

The logical disk number.

vscsiaTargetARN :: Lens' VolumeiSCSIAttributes (Maybe Text)

The Amazon Resource Name (ARN) of the volume target.

vscsiaChapEnabled :: Lens' VolumeiSCSIAttributes (Maybe Bool)

Indicates whether mutual CHAP is enabled for the iSCSI target.

vscsiaNetworkInterfaceId :: Lens' VolumeiSCSIAttributes (Maybe Text)

The network interface identifier.

vscsiaNetworkInterfacePort :: Lens' VolumeiSCSIAttributes (Maybe Int)

The port used to communicate with iSCSI targets.