Network.AWS.StorageGateway.ActivateGateway

Description

Activates the gateway you previously deployed on your host. For more information, see Activate the AWS Storage Gateway . In the activation process, you specify information such as the region you want to use for storing snapshots or tapes, the time zone for scheduled snapshots the gateway snapshot schedule window, an activation key, and a name for your gateway. The activation process also associates your gateway with your account; for more information, see UpdateGatewayInformation .

Creating a Request

activateGateway

Arguments

:: Text
agActivationKey

-> Text
agGatewayName

-> Text
agGatewayTimezone

-> Text
agGatewayRegion

-> ActivateGateway

Creates a value of ActivateGateway with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

agMediumChangerType - The value that indicates the type of medium changer to use for tape gateway. This field is optional. Valid Values: STK-L700, AWS-Gateway-VTL

agTapeDriveType - The value that indicates the type of tape drive to use for tape gateway. This field is optional. Valid Values: IBM-ULT3580-TD5

agGatewayType - A value that defines the type of gateway to activate. The type specified is critical to all later functions of the gateway and cannot be changed after activation. The default value is STORED . Valid Values: STORED, CACHED, VTL, FILE_S3

agActivationKey - Your gateway activation key. You can obtain the activation key by sending an HTTP GET request with redirects enabled to the gateway IP address (port 80). The redirect URL returned in the response provides you the activation key for your gateway in the query string parameter activationKey . It may also include other activation-related parameters, however, these are merely defaults -- the arguments you pass to the ActivateGateway API call determine the actual configuration of your gateway.

agGatewayName - The name you configured for your gateway.

agGatewayTimezone - A value that indicates the time zone you want to set for the gateway. The time zone is of the format GMT-hr:mm or "GMT+hr:mm". For example, GMT-4:00 indicates the time is 4 hours behind GMT. GMT+2:00 indicates the time is 2 hours ahead of GMT. The time zone is used, for example, for scheduling snapshots and your gateway's maintenance schedule.

agGatewayRegion - A value that indicates the region where you want to store your data. The gateway region specified must be the same region as the region in your Host header in the request. For more information about available regions and endpoints for AWS Storage Gateway, see Regions and Endpoints in the Amazon Web Services Glossary . Valid Values: "us-east-1", "us-east-2", "us-west-1", "us-west-2", "ca-central-1", "eu-west-1", "eu-central-1", "eu-west-2", "ap-northeast-1", "ap-northeast-2", "ap-southeast-1", "ap-southeast-2", "ap-south-1", "sa-east-1"

data ActivateGateway

A JSON object containing one or more of the following fields:

'ActivateGatewayInput$ActivationKey'

'ActivateGatewayInput$GatewayName'

'ActivateGatewayInput$GatewayRegion'

'ActivateGatewayInput$GatewayTimezone'

'ActivateGatewayInput$GatewayType'

'ActivateGatewayInput$TapeDriveType'

'ActivateGatewayInput$MediumChangerType'

See: activateGateway smart constructor.

Request Lenses

agMediumChangerType :: Lens' ActivateGateway (Maybe Text)

The value that indicates the type of medium changer to use for tape gateway. This field is optional. Valid Values: STK-L700, AWS-Gateway-VTL

agTapeDriveType :: Lens' ActivateGateway (Maybe Text)

The value that indicates the type of tape drive to use for tape gateway. This field is optional. Valid Values: IBM-ULT3580-TD5

agGatewayType :: Lens' ActivateGateway (Maybe Text)

A value that defines the type of gateway to activate. The type specified is critical to all later functions of the gateway and cannot be changed after activation. The default value is STORED . Valid Values: STORED, CACHED, VTL, FILE_S3

agActivationKey :: Lens' ActivateGateway Text

Your gateway activation key. You can obtain the activation key by sending an HTTP GET request with redirects enabled to the gateway IP address (port 80). The redirect URL returned in the response provides you the activation key for your gateway in the query string parameter activationKey . It may also include other activation-related parameters, however, these are merely defaults -- the arguments you pass to the ActivateGateway API call determine the actual configuration of your gateway.

agGatewayName :: Lens' ActivateGateway Text

The name you configured for your gateway.

agGatewayTimezone :: Lens' ActivateGateway Text

A value that indicates the time zone you want to set for the gateway. The time zone is of the format GMT-hr:mm or "GMT+hr:mm". For example, GMT-4:00 indicates the time is 4 hours behind GMT. GMT+2:00 indicates the time is 2 hours ahead of GMT. The time zone is used, for example, for scheduling snapshots and your gateway's maintenance schedule.

agGatewayRegion :: Lens' ActivateGateway Text

A value that indicates the region where you want to store your data. The gateway region specified must be the same region as the region in your Host header in the request. For more information about available regions and endpoints for AWS Storage Gateway, see Regions and Endpoints in the Amazon Web Services Glossary . Valid Values: "us-east-1", "us-east-2", "us-west-1", "us-west-2", "ca-central-1", "eu-west-1", "eu-central-1", "eu-west-2", "ap-northeast-1", "ap-northeast-2", "ap-southeast-1", "ap-southeast-2", "ap-south-1", "sa-east-1"

Destructuring the Response

activateGatewayResponse

Arguments

:: Int
agrsResponseStatus

-> ActivateGatewayResponse

Creates a value of ActivateGatewayResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

agrsGatewayARN - Undocumented member.

agrsResponseStatus - -- | The response status code.

data ActivateGatewayResponse

AWS Storage Gateway returns the Amazon Resource Name (ARN) of the activated gateway. It is a string made of information such as your account, gateway name, and region. This ARN is used to reference the gateway in other API operations as well as resource-based authorization.

See: activateGatewayResponse smart constructor.

Response Lenses

agrsGatewayARN :: Lens' ActivateGatewayResponse (Maybe Text)

Undocumented member.

agrsResponseStatus :: Lens' ActivateGatewayResponse Int

- | The response status code.