Network.AWS.StorageGateway.ActivateGateway

Description

This operation activates the gateway you previously deployed on your host. For more information, see Activate the AWS Storage Gateway. In the activation process, you specify information such as the region you want to use for storing snapshots, the time zone for scheduled snapshots the gateway snapshot schedule window, an activation key, and a name for your gateway. The activation process also associates your gateway with your account; for more information, see UpdateGatewayInformation.

You must turn on the gateway VM before you can activate your gateway.

http://docs.aws.amazon.com/storagegateway/latest/APIReference/API_ActivateGateway.html

Request

data ActivateGateway

Request constructor

activateGateway

Arguments

:: Text
agActivationKey

-> Text
agGatewayName

-> Text
agGatewayTimezone

-> Text
agGatewayRegion

-> ActivateGateway

ActivateGateway constructor.

The fields accessible through corresponding lenses are:

agActivationKey :: Text

agGatewayName :: Text

agGatewayRegion :: Text

agGatewayTimezone :: Text

agGatewayType :: Maybe Text

agMediumChangerType :: Maybe Text

agTapeDriveType :: Maybe Text

Request lenses

agActivationKey :: Lens' ActivateGateway Text

Your gateway activation key. You can obtain the activation key by sending an HTTP GET request with redirects enabled to the gateway IP address (port 80). The redirect URL returned in the response provides you the activation key for your gateway in the query string parameter activationKey. It may also include other activation-related parameters, however, these are merely defaults -- the arguments you pass to the ActivateGateway API call determine the actual configuration of your gateway.

agGatewayName :: Lens' ActivateGateway Text

agGatewayRegion :: Lens' ActivateGateway Text

One of the values that indicates the region where you want to store the snapshot backups. The gateway region specified must be the same region as the region in your Host header in the request. For more information about available regions and endpoints for AWS Storage Gateway, see Regions andEndpoints in the Amazon Web Services Glossary.

Valid Values: "us-east-1", "us-west-1", "us-west-2", "eu-west-1", "eu-central-1", "ap-northeast-1", "ap-southeast-1", "ap-southeast-2", "sa-east-1"

agGatewayTimezone :: Lens' ActivateGateway Text

One of the values that indicates the time zone you want to set for the gateway. The time zone is used, for example, for scheduling snapshots and your gateway's maintenance schedule.

agGatewayType :: Lens' ActivateGateway (Maybe Text)

One of the values that defines the type of gateway to activate. The type specified is critical to all later functions of the gateway and cannot be changed after activation. The default value is STORED.

agMediumChangerType :: Lens' ActivateGateway (Maybe Text)

The value that indicates the type of medium changer to use for gateway-VTL. This field is optional.

Valid Values: STK-L700, AWS-Gateway-VTL

agTapeDriveType :: Lens' ActivateGateway (Maybe Text)

The value that indicates the type of tape drive to use for gateway-VTL. This field is optional.

Valid Values: IBM-ULT3580-TD5

Response

data ActivateGatewayResponse

Response constructor

activateGatewayResponse :: ActivateGatewayResponse

ActivateGatewayResponse constructor.

The fields accessible through corresponding lenses are:

agrGatewayARN :: Maybe Text

Response lenses

agrGatewayARN :: Lens' ActivateGatewayResponse (Maybe Text)