Network.AWS.StorageGateway.AddUploadBuffer

Description

Configures one or more gateway local disks as upload buffer for a specified gateway. This operation is supported for the stored volume, cached volume and tape gateway architectures.

In the request, you specify the gateway Amazon Resource Name (ARN) to which you want to add upload buffer, and one or more disk IDs that you want to configure as upload buffer.

Creating a Request

addUploadBuffer

Arguments

:: Text
aubGatewayARN

-> AddUploadBuffer

Creates a value of AddUploadBuffer with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aubGatewayARN - Undocumented member.

aubDiskIds - Undocumented member.

data AddUploadBuffer

See: addUploadBuffer smart constructor.

Request Lenses

aubGatewayARN :: Lens' AddUploadBuffer Text

Undocumented member.

aubDiskIds :: Lens' AddUploadBuffer [Text]

Undocumented member.

Destructuring the Response

addUploadBufferResponse

Arguments

:: Int
aubrsResponseStatus

-> AddUploadBufferResponse

Creates a value of AddUploadBufferResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aubrsGatewayARN - Undocumented member.

aubrsResponseStatus - -- | The response status code.

data AddUploadBufferResponse

See: addUploadBufferResponse smart constructor.

Response Lenses

aubrsGatewayARN :: Lens' AddUploadBufferResponse (Maybe Text)

Undocumented member.

aubrsResponseStatus :: Lens' AddUploadBufferResponse Int

- | The response status code.