Network.AWS.StorageGateway.AddUploadBuffer

Description

This operation configures one or more gateway local disks as upload buffer for a specified gateway. This operation is supported for both the gateway-stored and gateway-cached volume architectures.

In the request, you specify the gateway Amazon Resource Name (ARN) to which you want to add upload buffer, and one or more disk IDs that you want to configure as upload buffer.

http://docs.aws.amazon.com/storagegateway/latest/APIReference/API_AddUploadBuffer.html

Request

data AddUploadBuffer

Request constructor

addUploadBuffer

Arguments

:: Text
aubGatewayARN

-> AddUploadBuffer

AddUploadBuffer constructor.

The fields accessible through corresponding lenses are:

aubDiskIds :: [Text]

aubGatewayARN :: Text

Request lenses

aubDiskIds :: Lens' AddUploadBuffer [Text]

aubGatewayARN :: Lens' AddUploadBuffer Text

Response

data AddUploadBufferResponse

Response constructor

addUploadBufferResponse :: AddUploadBufferResponse

AddUploadBufferResponse constructor.

The fields accessible through corresponding lenses are:

aubrGatewayARN :: Maybe Text

Response lenses

aubrGatewayARN :: Lens' AddUploadBufferResponse (Maybe Text)