Network.AWS.StorageGateway.AddWorkingStorage

Description

Configures one or more gateway local disks as working storage for a gateway. This operation is only supported in the stored volume gateway architecture. This operation is deprecated in cached volume API version 20120630. Use AddUploadBuffer instead.

In the request, you specify the gateway Amazon Resource Name (ARN) to which you want to add working storage, and one or more disk IDs that you want to configure as working storage.

Creating a Request

addWorkingStorage

Arguments

:: Text
awsGatewayARN

-> AddWorkingStorage

Creates a value of AddWorkingStorage with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

awsGatewayARN - Undocumented member.

awsDiskIds - An array of strings that identify disks that are to be configured as working storage. Each string have a minimum length of 1 and maximum length of 300. You can get the disk IDs from the ListLocalDisks API.

data AddWorkingStorage

A JSON object containing one or more of the following fields:

'AddWorkingStorageInput$DiskIds'

See: addWorkingStorage smart constructor.

Request Lenses

awsGatewayARN :: Lens' AddWorkingStorage Text

Undocumented member.

awsDiskIds :: Lens' AddWorkingStorage [Text]

An array of strings that identify disks that are to be configured as working storage. Each string have a minimum length of 1 and maximum length of 300. You can get the disk IDs from the ListLocalDisks API.

Destructuring the Response

addWorkingStorageResponse

Arguments

:: Int
awsrsResponseStatus

-> AddWorkingStorageResponse

Creates a value of AddWorkingStorageResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

awsrsGatewayARN - Undocumented member.

awsrsResponseStatus - -- | The response status code.

data AddWorkingStorageResponse

A JSON object containing the of the gateway for which working storage was configured.

See: addWorkingStorageResponse smart constructor.

Response Lenses

awsrsGatewayARN :: Lens' AddWorkingStorageResponse (Maybe Text)

Undocumented member.

awsrsResponseStatus :: Lens' AddWorkingStorageResponse Int

- | The response status code.