Network.AWS.StorageGateway.CancelRetrieval

Description

Cancels retrieval of a virtual tape from the virtual tape shelf (VTS) to a gateway after the retrieval process is initiated. The virtual tape is returned to the VTS.

Creating a Request

cancelRetrieval

Arguments

:: Text
crGatewayARN

-> Text
crTapeARN

-> CancelRetrieval

Creates a value of CancelRetrieval with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

crGatewayARN - Undocumented member.

crTapeARN - The Amazon Resource Name (ARN) of the virtual tape you want to cancel retrieval for.

data CancelRetrieval

CancelRetrievalInput

See: cancelRetrieval smart constructor.

Request Lenses

crGatewayARN :: Lens' CancelRetrieval Text

Undocumented member.

crTapeARN :: Lens' CancelRetrieval Text

The Amazon Resource Name (ARN) of the virtual tape you want to cancel retrieval for.

Destructuring the Response

cancelRetrievalResponse

Arguments

:: Int
crrsResponseStatus

-> CancelRetrievalResponse

Creates a value of CancelRetrievalResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

crrsTapeARN - The Amazon Resource Name (ARN) of the virtual tape for which retrieval was canceled.

crrsResponseStatus - -- | The response status code.

data CancelRetrievalResponse

CancelRetrievalOutput

See: cancelRetrievalResponse smart constructor.

Response Lenses

crrsTapeARN :: Lens' CancelRetrievalResponse (Maybe Text)

The Amazon Resource Name (ARN) of the virtual tape for which retrieval was canceled.

crrsResponseStatus :: Lens' CancelRetrievalResponse Int

- | The response status code.