Network.AWS.StorageGateway.CancelRetrieval

Description

Cancels retrieval of a virtual tape from the virtual tape shelf (VTS) to a gateway after the retrieval process is initiated. The virtual tape is returned to the VTS.

http://docs.aws.amazon.com/storagegateway/latest/APIReference/API_CancelRetrieval.html

Request

data CancelRetrieval

Request constructor

cancelRetrieval

Arguments

:: Text
crGatewayARN

-> Text
crTapeARN

-> CancelRetrieval

CancelRetrieval constructor.

The fields accessible through corresponding lenses are:

crGatewayARN :: Text

crTapeARN :: Text

Request lenses

crGatewayARN :: Lens' CancelRetrieval Text

crTapeARN :: Lens' CancelRetrieval Text

The Amazon Resource Name (ARN) of the virtual tape you want to cancel retrieval for.

Response

data CancelRetrievalResponse

Response constructor

cancelRetrievalResponse :: CancelRetrievalResponse

CancelRetrievalResponse constructor.

The fields accessible through corresponding lenses are:

crrTapeARN :: Maybe Text

Response lenses

crrTapeARN :: Lens' CancelRetrievalResponse (Maybe Text)

The Amazon Resource Name (ARN) of the virtual tape for which retrieval was canceled.