Network.AWS.StorageGateway.CreateCachediSCSIVolume

Description

This operation creates a cached volume on a specified cached gateway. This operation is supported only for the gateway-cached volume architecture.

Cache storage must be allocated to the gateway before you can create a cached volume. Use the AddCache operation to add cache storage to a gateway. In the request, you must specify the gateway, size of the volume in bytes, the iSCSI target name, an IP address on which to expose the target, and a unique client token. In response, AWS Storage Gateway creates the volume and returns information about it such as the volume Amazon Resource Name (ARN), its size, and the iSCSI target ARN that initiators can use to connect to the volume target.

http://docs.aws.amazon.com/storagegateway/latest/APIReference/API_CreateCachediSCSIVolume.html

Request

data CreateCachediSCSIVolume

Request constructor

createCachediSCSIVolume

Arguments

:: Text
ccscsivGatewayARN

-> Integer
ccscsivVolumeSizeInBytes

-> Text
ccscsivTargetName

-> Text
ccscsivNetworkInterfaceId

-> Text
ccscsivClientToken

-> CreateCachediSCSIVolume

CreateCachediSCSIVolume constructor.

The fields accessible through corresponding lenses are:

ccscsivClientToken :: Text

ccscsivGatewayARN :: Text

ccscsivNetworkInterfaceId :: Text

ccscsivSnapshotId :: Maybe Text

ccscsivTargetName :: Text

ccscsivVolumeSizeInBytes :: Integer

Request lenses

ccscsivClientToken :: Lens' CreateCachediSCSIVolume Text

ccscsivGatewayARN :: Lens' CreateCachediSCSIVolume Text

ccscsivNetworkInterfaceId :: Lens' CreateCachediSCSIVolume Text

ccscsivSnapshotId :: Lens' CreateCachediSCSIVolume (Maybe Text)

ccscsivTargetName :: Lens' CreateCachediSCSIVolume Text

ccscsivVolumeSizeInBytes :: Lens' CreateCachediSCSIVolume Integer

Response

data CreateCachediSCSIVolumeResponse

Response constructor

createCachediSCSIVolumeResponse :: CreateCachediSCSIVolumeResponse

CreateCachediSCSIVolumeResponse constructor.

The fields accessible through corresponding lenses are:

ccscsivrTargetARN :: Maybe Text

ccscsivrVolumeARN :: Maybe Text

Response lenses

ccscsivrTargetARN :: Lens' CreateCachediSCSIVolumeResponse (Maybe Text)

ccscsivrVolumeARN :: Lens' CreateCachediSCSIVolumeResponse (Maybe Text)