Network.AWS.StorageGateway.CreateSnapshot

Description

This operation initiates a snapshot of a volume. AWS Storage Gateway provides the ability to back up point-in-time snapshots of your data to Amazon Simple Storage (S3) for durable off-site recovery, as well as import the data to an Amazon Elastic Block Store (EBS) volume in Amazon Elastic Compute Cloud (EC2). You can take snapshots of your gateway volume on a scheduled or ad-hoc basis. This API enables you to take ad-hoc snapshot. For more information, see Working With Snapshots in the AWS Storage Gateway Console. In the CreateSnapshot request you identify the volume by providing its Amazon Resource Name (ARN). You must also provide description for the snapshot. When AWS Storage Gateway takes the snapshot of specified volume, the snapshot and description appears in the AWS Storage Gateway Console. In response, AWS Storage Gateway returns you a snapshot ID. You can use this snapshot ID to check the snapshot progress or later use it when you want to create a volume from a snapshot. To list or delete a snapshot, you must use the Amazon EC2 API. For more information, . Example Request The following example sends a CreateSnapshot request to take snapshot of the specified an example volume. POST HTTP1.1 Host: storagegateway.us-east-1.amazonaws.com x-amz-Date: 20120425T120000Z Authorization: CSOC7TJPLR0OOKIRLGOHVAICUFVV4KQNSO5AEMVJF66Q9ASUAAJG Content-type: application/x-amz-json-1.1 x-amz-target: StorageGateway_20120630.CreateSnapshot { VolumeARN: "arn:aws:storagegateway:us-east-1:111122223333:gatewaymygatewayvolume/vol-1122AABB", SnapshotDescription: "snapshot description" } HTTP/1.1 200 OK x-amzn-RequestId: CSOC7TJPLR0OOKIRLGOHVAICUFVV4KQNSO5AEMVJF66Q9ASUAAJG Date: Wed, 25 Apr 2012 12:00:02 GMT Content-type: application/x-amz-json-1.1 Content-length: 128 { VolumeARN: "arn:aws:storagegateway:us-east-1:111122223333:gatewaymygatewayvolume/vol-1122AABB", SnapshotId: "snap-78e22663" }.

Request

data CreateSnapshot

A JSON object containing one or more of the following fields: CreateSnapshotInput$SnapshotDescription CreateSnapshotInput$VolumeARN.

Request constructor

createSnapshot

Arguments

:: Text
csVolumeARN

-> Text
csSnapshotDescription

-> CreateSnapshot

Smart constructor for the minimum required parameters to construct a valid CreateSnapshot request.

The fields accessible through corresponding lenses are:

VolumeARN :: Text

SnapshotDescription :: Text

Request lenses

csVolumeARN :: Lens' CreateSnapshot Text

The Amazon Resource Name (ARN) of the volume. Use the ListVolumes operation to return a list of gateway volumes.

csSnapshotDescription :: Lens' CreateSnapshot Text

Textual description of the snapshot that appears in the Amazon EC2 console, Elastic Block Store snapshots panel in the Description field, and in the AWS Storage Gateway snapshot Details pane, Description field.

Response

data CreateSnapshotResponse

A JSON object containing the following fields:.

Response constructor

createSnapshotResponse :: CreateSnapshotResponse

Smart constructor for the minimum required parameters to construct a valid CreateSnapshotResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

VolumeARN :: Maybe Text

SnapshotId :: Maybe Text

Response lenses

csrVolumeARN :: Lens' CreateSnapshotResponse (Maybe Text)

The Amazon Resource Name (ARN) of the volume of which the snapshot was taken.

csrSnapshotId :: Lens' CreateSnapshotResponse (Maybe Text)

The snapshot ID that is used to refer to the snapshot in future operations such as describing snapshots (Amazon Elastic Compute Cloud API DescribeSnapshots) or creating a volume from a snapshot (CreateStorediSCSIVolume).