Network.AWS.StorageGateway.CreateSnapshotFromVolumeRecoveryPoint

Description

Initiates a snapshot of a gateway from a volume recovery point. This operation is only supported in the cached volume gateway architecture.

A volume recovery point is a point in time at which all data of the volume is consistent and from which you can create a snapshot. To get a list of volume recovery point for cached volume gateway, use ListVolumeRecoveryPoints .

In the CreateSnapshotFromVolumeRecoveryPoint request, you identify the volume by providing its Amazon Resource Name (ARN). You must also provide a description for the snapshot. When the gateway takes a snapshot of the specified volume, the snapshot and its description appear in the AWS Storage Gateway console. In response, the gateway returns you a snapshot ID. You can use this snapshot ID to check the snapshot progress or later use it when you want to create a volume from a snapshot.

Creating a Request

createSnapshotFromVolumeRecoveryPoint

Arguments

:: Text
csfvrpVolumeARN

-> Text
csfvrpSnapshotDescription

-> CreateSnapshotFromVolumeRecoveryPoint

Creates a value of CreateSnapshotFromVolumeRecoveryPoint with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

csfvrpVolumeARN - Undocumented member.

csfvrpSnapshotDescription - Undocumented member.

data CreateSnapshotFromVolumeRecoveryPoint

See: createSnapshotFromVolumeRecoveryPoint smart constructor.

Request Lenses

csfvrpVolumeARN :: Lens' CreateSnapshotFromVolumeRecoveryPoint Text

Undocumented member.

csfvrpSnapshotDescription :: Lens' CreateSnapshotFromVolumeRecoveryPoint Text

Undocumented member.

Destructuring the Response

createSnapshotFromVolumeRecoveryPointResponse

Arguments

:: Int
csfvrprsResponseStatus

-> CreateSnapshotFromVolumeRecoveryPointResponse

Creates a value of CreateSnapshotFromVolumeRecoveryPointResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

csfvrprsVolumeRecoveryPointTime - Undocumented member.

csfvrprsVolumeARN - Undocumented member.

csfvrprsSnapshotId - Undocumented member.

csfvrprsResponseStatus - -- | The response status code.

data CreateSnapshotFromVolumeRecoveryPointResponse

See: createSnapshotFromVolumeRecoveryPointResponse smart constructor.

Response Lenses

csfvrprsVolumeRecoveryPointTime :: Lens' CreateSnapshotFromVolumeRecoveryPointResponse (Maybe Text)

Undocumented member.

csfvrprsVolumeARN :: Lens' CreateSnapshotFromVolumeRecoveryPointResponse (Maybe Text)

Undocumented member.

csfvrprsSnapshotId :: Lens' CreateSnapshotFromVolumeRecoveryPointResponse (Maybe Text)

Undocumented member.

csfvrprsResponseStatus :: Lens' CreateSnapshotFromVolumeRecoveryPointResponse Int

- | The response status code.