Network.AWS.StorageGateway.CreateStorediSCSIVolume

Description

Creates a volume on a specified gateway. This operation is only supported in the stored volume gateway architecture.

The size of the volume to create is inferred from the disk size. You can choose to preserve existing data on the disk, create volume from an existing snapshot, or create an empty volume. If you choose to create an empty gateway volume, then any existing data on the disk is erased.

In the request you must specify the gateway and the disk information on which you are creating the volume. In response, the gateway creates the volume and returns volume information such as the volume Amazon Resource Name (ARN), its size, and the iSCSI target ARN that initiators can use to connect to the volume target.

Creating a Request

createStorediSCSIVolume

Arguments

:: Text
csscsivGatewayARN

-> Text
csscsivDiskId

-> Bool
csscsivPreserveExistingData

-> Text
csscsivTargetName

-> Text
csscsivNetworkInterfaceId

-> CreateStorediSCSIVolume

Creates a value of CreateStorediSCSIVolume with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

csscsivSnapshotId - The snapshot ID (e.g. "snap-1122aabb") of the snapshot to restore as the new stored volume. Specify this field if you want to create the iSCSI storage volume from a snapshot otherwise do not include this field. To list snapshots for your account use DescribeSnapshots in the Amazon Elastic Compute Cloud API Reference .

csscsivGatewayARN - Undocumented member.

csscsivDiskId - The unique identifier for the gateway local disk that is configured as a stored volume. Use ListLocalDisks to list disk IDs for a gateway.

csscsivPreserveExistingData - Specify this field as true if you want to preserve the data on the local disk. Otherwise, specifying this field as false creates an empty volume. Valid Values: true, false

csscsivTargetName - The name of the iSCSI target used by initiators to connect to the target and as a suffix for the target ARN. For example, specifying TargetName as myvolume results in the target ARN of arn:aws:storagegateway:us-east-2:111122223333:gatewaysgw-12A3456Btarget/iqn.1997-05.com.amazon:myvolume. The target name must be unique across all volumes of a gateway.

csscsivNetworkInterfaceId - The network interface of the gateway on which to expose the iSCSI target. Only IPv4 addresses are accepted. Use DescribeGatewayInformation to get a list of the network interfaces available on a gateway. Valid Values: A valid IP address.

data CreateStorediSCSIVolume

A JSON object containing one or more of the following fields:

'CreateStorediSCSIVolumeInput$DiskId'

'CreateStorediSCSIVolumeInput$NetworkInterfaceId'

'CreateStorediSCSIVolumeInput$PreserveExistingData'

'CreateStorediSCSIVolumeInput$SnapshotId'

'CreateStorediSCSIVolumeInput$TargetName'

See: createStorediSCSIVolume smart constructor.

Request Lenses

csscsivSnapshotId :: Lens' CreateStorediSCSIVolume (Maybe Text)

The snapshot ID (e.g. "snap-1122aabb") of the snapshot to restore as the new stored volume. Specify this field if you want to create the iSCSI storage volume from a snapshot otherwise do not include this field. To list snapshots for your account use DescribeSnapshots in the Amazon Elastic Compute Cloud API Reference .

csscsivGatewayARN :: Lens' CreateStorediSCSIVolume Text

Undocumented member.

csscsivDiskId :: Lens' CreateStorediSCSIVolume Text

The unique identifier for the gateway local disk that is configured as a stored volume. Use ListLocalDisks to list disk IDs for a gateway.

csscsivPreserveExistingData :: Lens' CreateStorediSCSIVolume Bool

Specify this field as true if you want to preserve the data on the local disk. Otherwise, specifying this field as false creates an empty volume. Valid Values: true, false

csscsivTargetName :: Lens' CreateStorediSCSIVolume Text

The name of the iSCSI target used by initiators to connect to the target and as a suffix for the target ARN. For example, specifying TargetName as myvolume results in the target ARN of arn:aws:storagegateway:us-east-2:111122223333:gatewaysgw-12A3456Btarget/iqn.1997-05.com.amazon:myvolume. The target name must be unique across all volumes of a gateway.

csscsivNetworkInterfaceId :: Lens' CreateStorediSCSIVolume Text

The network interface of the gateway on which to expose the iSCSI target. Only IPv4 addresses are accepted. Use DescribeGatewayInformation to get a list of the network interfaces available on a gateway. Valid Values: A valid IP address.

Destructuring the Response

createStorediSCSIVolumeResponse

Arguments

:: Int
csscsivrsResponseStatus

-> CreateStorediSCSIVolumeResponse

Creates a value of CreateStorediSCSIVolumeResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

csscsivrsTargetARN - he Amazon Resource Name (ARN) of the volume target that includes the iSCSI name that initiators can use to connect to the target.

csscsivrsVolumeARN - The Amazon Resource Name (ARN) of the configured volume.

csscsivrsVolumeSizeInBytes - The size of the volume in bytes.

csscsivrsResponseStatus - -- | The response status code.

data CreateStorediSCSIVolumeResponse

A JSON object containing the following fields:

See: createStorediSCSIVolumeResponse smart constructor.

Response Lenses

csscsivrsTargetARN :: Lens' CreateStorediSCSIVolumeResponse (Maybe Text)

he Amazon Resource Name (ARN) of the volume target that includes the iSCSI name that initiators can use to connect to the target.

csscsivrsVolumeARN :: Lens' CreateStorediSCSIVolumeResponse (Maybe Text)

The Amazon Resource Name (ARN) of the configured volume.

csscsivrsVolumeSizeInBytes :: Lens' CreateStorediSCSIVolumeResponse (Maybe Integer)

The size of the volume in bytes.

csscsivrsResponseStatus :: Lens' CreateStorediSCSIVolumeResponse Int

- | The response status code.