Network.AWS.StorageGateway.CreateStorediSCSIVolume

Description

This operation creates a volume on a specified gateway. This operation is supported only for the gateway-stored volume architecture.

The size of the volume to create is inferred from the disk size. You can choose to preserve existing data on the disk, create volume from an existing snapshot, or create an empty volume. If you choose to create an empty gateway volume, then any existing data on the disk is erased.

In the request you must specify the gateway and the disk information on which you are creating the volume. In response, AWS Storage Gateway creates the volume and returns volume information such as the volume Amazon Resource Name (ARN), its size, and the iSCSI target ARN that initiators can use to connect to the volume target.

http://docs.aws.amazon.com/storagegateway/latest/APIReference/API_CreateStorediSCSIVolume.html

Request

data CreateStorediSCSIVolume

Request constructor

createStorediSCSIVolume

Arguments

:: Text
csscsivGatewayARN

-> Text
csscsivDiskId

-> Bool
csscsivPreserveExistingData

-> Text
csscsivTargetName

-> Text
csscsivNetworkInterfaceId

-> CreateStorediSCSIVolume

CreateStorediSCSIVolume constructor.

The fields accessible through corresponding lenses are:

csscsivDiskId :: Text

csscsivGatewayARN :: Text

csscsivNetworkInterfaceId :: Text

csscsivPreserveExistingData :: Bool

csscsivSnapshotId :: Maybe Text

csscsivTargetName :: Text

Request lenses

csscsivDiskId :: Lens' CreateStorediSCSIVolume Text

The unique identifier for the gateway local disk that is configured as a stored volume. Use ListLocalDisks to list disk IDs for a gateway.

csscsivGatewayARN :: Lens' CreateStorediSCSIVolume Text

csscsivNetworkInterfaceId :: Lens' CreateStorediSCSIVolume Text

The network interface of the gateway on which to expose the iSCSI target. Only IPv4 addresses are accepted. Use DescribeGatewayInformation to get a list of the network interfaces available on a gateway.

Valid Values: A valid IP address.

csscsivPreserveExistingData :: Lens' CreateStorediSCSIVolume Bool

Specify this field as true if you want to preserve the data on the local disk. Otherwise, specifying this field as false creates an empty volume.

Valid Values: true, false

csscsivSnapshotId :: Lens' CreateStorediSCSIVolume (Maybe Text)

The snapshot ID (e.g. "snap-1122aabb") of the snapshot to restore as the new stored volume. Specify this field if you want to create the iSCSI storage volume from a snapshot otherwise do not include this field. To list snapshots for your account use DescribeSnapshots in the Amazon Elastic Compute CloudAPI Reference.

csscsivTargetName :: Lens' CreateStorediSCSIVolume Text

The name of the iSCSI target used by initiators to connect to the target and as a suffix for the target ARN. For example, specifying TargetName as myvolume results in the target ARN of arn:aws:storagegateway:us-east-1:111122223333:gatewaymygatewaytarget/iqn.1997-05.com.amazon:myvolume. The target name must be unique across all volumes of a gateway.

Response

data CreateStorediSCSIVolumeResponse

Response constructor

createStorediSCSIVolumeResponse :: CreateStorediSCSIVolumeResponse

CreateStorediSCSIVolumeResponse constructor.

The fields accessible through corresponding lenses are:

csscsivrTargetARN :: Maybe Text

csscsivrVolumeARN :: Maybe Text

csscsivrVolumeSizeInBytes :: Maybe Integer

Response lenses

csscsivrTargetARN :: Lens' CreateStorediSCSIVolumeResponse (Maybe Text)

he Amazon Resource Name (ARN) of the volume target that includes the iSCSI name that initiators can use to connect to the target.

csscsivrVolumeARN :: Lens' CreateStorediSCSIVolumeResponse (Maybe Text)

The Amazon Resource Name (ARN) of the configured volume.

csscsivrVolumeSizeInBytes :: Lens' CreateStorediSCSIVolumeResponse (Maybe Integer)

The size of the volume in bytes.