Network.AWS.StorageGateway.CreateTapes

Description

Creates one or more virtual tapes. You write data to the virtual tapes and then archive the tapes. This operation is only supported in the tape gateway architecture.

Creating a Request

createTapes

Arguments

:: Text
ctGatewayARN

-> Integer
ctTapeSizeInBytes

-> Text
ctClientToken

-> Natural
ctNumTapesToCreate

-> Text
ctTapeBarcodePrefix

-> CreateTapes

Creates a value of CreateTapes with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ctGatewayARN - The unique Amazon Resource Name (ARN) that represents the gateway to associate the virtual tapes with. Use the ListGateways operation to return a list of gateways for your account and region.

ctTapeSizeInBytes - The size, in bytes, of the virtual tapes that you want to create.

ctClientToken - A unique identifier that you use to retry a request. If you retry a request, use the same ClientToken you specified in the initial request.

ctNumTapesToCreate - The number of virtual tapes that you want to create.

ctTapeBarcodePrefix - A prefix that you append to the barcode of the virtual tape you are creating. This prefix makes the barcode unique.

data CreateTapes

CreateTapesInput

See: createTapes smart constructor.

Request Lenses

ctGatewayARN :: Lens' CreateTapes Text

The unique Amazon Resource Name (ARN) that represents the gateway to associate the virtual tapes with. Use the ListGateways operation to return a list of gateways for your account and region.

ctTapeSizeInBytes :: Lens' CreateTapes Integer

The size, in bytes, of the virtual tapes that you want to create.

ctClientToken :: Lens' CreateTapes Text

A unique identifier that you use to retry a request. If you retry a request, use the same ClientToken you specified in the initial request.

ctNumTapesToCreate :: Lens' CreateTapes Natural

The number of virtual tapes that you want to create.

ctTapeBarcodePrefix :: Lens' CreateTapes Text

A prefix that you append to the barcode of the virtual tape you are creating. This prefix makes the barcode unique.

Destructuring the Response

createTapesResponse

Arguments

:: Int
ctrsResponseStatus

-> CreateTapesResponse

Creates a value of CreateTapesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ctrsTapeARNs - A list of unique Amazon Resource Names (ARNs) that represents the virtual tapes that were created.

ctrsResponseStatus - -- | The response status code.

data CreateTapesResponse

CreateTapeOutput

See: createTapesResponse smart constructor.

Response Lenses

ctrsTapeARNs :: Lens' CreateTapesResponse [Text]

A list of unique Amazon Resource Names (ARNs) that represents the virtual tapes that were created.

ctrsResponseStatus :: Lens' CreateTapesResponse Int

- | The response status code.