Network.AWS.StorageGateway.DeleteBandwidthRateLimit

Description

This operation deletes the bandwidth rate limits of a gateway. You can delete either the upload and download bandwidth rate limit, or you can delete both. If you delete only one of the limits, the other limit remains unchanged. To specify which gateway to work with, use the Amazon Resource Name (ARN) of the gateway in your request.

http://docs.aws.amazon.com/storagegateway/latest/APIReference/API_DeleteBandwidthRateLimit.html

Request

data DeleteBandwidthRateLimit

Request constructor

deleteBandwidthRateLimit

Arguments

:: Text
dbrl1GatewayARN

-> Text
dbrl1BandwidthType

-> DeleteBandwidthRateLimit

DeleteBandwidthRateLimit constructor.

The fields accessible through corresponding lenses are:

dbrl1BandwidthType :: Text

dbrl1GatewayARN :: Text

Request lenses

dbrl1BandwidthType :: Lens' DeleteBandwidthRateLimit Text

dbrl1GatewayARN :: Lens' DeleteBandwidthRateLimit Text

Response

data DeleteBandwidthRateLimitResponse

Response constructor

deleteBandwidthRateLimitResponse :: DeleteBandwidthRateLimitResponse

DeleteBandwidthRateLimitResponse constructor.

The fields accessible through corresponding lenses are:

dbrlr1GatewayARN :: Maybe Text

Response lenses

dbrlr1GatewayARN :: Lens' DeleteBandwidthRateLimitResponse (Maybe Text)