Network.AWS.StorageGateway.DeleteSnapshotSchedule

Description

This operation deletes a snapshot of a volume.

You can take snapshots of your gateway volumes on a scheduled or ad-hoc basis. This API enables you to delete a snapshot schedule for a volume. For more information, see Working with Snapshots. In the DeleteSnapshotSchedule request, you identify the volume by providing its Amazon Resource Name (ARN).

To list or delete a snapshot, you must use the Amazon EC2 API. in AmazonElastic Compute Cloud API Reference.

http://docs.aws.amazon.com/storagegateway/latest/APIReference/API_DeleteSnapshotSchedule.html

Request

data DeleteSnapshotSchedule

Request constructor

deleteSnapshotSchedule

Arguments

:: Text
dss1VolumeARN

-> DeleteSnapshotSchedule

DeleteSnapshotSchedule constructor.

The fields accessible through corresponding lenses are:

dss1VolumeARN :: Text

Request lenses

dss1VolumeARN :: Lens' DeleteSnapshotSchedule Text

Response

data DeleteSnapshotScheduleResponse

Response constructor

deleteSnapshotScheduleResponse :: DeleteSnapshotScheduleResponse

DeleteSnapshotScheduleResponse constructor.

The fields accessible through corresponding lenses are:

dssr1VolumeARN :: Maybe Text

Response lenses

dssr1VolumeARN :: Lens' DeleteSnapshotScheduleResponse (Maybe Text)