Network.AWS.StorageGateway.DeleteTape

Description

Deletes the specified virtual tape. This operation is only supported in the tape gateway architecture.

Creating a Request

deleteTape

Arguments

:: Text
delGatewayARN

-> Text
delTapeARN

-> DeleteTape

Creates a value of DeleteTape with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

delGatewayARN - The unique Amazon Resource Name (ARN) of the gateway that the virtual tape to delete is associated with. Use the ListGateways operation to return a list of gateways for your account and region.

delTapeARN - The Amazon Resource Name (ARN) of the virtual tape to delete.

data DeleteTape

DeleteTapeInput

See: deleteTape smart constructor.

Request Lenses

delGatewayARN :: Lens' DeleteTape Text

The unique Amazon Resource Name (ARN) of the gateway that the virtual tape to delete is associated with. Use the ListGateways operation to return a list of gateways for your account and region.

delTapeARN :: Lens' DeleteTape Text

The Amazon Resource Name (ARN) of the virtual tape to delete.

Destructuring the Response

deleteTapeResponse

Arguments

:: Int
dtrsResponseStatus

-> DeleteTapeResponse

Creates a value of DeleteTapeResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dtrsTapeARN - The Amazon Resource Name (ARN) of the deleted virtual tape.

dtrsResponseStatus - -- | The response status code.

data DeleteTapeResponse

DeleteTapeOutput

See: deleteTapeResponse smart constructor.

Response Lenses

dtrsTapeARN :: Lens' DeleteTapeResponse (Maybe Text)

The Amazon Resource Name (ARN) of the deleted virtual tape.

dtrsResponseStatus :: Lens' DeleteTapeResponse Int

- | The response status code.