Network.AWS.StorageGateway.DeleteTapeArchive

Description

Deletes the specified virtual tape from the virtual tape shelf (VTS). This operation is only supported in the tape gateway architecture.

Creating a Request

deleteTapeArchive

Arguments

:: Text
dtaTapeARN

-> DeleteTapeArchive

Creates a value of DeleteTapeArchive with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dtaTapeARN - The Amazon Resource Name (ARN) of the virtual tape to delete from the virtual tape shelf (VTS).

data DeleteTapeArchive

DeleteTapeArchiveInput

See: deleteTapeArchive smart constructor.

Request Lenses

dtaTapeARN :: Lens' DeleteTapeArchive Text

The Amazon Resource Name (ARN) of the virtual tape to delete from the virtual tape shelf (VTS).

Destructuring the Response

deleteTapeArchiveResponse

Arguments

:: Int
dtatrsResponseStatus

-> DeleteTapeArchiveResponse

Creates a value of DeleteTapeArchiveResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dtatrsTapeARN - The Amazon Resource Name (ARN) of the virtual tape that was deleted from the virtual tape shelf (VTS).

dtatrsResponseStatus - -- | The response status code.

data DeleteTapeArchiveResponse

DeleteTapeArchiveOutput

See: deleteTapeArchiveResponse smart constructor.

Response Lenses

dtatrsTapeARN :: Lens' DeleteTapeArchiveResponse (Maybe Text)

The Amazon Resource Name (ARN) of the virtual tape that was deleted from the virtual tape shelf (VTS).

dtatrsResponseStatus :: Lens' DeleteTapeArchiveResponse Int

- | The response status code.