Network.AWS.StorageGateway.DescribeBandwidthRateLimit

Description

Returns the bandwidth rate limits of a gateway. By default, these limits are not set, which means no bandwidth rate limiting is in effect.

This operation only returns a value for a bandwidth rate limit only if the limit is set. If no limits are set for the gateway, then this operation returns only the gateway ARN in the response body. To specify which gateway to describe, use the Amazon Resource Name (ARN) of the gateway in your request.

Creating a Request

describeBandwidthRateLimit

Arguments

:: Text
dbrlGatewayARN

-> DescribeBandwidthRateLimit

Creates a value of DescribeBandwidthRateLimit with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dbrlGatewayARN - Undocumented member.

data DescribeBandwidthRateLimit

A JSON object containing the of the gateway.

See: describeBandwidthRateLimit smart constructor.

Request Lenses

dbrlGatewayARN :: Lens' DescribeBandwidthRateLimit Text

Undocumented member.

Destructuring the Response

describeBandwidthRateLimitResponse

Arguments

:: Int
dbrlrsResponseStatus

-> DescribeBandwidthRateLimitResponse

Creates a value of DescribeBandwidthRateLimitResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dbrlrsGatewayARN - Undocumented member.

dbrlrsAverageUploadRateLimitInBitsPerSec - The average upload bandwidth rate limit in bits per second. This field does not appear in the response if the upload rate limit is not set.

dbrlrsAverageDownloadRateLimitInBitsPerSec - The average download bandwidth rate limit in bits per second. This field does not appear in the response if the download rate limit is not set.

dbrlrsResponseStatus - -- | The response status code.

data DescribeBandwidthRateLimitResponse

A JSON object containing the following fields:

See: describeBandwidthRateLimitResponse smart constructor.

Response Lenses

dbrlrsGatewayARN :: Lens' DescribeBandwidthRateLimitResponse (Maybe Text)

Undocumented member.

dbrlrsAverageUploadRateLimitInBitsPerSec :: Lens' DescribeBandwidthRateLimitResponse (Maybe Natural)

The average upload bandwidth rate limit in bits per second. This field does not appear in the response if the upload rate limit is not set.

dbrlrsAverageDownloadRateLimitInBitsPerSec :: Lens' DescribeBandwidthRateLimitResponse (Maybe Natural)

The average download bandwidth rate limit in bits per second. This field does not appear in the response if the download rate limit is not set.

dbrlrsResponseStatus :: Lens' DescribeBandwidthRateLimitResponse Int

- | The response status code.