Network.AWS.StorageGateway.DescribeBandwidthRateLimit

Description

This operation returns the bandwidth rate limits of a gateway. By default, these limits are not set, which means no bandwidth rate limiting is in effect.

This operation only returns a value for a bandwidth rate limit only if the limit is set. If no limits are set for the gateway, then this operation returns only the gateway ARN in the response body. To specify which gateway to describe, use the Amazon Resource Name (ARN) of the gateway in your request.

http://docs.aws.amazon.com/storagegateway/latest/APIReference/API_DescribeBandwidthRateLimit.html

Request

data DescribeBandwidthRateLimit

Request constructor

describeBandwidthRateLimit

Arguments

:: Text
dbrlGatewayARN

-> DescribeBandwidthRateLimit

DescribeBandwidthRateLimit constructor.

The fields accessible through corresponding lenses are:

dbrlGatewayARN :: Text

Request lenses

dbrlGatewayARN :: Lens' DescribeBandwidthRateLimit Text

Response

data DescribeBandwidthRateLimitResponse

Response constructor

describeBandwidthRateLimitResponse :: DescribeBandwidthRateLimitResponse

DescribeBandwidthRateLimitResponse constructor.

The fields accessible through corresponding lenses are:

dbrlrAverageDownloadRateLimitInBitsPerSec :: Maybe Natural

dbrlrAverageUploadRateLimitInBitsPerSec :: Maybe Natural

dbrlrGatewayARN :: Maybe Text

Response lenses

dbrlrAverageDownloadRateLimitInBitsPerSec :: Lens' DescribeBandwidthRateLimitResponse (Maybe Natural)

The average download bandwidth rate limit in bits per second. This field does not appear in the response if the download rate limit is not set.

dbrlrAverageUploadRateLimitInBitsPerSec :: Lens' DescribeBandwidthRateLimitResponse (Maybe Natural)

The average upload bandwidth rate limit in bits per second. This field does not appear in the response if the upload rate limit is not set.

dbrlrGatewayARN :: Lens' DescribeBandwidthRateLimitResponse (Maybe Text)