Network.AWS.StorageGateway.DescribeCache

Description

This operation returns information about the cache of a gateway. This operation is supported only for the gateway-cached volume architecture.

The response includes disk IDs that are configured as cache, and it includes the amount of cache allocated and used.

http://docs.aws.amazon.com/storagegateway/latest/APIReference/API_DescribeCache.html

Request

data DescribeCache

Request constructor

describeCache

Arguments

:: Text
dcGatewayARN

-> DescribeCache

DescribeCache constructor.

The fields accessible through corresponding lenses are:

dcGatewayARN :: Text

Request lenses

dcGatewayARN :: Lens' DescribeCache Text

Response

data DescribeCacheResponse

Response constructor

describeCacheResponse :: DescribeCacheResponse

DescribeCacheResponse constructor.

The fields accessible through corresponding lenses are:

dcrCacheAllocatedInBytes :: Maybe Integer

dcrCacheDirtyPercentage :: Maybe Double

dcrCacheHitPercentage :: Maybe Double

dcrCacheMissPercentage :: Maybe Double

dcrCacheUsedPercentage :: Maybe Double

dcrDiskIds :: [Text]

dcrGatewayARN :: Maybe Text

Response lenses

dcrCacheAllocatedInBytes :: Lens' DescribeCacheResponse (Maybe Integer)

dcrCacheDirtyPercentage :: Lens' DescribeCacheResponse (Maybe Double)

dcrCacheHitPercentage :: Lens' DescribeCacheResponse (Maybe Double)

dcrCacheMissPercentage :: Lens' DescribeCacheResponse (Maybe Double)

dcrCacheUsedPercentage :: Lens' DescribeCacheResponse (Maybe Double)

dcrDiskIds :: Lens' DescribeCacheResponse [Text]

dcrGatewayARN :: Lens' DescribeCacheResponse (Maybe Text)