Network.AWS.StorageGateway.DescribeGatewayInformation

Description

Returns metadata about a gateway such as its name, network interfaces, configured time zone, and the state (whether the gateway is running or not). To specify which gateway to describe, use the Amazon Resource Name (ARN) of the gateway in your request.

Creating a Request

describeGatewayInformation

Arguments

:: Text
dgiGatewayARN

-> DescribeGatewayInformation

Creates a value of DescribeGatewayInformation with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dgiGatewayARN - Undocumented member.

data DescribeGatewayInformation

A JSON object containing the id of the gateway.

See: describeGatewayInformation smart constructor.

Request Lenses

dgiGatewayARN :: Lens' DescribeGatewayInformation Text

Undocumented member.

Destructuring the Response

describeGatewayInformationResponse

Arguments

:: Int
dgirsResponseStatus

-> DescribeGatewayInformationResponse

Creates a value of DescribeGatewayInformationResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dgirsGatewayState - A value that indicates the operating state of the gateway.

dgirsGatewayARN - Undocumented member.

dgirsGatewayNetworkInterfaces - A NetworkInterface array that contains descriptions of the gateway network interfaces.

dgirsNextUpdateAvailabilityDate - The date on which an update to the gateway is available. This date is in the time zone of the gateway. If the gateway is not available for an update this field is not returned in the response.

dgirsLastSoftwareUpdate - The date on which the last software update was applied to the gateway. If the gateway has never been updated, this field does not return a value in the response.

dgirsGatewayName - The name you configured for your gateway.

dgirsGatewayId - The unique identifier assigned to your gateway during activation. This ID becomes part of the gateway Amazon Resource Name (ARN), which you use as input for other operations.

dgirsGatewayType - The type of the gateway.

dgirsGatewayTimezone - A value that indicates the time zone configured for the gateway.

dgirsResponseStatus - -- | The response status code.

data DescribeGatewayInformationResponse

A JSON object containing the following fields:

See: describeGatewayInformationResponse smart constructor.

Response Lenses

dgirsGatewayState :: Lens' DescribeGatewayInformationResponse (Maybe Text)

A value that indicates the operating state of the gateway.

dgirsGatewayARN :: Lens' DescribeGatewayInformationResponse (Maybe Text)

Undocumented member.

dgirsGatewayNetworkInterfaces :: Lens' DescribeGatewayInformationResponse [NetworkInterface]

A NetworkInterface array that contains descriptions of the gateway network interfaces.

dgirsNextUpdateAvailabilityDate :: Lens' DescribeGatewayInformationResponse (Maybe Text)

The date on which an update to the gateway is available. This date is in the time zone of the gateway. If the gateway is not available for an update this field is not returned in the response.

dgirsLastSoftwareUpdate :: Lens' DescribeGatewayInformationResponse (Maybe Text)

The date on which the last software update was applied to the gateway. If the gateway has never been updated, this field does not return a value in the response.

dgirsGatewayName :: Lens' DescribeGatewayInformationResponse (Maybe Text)

The name you configured for your gateway.

dgirsGatewayId :: Lens' DescribeGatewayInformationResponse (Maybe Text)

The unique identifier assigned to your gateway during activation. This ID becomes part of the gateway Amazon Resource Name (ARN), which you use as input for other operations.

dgirsGatewayType :: Lens' DescribeGatewayInformationResponse (Maybe Text)

The type of the gateway.

dgirsGatewayTimezone :: Lens' DescribeGatewayInformationResponse (Maybe Text)

A value that indicates the time zone configured for the gateway.

dgirsResponseStatus :: Lens' DescribeGatewayInformationResponse Int

- | The response status code.