Network.AWS.StorageGateway.DescribeMaintenanceStartTime

Description

Returns your gateway's weekly maintenance start time including the day and time of the week. Note that values are in terms of the gateway's time zone.

Creating a Request

describeMaintenanceStartTime

Arguments

:: Text
dmstGatewayARN

-> DescribeMaintenanceStartTime

Creates a value of DescribeMaintenanceStartTime with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dmstGatewayARN - Undocumented member.

data DescribeMaintenanceStartTime

A JSON object containing the of the gateway.

See: describeMaintenanceStartTime smart constructor.

Request Lenses

dmstGatewayARN :: Lens' DescribeMaintenanceStartTime Text

Undocumented member.

Destructuring the Response

describeMaintenanceStartTimeResponse

Arguments

:: Int
dmstrsResponseStatus

-> DescribeMaintenanceStartTimeResponse

Creates a value of DescribeMaintenanceStartTimeResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dmstrsGatewayARN - Undocumented member.

dmstrsMinuteOfHour - The minute component of the maintenance start time represented as mm , where mm is the minute (0 to 59). The minute of the hour is in the time zone of the gateway.

dmstrsHourOfDay - The hour component of the maintenance start time represented as hh , where hh is the hour (0 to 23). The hour of the day is in the time zone of the gateway.

dmstrsTimezone - Undocumented member.

dmstrsDayOfWeek - An ordinal number between 0 and 6 that represents the day of the week, where 0 represents Sunday and 6 represents Saturday. The day of week is in the time zone of the gateway.

dmstrsResponseStatus - -- | The response status code.

data DescribeMaintenanceStartTimeResponse

A JSON object containing the following fields:

'DescribeMaintenanceStartTimeOutput$DayOfWeek'

'DescribeMaintenanceStartTimeOutput$HourOfDay'

'DescribeMaintenanceStartTimeOutput$MinuteOfHour'

'DescribeMaintenanceStartTimeOutput$Timezone'

See: describeMaintenanceStartTimeResponse smart constructor.

Response Lenses

dmstrsGatewayARN :: Lens' DescribeMaintenanceStartTimeResponse (Maybe Text)

Undocumented member.

dmstrsMinuteOfHour :: Lens' DescribeMaintenanceStartTimeResponse (Maybe Natural)

The minute component of the maintenance start time represented as mm , where mm is the minute (0 to 59). The minute of the hour is in the time zone of the gateway.

dmstrsHourOfDay :: Lens' DescribeMaintenanceStartTimeResponse (Maybe Natural)

The hour component of the maintenance start time represented as hh , where hh is the hour (0 to 23). The hour of the day is in the time zone of the gateway.

dmstrsTimezone :: Lens' DescribeMaintenanceStartTimeResponse (Maybe Text)

Undocumented member.

dmstrsDayOfWeek :: Lens' DescribeMaintenanceStartTimeResponse (Maybe Natural)

An ordinal number between 0 and 6 that represents the day of the week, where 0 represents Sunday and 6 represents Saturday. The day of week is in the time zone of the gateway.

dmstrsResponseStatus :: Lens' DescribeMaintenanceStartTimeResponse Int

- | The response status code.