Network.AWS.StorageGateway.DescribeSnapshotSchedule

Description

This operation describes the snapshot schedule for the specified gateway volume. The snapshot schedule information includes intervals at which snapshots are automatically initiated on the volume.

http://docs.aws.amazon.com/storagegateway/latest/APIReference/API_DescribeSnapshotSchedule.html

Request

data DescribeSnapshotSchedule

Request constructor

describeSnapshotSchedule

Arguments

:: Text
dssVolumeARN

-> DescribeSnapshotSchedule

DescribeSnapshotSchedule constructor.

The fields accessible through corresponding lenses are:

dssVolumeARN :: Text

Request lenses

dssVolumeARN :: Lens' DescribeSnapshotSchedule Text

The Amazon Resource Name (ARN) of the volume. Use the ListVolumes operation to return a list of gateway volumes.

Response

data DescribeSnapshotScheduleResponse

Response constructor

describeSnapshotScheduleResponse :: DescribeSnapshotScheduleResponse

DescribeSnapshotScheduleResponse constructor.

The fields accessible through corresponding lenses are:

dssrDescription :: Maybe Text

dssrRecurrenceInHours :: Maybe Natural

dssrStartAt :: Maybe Natural

dssrTimezone :: Maybe Text

dssrVolumeARN :: Maybe Text

Response lenses

dssrDescription :: Lens' DescribeSnapshotScheduleResponse (Maybe Text)

dssrRecurrenceInHours :: Lens' DescribeSnapshotScheduleResponse (Maybe Natural)

dssrStartAt :: Lens' DescribeSnapshotScheduleResponse (Maybe Natural)

dssrTimezone :: Lens' DescribeSnapshotScheduleResponse (Maybe Text)

dssrVolumeARN :: Lens' DescribeSnapshotScheduleResponse (Maybe Text)