Network.AWS.StorageGateway.DescribeTapeArchives

Description

Returns a description of specified virtual tapes in the virtual tape shelf (VTS). This operation is only supported in the tape gateway architecture.

If a specific TapeARN is not specified, AWS Storage Gateway returns a description of all virtual tapes found in the VTS associated with your account.

This operation returns paginated results.

Creating a Request

describeTapeArchives :: DescribeTapeArchives

Creates a value of DescribeTapeArchives with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dtaMarker - An opaque string that indicates the position at which to begin describing virtual tapes.

dtaLimit - Specifies that the number of virtual tapes descried be limited to the specified number.

dtaTapeARNs - Specifies one or more unique Amazon Resource Names (ARNs) that represent the virtual tapes you want to describe.

data DescribeTapeArchives

DescribeTapeArchivesInput

See: describeTapeArchives smart constructor.

Request Lenses

dtaMarker :: Lens' DescribeTapeArchives (Maybe Text)

An opaque string that indicates the position at which to begin describing virtual tapes.

dtaLimit :: Lens' DescribeTapeArchives (Maybe Natural)

Specifies that the number of virtual tapes descried be limited to the specified number.

dtaTapeARNs :: Lens' DescribeTapeArchives [Text]

Specifies one or more unique Amazon Resource Names (ARNs) that represent the virtual tapes you want to describe.

Destructuring the Response

describeTapeArchivesResponse

Arguments

:: Int
dtarsResponseStatus

-> DescribeTapeArchivesResponse

Creates a value of DescribeTapeArchivesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dtarsTapeArchives - An array of virtual tape objects in the virtual tape shelf (VTS). The description includes of the Amazon Resource Name(ARN) of the virtual tapes. The information returned includes the Amazon Resource Names (ARNs) of the tapes, size of the tapes, status of the tapes, progress of the description and tape barcode.

dtarsMarker - An opaque string that indicates the position at which the virtual tapes that were fetched for description ended. Use this marker in your next request to fetch the next set of virtual tapes in the virtual tape shelf (VTS). If there are no more virtual tapes to describe, this field does not appear in the response.

dtarsResponseStatus - -- | The response status code.

data DescribeTapeArchivesResponse

DescribeTapeArchivesOutput

See: describeTapeArchivesResponse smart constructor.

Response Lenses

dtarsTapeArchives :: Lens' DescribeTapeArchivesResponse [TapeArchive]

An array of virtual tape objects in the virtual tape shelf (VTS). The description includes of the Amazon Resource Name(ARN) of the virtual tapes. The information returned includes the Amazon Resource Names (ARNs) of the tapes, size of the tapes, status of the tapes, progress of the description and tape barcode.

dtarsMarker :: Lens' DescribeTapeArchivesResponse (Maybe Text)

An opaque string that indicates the position at which the virtual tapes that were fetched for description ended. Use this marker in your next request to fetch the next set of virtual tapes in the virtual tape shelf (VTS). If there are no more virtual tapes to describe, this field does not appear in the response.

dtarsResponseStatus :: Lens' DescribeTapeArchivesResponse Int

- | The response status code.