Network.AWS.StorageGateway.DescribeTapeRecoveryPoints

Description

Returns a list of virtual tape recovery points that are available for the specified tape gateway.

A recovery point is a point-in-time view of a virtual tape at which all the data on the virtual tape is consistent. If your gateway crashes, virtual tapes that have recovery points can be recovered to a new gateway. This operation is only supported in the tape gateway architecture.

This operation returns paginated results.

Creating a Request

describeTapeRecoveryPoints

Arguments

:: Text
dtrpGatewayARN

-> DescribeTapeRecoveryPoints

Creates a value of DescribeTapeRecoveryPoints with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dtrpMarker - An opaque string that indicates the position at which to begin describing the virtual tape recovery points.

dtrpLimit - Specifies that the number of virtual tape recovery points that are described be limited to the specified number.

dtrpGatewayARN - Undocumented member.

data DescribeTapeRecoveryPoints

DescribeTapeRecoveryPointsInput

See: describeTapeRecoveryPoints smart constructor.

Request Lenses

dtrpMarker :: Lens' DescribeTapeRecoveryPoints (Maybe Text)

An opaque string that indicates the position at which to begin describing the virtual tape recovery points.

dtrpLimit :: Lens' DescribeTapeRecoveryPoints (Maybe Natural)

Specifies that the number of virtual tape recovery points that are described be limited to the specified number.

dtrpGatewayARN :: Lens' DescribeTapeRecoveryPoints Text

Undocumented member.

Destructuring the Response

describeTapeRecoveryPointsResponse

Arguments

:: Int
dtrprsResponseStatus

-> DescribeTapeRecoveryPointsResponse

Creates a value of DescribeTapeRecoveryPointsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dtrprsTapeRecoveryPointInfos - An array of TapeRecoveryPointInfos that are available for the specified gateway.

dtrprsGatewayARN - Undocumented member.

dtrprsMarker - An opaque string that indicates the position at which the virtual tape recovery points that were listed for description ended. Use this marker in your next request to list the next set of virtual tape recovery points in the list. If there are no more recovery points to describe, this field does not appear in the response.

dtrprsResponseStatus - -- | The response status code.

data DescribeTapeRecoveryPointsResponse

DescribeTapeRecoveryPointsOutput

See: describeTapeRecoveryPointsResponse smart constructor.

Response Lenses

dtrprsTapeRecoveryPointInfos :: Lens' DescribeTapeRecoveryPointsResponse [TapeRecoveryPointInfo]

An array of TapeRecoveryPointInfos that are available for the specified gateway.

dtrprsGatewayARN :: Lens' DescribeTapeRecoveryPointsResponse (Maybe Text)

Undocumented member.

dtrprsMarker :: Lens' DescribeTapeRecoveryPointsResponse (Maybe Text)

An opaque string that indicates the position at which the virtual tape recovery points that were listed for description ended. Use this marker in your next request to list the next set of virtual tape recovery points in the list. If there are no more recovery points to describe, this field does not appear in the response.

dtrprsResponseStatus :: Lens' DescribeTapeRecoveryPointsResponse Int

- | The response status code.