Network.AWS.StorageGateway.DescribeTapeRecoveryPoints

Description

Returns a list of virtual tape recovery points that are available for the specified gateway-VTL.

A recovery point is a point in time view of a virtual tape at which all the data on the virtual tape is consistent. If your gateway crashes, virtual tapes that have recovery points can be recovered to a new gateway.

http://docs.aws.amazon.com/storagegateway/latest/APIReference/API_DescribeTapeRecoveryPoints.html

Request

data DescribeTapeRecoveryPoints

Request constructor

describeTapeRecoveryPoints

Arguments

:: Text
dtrpGatewayARN

-> DescribeTapeRecoveryPoints

DescribeTapeRecoveryPoints constructor.

The fields accessible through corresponding lenses are:

dtrpGatewayARN :: Text

dtrpLimit :: Maybe Natural

dtrpMarker :: Maybe Text

Request lenses

dtrpGatewayARN :: Lens' DescribeTapeRecoveryPoints Text

dtrpLimit :: Lens' DescribeTapeRecoveryPoints (Maybe Natural)

Specifies that the number of virtual tape recovery points that are described be limited to the specified number.

dtrpMarker :: Lens' DescribeTapeRecoveryPoints (Maybe Text)

An opaque string that indicates the position at which to begin describing the virtual tape recovery points.

Response

data DescribeTapeRecoveryPointsResponse

Response constructor

describeTapeRecoveryPointsResponse :: DescribeTapeRecoveryPointsResponse

DescribeTapeRecoveryPointsResponse constructor.

The fields accessible through corresponding lenses are:

dtrprGatewayARN :: Maybe Text

dtrprMarker :: Maybe Text

dtrprTapeRecoveryPointInfos :: [TapeRecoveryPointInfo]

Response lenses

dtrprGatewayARN :: Lens' DescribeTapeRecoveryPointsResponse (Maybe Text)

dtrprMarker :: Lens' DescribeTapeRecoveryPointsResponse (Maybe Text)

An opaque string that indicates the position at which the virtual tape recovery points that were listed for description ended.

Use this marker in your next request to list the next set of virtual tape recovery points in the list. If there are no more recovery points to describe, this field does not appear in the response.

dtrprTapeRecoveryPointInfos :: Lens' DescribeTapeRecoveryPointsResponse [TapeRecoveryPointInfo]

An array of TapeRecoveryPointInfos that are available for the specified gateway.