Network.AWS.StorageGateway.DescribeTapes

Description

Returns a description of the specified Amazon Resource Name (ARN) of virtual tapes. If a TapeARN is not specified, returns a description of all virtual tapes associated with the specified gateway.

http://docs.aws.amazon.com/storagegateway/latest/APIReference/API_DescribeTapes.html

Request

data DescribeTapes

Request constructor

describeTapes

Arguments

:: Text
dtGatewayARN

-> DescribeTapes

DescribeTapes constructor.

The fields accessible through corresponding lenses are:

dtGatewayARN :: Text

dtLimit :: Maybe Natural

dtMarker :: Maybe Text

dtTapeARNs :: [Text]

Request lenses

dtGatewayARN :: Lens' DescribeTapes Text

dtLimit :: Lens' DescribeTapes (Maybe Natural)

Specifies that the number of virtual tapes described be limited to the specified number.

Amazon Web Services may impose its own limit, if this field is not set.

dtMarker :: Lens' DescribeTapes (Maybe Text)

A marker value, obtained in a previous call to DescribeTapes. This marker indicates which page of results to retrieve.

If not specified, the first page of results is retrieved.

dtTapeARNs :: Lens' DescribeTapes [Text]

Specifies one or more unique Amazon Resource Names (ARNs) that represent the virtual tapes you want to describe. If this parameter is not specified, AWS Storage Gateway returns a description of all virtual tapes associated with the specified gateway.

Response

data DescribeTapesResponse

Response constructor

describeTapesResponse :: DescribeTapesResponse

DescribeTapesResponse constructor.

The fields accessible through corresponding lenses are:

dtrMarker :: Maybe Text

dtrTapes :: [Tape]

Response lenses

dtrMarker :: Lens' DescribeTapesResponse (Maybe Text)

An opaque string which can be used as part of a subsequent DescribeTapes call to retrieve the next page of results.

If a response does not contain a marker, then there are no more results to be retrieved.

dtrTapes :: Lens' DescribeTapesResponse [Tape]

An array of virtual tape descriptions.