Network.AWS.StorageGateway.DescribeVTLDevices

Description

Returns a description of virtual tape library (VTL) devices for the specified tape gateway. In the response, AWS Storage Gateway returns VTL device information.

This operation is only supported in the tape gateway architecture.

This operation returns paginated results.

Creating a Request

describeVTLDevices

Arguments

:: Text
dvtldGatewayARN

-> DescribeVTLDevices

Creates a value of DescribeVTLDevices with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dvtldMarker - An opaque string that indicates the position at which to begin describing the VTL devices.

dvtldLimit - Specifies that the number of VTL devices described be limited to the specified number.

dvtldVTLDeviceARNs - An array of strings, where each string represents the Amazon Resource Name (ARN) of a VTL device.

dvtldGatewayARN - Undocumented member.

data DescribeVTLDevices

DescribeVTLDevicesInput

See: describeVTLDevices smart constructor.

Request Lenses

dvtldMarker :: Lens' DescribeVTLDevices (Maybe Text)

An opaque string that indicates the position at which to begin describing the VTL devices.

dvtldLimit :: Lens' DescribeVTLDevices (Maybe Natural)

Specifies that the number of VTL devices described be limited to the specified number.

dvtldVTLDeviceARNs :: Lens' DescribeVTLDevices [Text]

An array of strings, where each string represents the Amazon Resource Name (ARN) of a VTL device.

dvtldGatewayARN :: Lens' DescribeVTLDevices Text

Undocumented member.

Destructuring the Response

describeVTLDevicesResponse

Arguments

:: Int
dvtldrsResponseStatus

-> DescribeVTLDevicesResponse

Creates a value of DescribeVTLDevicesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dvtldrsVTLDevices - An array of VTL device objects composed of the Amazon Resource Name(ARN) of the VTL devices.

dvtldrsGatewayARN - Undocumented member.

dvtldrsMarker - An opaque string that indicates the position at which the VTL devices that were fetched for description ended. Use the marker in your next request to fetch the next set of VTL devices in the list. If there are no more VTL devices to describe, this field does not appear in the response.

dvtldrsResponseStatus - -- | The response status code.

data DescribeVTLDevicesResponse

DescribeVTLDevicesOutput

See: describeVTLDevicesResponse smart constructor.

Response Lenses

dvtldrsVTLDevices :: Lens' DescribeVTLDevicesResponse [VTLDevice]

An array of VTL device objects composed of the Amazon Resource Name(ARN) of the VTL devices.

dvtldrsGatewayARN :: Lens' DescribeVTLDevicesResponse (Maybe Text)

Undocumented member.

dvtldrsMarker :: Lens' DescribeVTLDevicesResponse (Maybe Text)

An opaque string that indicates the position at which the VTL devices that were fetched for description ended. Use the marker in your next request to fetch the next set of VTL devices in the list. If there are no more VTL devices to describe, this field does not appear in the response.

dvtldrsResponseStatus :: Lens' DescribeVTLDevicesResponse Int

- | The response status code.