Network.AWS.StorageGateway.DescribeWorkingStorage

Description

This operation returns information about the working storage of a gateway. This operation is supported only for the gateway-stored volume architecture. This operation is deprecated in cached-volumes API version (20120630). Use DescribeUploadBuffer instead.

Working storage is also referred to as upload buffer. You can also use the DescribeUploadBuffer operation to add upload buffer to a stored-volume gateway.

The response includes disk IDs that are configured as working storage, and it includes the amount of working storage allocated and used.

http://docs.aws.amazon.com/storagegateway/latest/APIReference/API_DescribeWorkingStorage.html

Request

data DescribeWorkingStorage

Request constructor

describeWorkingStorage

Arguments

:: Text
dwsGatewayARN

-> DescribeWorkingStorage

DescribeWorkingStorage constructor.

The fields accessible through corresponding lenses are:

dwsGatewayARN :: Text

Request lenses

dwsGatewayARN :: Lens' DescribeWorkingStorage Text

Response

data DescribeWorkingStorageResponse

Response constructor

describeWorkingStorageResponse :: DescribeWorkingStorageResponse

DescribeWorkingStorageResponse constructor.

The fields accessible through corresponding lenses are:

dwsrDiskIds :: [Text]

dwsrGatewayARN :: Maybe Text

dwsrWorkingStorageAllocatedInBytes :: Maybe Integer

dwsrWorkingStorageUsedInBytes :: Maybe Integer

Response lenses

dwsrDiskIds :: Lens' DescribeWorkingStorageResponse [Text]

An array of the gateway's local disk IDs that are configured as working storage. Each local disk ID is specified as a string (minimum length of 1 and maximum length of 300). If no local disks are configured as working storage, then the DiskIds array is empty.

dwsrGatewayARN :: Lens' DescribeWorkingStorageResponse (Maybe Text)

dwsrWorkingStorageAllocatedInBytes :: Lens' DescribeWorkingStorageResponse (Maybe Integer)

The total working storage in bytes allocated for the gateway. If no working storage is configured for the gateway, this field returns 0.

dwsrWorkingStorageUsedInBytes :: Lens' DescribeWorkingStorageResponse (Maybe Integer)

The total working storage in bytes in use by the gateway. If no working storage is configured for the gateway, this field returns 0.