Network.AWS.StorageGateway.ListLocalDisks

Description

This operation returns a list of the local disks of a gateway. To specify which gateway to describe you use the Amazon Resource Name (ARN) of the gateway in the body of the request. The request returns all disks, specifying which are configured as working storage, stored volume or not configured at all. Example Request The following example shows a request that returns information about a gateway's local disks. POST HTTP1.1 Host: storagegateway.us-east-1.amazonaws.com x-amz-Date: 20120425T120000Z Authorization: CSOC7TJPLR0OOKIRLGOHVAICUFVV4KQNSO5AEMVJF66Q9ASUAAJG Content-type: application/x-amz-json-1.1 x-amz-target: StorageGateway_20120630.ListLocalDisks { GatewayARN: "arn:aws:storagegateway:us-east-1:111122223333:gateway/mygateway" } HTTP/1.1 200 OK x-amzn-RequestId: CSOC7TJPLR0OOKIRLGOHVAICUFVV4KQNSO5AEMVJF66Q9ASUAAJG Date: Wed, 25 Apr 2012 12:00:02 GMT Content-type: application/x-amz-json-1.1 Content-length: 398 { Disks: [ { DiskAllocationType: UPLOAD_BUFFER, DiskId: "pci-0000:03:00.0-scsi-0:0:0:0", DiskNode: "SCSI(0:0)", DiskPath: "devsda", DiskSizeInBytes: 1099511627776 }, { DiskAllocationType: STORED_iSCSI_VOLUME, DiskAllocationResource: "", DiskId: "pci-0000:03:00.0-scsi-0:0:1:0", DiskNode: "SCSI(0:1)", DiskPath: "devsdb", DiskSizeInBytes: 1099511627776 } ], GatewayARN: "arn:aws:storagegateway:us-east-1:111122223333:gateway/mygateway" }.

Request

data ListLocalDisks

A JSON object containing the of the gateway.

Request constructor

listLocalDisks

Arguments

:: Text
lldGatewayARN

-> ListLocalDisks

Smart constructor for the minimum required parameters to construct a valid ListLocalDisks request.

The fields accessible through corresponding lenses are:

GatewayARN :: Text

Request lenses

lldGatewayARN :: Lens' ListLocalDisks Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

Response

data ListLocalDisksResponse

Response constructor

listLocalDisksResponse :: ListLocalDisksResponse

Smart constructor for the minimum required parameters to construct a valid ListLocalDisksResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

GatewayARN :: Maybe Text

Disks :: [DiskInformation]

Response lenses

lldrGatewayARN :: Lens' ListLocalDisksResponse (Maybe Text)

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

lldrDisks :: Lens' ListLocalDisksResponse [DiskInformation]