Network.AWS.StorageGateway.ListLocalDisks

Description

This operation returns a list of the gateway's local disks. To specify which gateway to describe, you use the Amazon Resource Name (ARN) of the gateway in the body of the request.

The request returns a list of all disks, specifying which are configured as working storage, cache storage, or stored volume or not configured at all. The response includes a DiskStatus field. This field can have a value of present (the disk is availble to use), missing (the disk is no longer connected to the gateway), or mismatch (the disk node is occupied by a disk that has incorrect metadata or the disk content is corrupted).

http://docs.aws.amazon.com/storagegateway/latest/APIReference/API_ListLocalDisks.html

Request

data ListLocalDisks

Request constructor

listLocalDisks

Arguments

:: Text
lldGatewayARN

-> ListLocalDisks

ListLocalDisks constructor.

The fields accessible through corresponding lenses are:

lldGatewayARN :: Text

Request lenses

lldGatewayARN :: Lens' ListLocalDisks Text

Response

data ListLocalDisksResponse

Response constructor

listLocalDisksResponse :: ListLocalDisksResponse

ListLocalDisksResponse constructor.

The fields accessible through corresponding lenses are:

lldrDisks :: [Disk]

lldrGatewayARN :: Maybe Text

Response lenses

lldrDisks :: Lens' ListLocalDisksResponse [Disk]

lldrGatewayARN :: Lens' ListLocalDisksResponse (Maybe Text)