Network.AWS.StorageGateway.RetrieveTapeArchive

Description

Retrieves an archived virtual tape from the virtual tape shelf (VTS) to a gateway-VTL. Virtual tapes archived in the VTS are not associated with any gateway. However after a tape is retrieved, it is associated with a gateway, even though it is also listed in the VTS.

Once a tape is successfully retrieved to a gateway, it cannot be retrieved again to another gateway. You must archive the tape again before you can retrieve it to another gateway.

http://docs.aws.amazon.com/storagegateway/latest/APIReference/API_RetrieveTapeArchive.html

Request

data RetrieveTapeArchive

Request constructor

retrieveTapeArchive

Arguments

:: Text
rtaTapeARN

-> Text
rtaGatewayARN

-> RetrieveTapeArchive

RetrieveTapeArchive constructor.

The fields accessible through corresponding lenses are:

rtaGatewayARN :: Text

rtaTapeARN :: Text

Request lenses

rtaGatewayARN :: Lens' RetrieveTapeArchive Text

The Amazon Resource Name (ARN) of the gateway you want to retrieve the virtual tape to. Use the ListGateways operation to return a list of gateways for your account and region.

You retrieve archived virtual tapes to only one gateway and the gateway must be a gateway-VTL.

rtaTapeARN :: Lens' RetrieveTapeArchive Text

The Amazon Resource Name (ARN) of the virtual tape you want to retrieve from the virtual tape shelf (VTS).

Response

data RetrieveTapeArchiveResponse

Response constructor

retrieveTapeArchiveResponse :: RetrieveTapeArchiveResponse

RetrieveTapeArchiveResponse constructor.

The fields accessible through corresponding lenses are:

rtarTapeARN :: Maybe Text

Response lenses

rtarTapeARN :: Lens' RetrieveTapeArchiveResponse (Maybe Text)

The Amazon Resource Name (ARN) of the retrieved virtual tape.