Network.AWS.StorageGateway.RetrieveTapeRecoveryPoint

Description

Retrieves the recovery point for the specified virtual tape. This operation is only supported in the tape gateway architecture.

A recovery point is a point in time view of a virtual tape at which all the data on the tape is consistent. If your gateway crashes, virtual tapes that have recovery points can be recovered to a new gateway.

Creating a Request

retrieveTapeRecoveryPoint

Arguments

:: Text
rtrpTapeARN

-> Text
rtrpGatewayARN

-> RetrieveTapeRecoveryPoint

Creates a value of RetrieveTapeRecoveryPoint with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rtrpTapeARN - The Amazon Resource Name (ARN) of the virtual tape for which you want to retrieve the recovery point.

rtrpGatewayARN - Undocumented member.

data RetrieveTapeRecoveryPoint

RetrieveTapeRecoveryPointInput

See: retrieveTapeRecoveryPoint smart constructor.

Request Lenses

rtrpTapeARN :: Lens' RetrieveTapeRecoveryPoint Text

The Amazon Resource Name (ARN) of the virtual tape for which you want to retrieve the recovery point.

rtrpGatewayARN :: Lens' RetrieveTapeRecoveryPoint Text

Undocumented member.

Destructuring the Response

retrieveTapeRecoveryPointResponse

Arguments

:: Int
rtrprsResponseStatus

-> RetrieveTapeRecoveryPointResponse

Creates a value of RetrieveTapeRecoveryPointResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rtrprsTapeARN - The Amazon Resource Name (ARN) of the virtual tape for which the recovery point was retrieved.

rtrprsResponseStatus - -- | The response status code.

data RetrieveTapeRecoveryPointResponse

RetrieveTapeRecoveryPointOutput

See: retrieveTapeRecoveryPointResponse smart constructor.

Response Lenses

rtrprsTapeARN :: Lens' RetrieveTapeRecoveryPointResponse (Maybe Text)

The Amazon Resource Name (ARN) of the virtual tape for which the recovery point was retrieved.

rtrprsResponseStatus :: Lens' RetrieveTapeRecoveryPointResponse Int

- | The response status code.