Network.AWS.StorageGateway.ShutdownGateway

Description

Shuts down a gateway. To specify which gateway to shut down, use the Amazon Resource Name (ARN) of the gateway in the body of your request.

The operation shuts down the gateway service component running in the gateway's virtual machine (VM) and not the host VM.

After the gateway is shutdown, you cannot call any other API except StartGateway , DescribeGatewayInformation , and ListGateways . For more information, see ActivateGateway . Your applications cannot read from or write to the gateway's storage volumes, and there are no snapshots taken.

If do not intend to use the gateway again, you must delete the gateway (using DeleteGateway ) to no longer pay software charges associated with the gateway.

Creating a Request

shutdownGateway

Arguments

:: Text
sGatewayARN

-> ShutdownGateway

Creates a value of ShutdownGateway with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sGatewayARN - Undocumented member.

data ShutdownGateway

A JSON object containing the of the gateway to shut down.

See: shutdownGateway smart constructor.

Request Lenses

sGatewayARN :: Lens' ShutdownGateway Text

Undocumented member.

Destructuring the Response

shutdownGatewayResponse

Arguments

:: Int
srsResponseStatus

-> ShutdownGatewayResponse

Creates a value of ShutdownGatewayResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

srsGatewayARN - Undocumented member.

srsResponseStatus - -- | The response status code.

data ShutdownGatewayResponse

A JSON object containing the of the gateway that was shut down.

See: shutdownGatewayResponse smart constructor.

Response Lenses

srsGatewayARN :: Lens' ShutdownGatewayResponse (Maybe Text)

Undocumented member.

srsResponseStatus :: Lens' ShutdownGatewayResponse Int

- | The response status code.