Network.AWS.StorageGateway.UpdateBandwidthRateLimit

Description

This operation updates the bandwidth rate limits of a gateway. You can update both the upload and download bandwidth rate limit or specify only one of the two. If you don't set a bandwidth rate limit, the existing rate limit remains.

By default, a gateway's bandwidth rate limits are not set. If you don't set any limit, the gateway does not have any limitations on its bandwidth usage and could potentially use the maximum available bandwidth.

To specify which gateway to update, use the Amazon Resource Name (ARN) of the gateway in your request.

http://docs.aws.amazon.com/storagegateway/latest/APIReference/API_UpdateBandwidthRateLimit.html

Request

data UpdateBandwidthRateLimit

Request constructor

updateBandwidthRateLimit

Arguments

:: Text
ubrlGatewayARN

-> UpdateBandwidthRateLimit

UpdateBandwidthRateLimit constructor.

The fields accessible through corresponding lenses are:

ubrlAverageDownloadRateLimitInBitsPerSec :: Maybe Natural

ubrlAverageUploadRateLimitInBitsPerSec :: Maybe Natural

ubrlGatewayARN :: Text

Request lenses

ubrlAverageDownloadRateLimitInBitsPerSec :: Lens' UpdateBandwidthRateLimit (Maybe Natural)

The average download bandwidth rate limit in bits per second.

ubrlAverageUploadRateLimitInBitsPerSec :: Lens' UpdateBandwidthRateLimit (Maybe Natural)

The average upload bandwidth rate limit in bits per second.

ubrlGatewayARN :: Lens' UpdateBandwidthRateLimit Text

Response

data UpdateBandwidthRateLimitResponse

Response constructor

updateBandwidthRateLimitResponse :: UpdateBandwidthRateLimitResponse

UpdateBandwidthRateLimitResponse constructor.

The fields accessible through corresponding lenses are:

ubrlrGatewayARN :: Maybe Text

Response lenses

ubrlrGatewayARN :: Lens' UpdateBandwidthRateLimitResponse (Maybe Text)