Network.AWS.StorageGateway.UpdateChapCredentials

Description

Updates the Challenge-Handshake Authentication Protocol (CHAP) credentials for a specified iSCSI target. By default, a gateway does not have CHAP enabled; however, for added security, you might use it.

Important: When you update CHAP credentials, all existing connections on the target are closed and initiators must reconnect with the new credentials.

Creating a Request

updateChapCredentials

Arguments

:: Text
uccTargetARN

-> Text
uccSecretToAuthenticateInitiator

-> Text
uccInitiatorName

-> UpdateChapCredentials

Creates a value of UpdateChapCredentials with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uccSecretToAuthenticateTarget - The secret key that the target must provide to participate in mutual CHAP with the initiator (e.g. Windows client). Byte constraints: Minimum bytes of 12. Maximum bytes of 16.

uccTargetARN - The Amazon Resource Name (ARN) of the iSCSI volume target. Use the DescribeStorediSCSIVolumes operation to return the TargetARN for specified VolumeARN.

uccSecretToAuthenticateInitiator - The secret key that the initiator (for example, the Windows client) must provide to participate in mutual CHAP with the target.

uccInitiatorName - The iSCSI initiator that connects to the target.

data UpdateChapCredentials

A JSON object containing one or more of the following fields:

'UpdateChapCredentialsInput$InitiatorName'

'UpdateChapCredentialsInput$SecretToAuthenticateInitiator'

'UpdateChapCredentialsInput$SecretToAuthenticateTarget'

'UpdateChapCredentialsInput$TargetARN'

See: updateChapCredentials smart constructor.

Request Lenses

uccSecretToAuthenticateTarget :: Lens' UpdateChapCredentials (Maybe Text)

The secret key that the target must provide to participate in mutual CHAP with the initiator (e.g. Windows client). Byte constraints: Minimum bytes of 12. Maximum bytes of 16.

uccTargetARN :: Lens' UpdateChapCredentials Text

The Amazon Resource Name (ARN) of the iSCSI volume target. Use the DescribeStorediSCSIVolumes operation to return the TargetARN for specified VolumeARN.

uccSecretToAuthenticateInitiator :: Lens' UpdateChapCredentials Text

The secret key that the initiator (for example, the Windows client) must provide to participate in mutual CHAP with the target.

uccInitiatorName :: Lens' UpdateChapCredentials Text

The iSCSI initiator that connects to the target.

Destructuring the Response

updateChapCredentialsResponse

Arguments

:: Int
uccrsResponseStatus

-> UpdateChapCredentialsResponse

Creates a value of UpdateChapCredentialsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uccrsTargetARN - The Amazon Resource Name (ARN) of the target. This is the same target specified in the request.

uccrsInitiatorName - The iSCSI initiator that connects to the target. This is the same initiator name specified in the request.

uccrsResponseStatus - -- | The response status code.

data UpdateChapCredentialsResponse

A JSON object containing the following fields:

See: updateChapCredentialsResponse smart constructor.

Response Lenses

uccrsTargetARN :: Lens' UpdateChapCredentialsResponse (Maybe Text)

The Amazon Resource Name (ARN) of the target. This is the same target specified in the request.

uccrsInitiatorName :: Lens' UpdateChapCredentialsResponse (Maybe Text)

The iSCSI initiator that connects to the target. This is the same initiator name specified in the request.

uccrsResponseStatus :: Lens' UpdateChapCredentialsResponse Int

- | The response status code.