Network.AWS.StorageGateway.UpdateGatewaySoftwareNow

Description

This operation updates the gateway virtual machine (VM) software. The request immediately triggers the software update.

When you make this request, you get a '200 OK' success response immediately. However, it might take some time for the update to complete. You can call DescribeGatewayInformation to verify the gateway is in the STATE_RUNNING state. A software update forces a system restart of your gateway. You can minimize the chance of any disruption to your applications by increasing your iSCSI Initiators' timeouts. For more information about increasing iSCSI Initiator timeouts for Windows and Linux, see Customizing Your Windows iSCSI Settings and Customizing Your Linux iSCSI Settings, respectively.

http://docs.aws.amazon.com/storagegateway/latest/APIReference/API_UpdateGatewaySoftwareNow.html

Request

data UpdateGatewaySoftwareNow

Request constructor

updateGatewaySoftwareNow

Arguments

:: Text
ugsnGatewayARN

-> UpdateGatewaySoftwareNow

UpdateGatewaySoftwareNow constructor.

The fields accessible through corresponding lenses are:

ugsnGatewayARN :: Text

Request lenses

ugsnGatewayARN :: Lens' UpdateGatewaySoftwareNow Text

Response

data UpdateGatewaySoftwareNowResponse

Response constructor

updateGatewaySoftwareNowResponse :: UpdateGatewaySoftwareNowResponse

UpdateGatewaySoftwareNowResponse constructor.

The fields accessible through corresponding lenses are:

ugsnrGatewayARN :: Maybe Text

Response lenses

ugsnrGatewayARN :: Lens' UpdateGatewaySoftwareNowResponse (Maybe Text)