Network.AWS.StorageGateway.UpdateMaintenanceStartTime

Description

Updates a gateway's weekly maintenance start time information, including day and time of the week. The maintenance time is the time in your gateway's time zone.

Creating a Request

updateMaintenanceStartTime

Arguments

:: Text
umstGatewayARN

-> Natural
umstHourOfDay

-> Natural
umstMinuteOfHour

-> Natural
umstDayOfWeek

-> UpdateMaintenanceStartTime

Creates a value of UpdateMaintenanceStartTime with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

umstGatewayARN - Undocumented member.

umstHourOfDay - The hour component of the maintenance start time represented as hh , where hh is the hour (00 to 23). The hour of the day is in the time zone of the gateway.

umstMinuteOfHour - The minute component of the maintenance start time represented as mm , where mm is the minute (00 to 59). The minute of the hour is in the time zone of the gateway.

umstDayOfWeek - The maintenance start time day of the week represented as an ordinal number from 0 to 6, where 0 represents Sunday and 6 Saturday.

data UpdateMaintenanceStartTime

A JSON object containing the following fields:

'UpdateMaintenanceStartTimeInput$DayOfWeek'

'UpdateMaintenanceStartTimeInput$HourOfDay'

'UpdateMaintenanceStartTimeInput$MinuteOfHour'

See: updateMaintenanceStartTime smart constructor.

Request Lenses

umstGatewayARN :: Lens' UpdateMaintenanceStartTime Text

Undocumented member.

umstHourOfDay :: Lens' UpdateMaintenanceStartTime Natural

The hour component of the maintenance start time represented as hh , where hh is the hour (00 to 23). The hour of the day is in the time zone of the gateway.

umstMinuteOfHour :: Lens' UpdateMaintenanceStartTime Natural

The minute component of the maintenance start time represented as mm , where mm is the minute (00 to 59). The minute of the hour is in the time zone of the gateway.

umstDayOfWeek :: Lens' UpdateMaintenanceStartTime Natural

The maintenance start time day of the week represented as an ordinal number from 0 to 6, where 0 represents Sunday and 6 Saturday.

Destructuring the Response

updateMaintenanceStartTimeResponse

Arguments

:: Int
umstrsResponseStatus

-> UpdateMaintenanceStartTimeResponse

Creates a value of UpdateMaintenanceStartTimeResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

umstrsGatewayARN - Undocumented member.

umstrsResponseStatus - -- | The response status code.

data UpdateMaintenanceStartTimeResponse

A JSON object containing the of the gateway whose maintenance start time is updated.

See: updateMaintenanceStartTimeResponse smart constructor.

Response Lenses

umstrsGatewayARN :: Lens' UpdateMaintenanceStartTimeResponse (Maybe Text)

Undocumented member.

umstrsResponseStatus :: Lens' UpdateMaintenanceStartTimeResponse Int

- | The response status code.