Network.AWS.StorageGateway.UpdateMaintenanceStartTime

Description

This operation updates a gateway's weekly maintenance start time information, including day and time of the week. The maintenance time is the time in your gateway's time zone.

http://docs.aws.amazon.com/storagegateway/latest/APIReference/API_UpdateMaintenanceStartTime.html

Request

data UpdateMaintenanceStartTime

Request constructor

updateMaintenanceStartTime

Arguments

:: Text
umstGatewayARN

-> Natural
umstHourOfDay

-> Natural
umstMinuteOfHour

-> Natural
umstDayOfWeek

-> UpdateMaintenanceStartTime

UpdateMaintenanceStartTime constructor.

The fields accessible through corresponding lenses are:

umstDayOfWeek :: Natural

umstGatewayARN :: Text

umstHourOfDay :: Natural

umstMinuteOfHour :: Natural

Request lenses

umstDayOfWeek :: Lens' UpdateMaintenanceStartTime Natural

The maintenance start time day of the week.

umstGatewayARN :: Lens' UpdateMaintenanceStartTime Text

umstHourOfDay :: Lens' UpdateMaintenanceStartTime Natural

The hour component of the maintenance start time represented as hh, where hh is the hour (00 to 23). The hour of the day is in the time zone of the gateway.

umstMinuteOfHour :: Lens' UpdateMaintenanceStartTime Natural

The minute component of the maintenance start time represented as mm, where mm is the minute (00 to 59). The minute of the hour is in the time zone of the gateway.

Response

data UpdateMaintenanceStartTimeResponse

Response constructor

updateMaintenanceStartTimeResponse :: UpdateMaintenanceStartTimeResponse

UpdateMaintenanceStartTimeResponse constructor.

The fields accessible through corresponding lenses are:

umstrGatewayARN :: Maybe Text

Response lenses

umstrGatewayARN :: Lens' UpdateMaintenanceStartTimeResponse (Maybe Text)