Network.AWS.StorageGateway.V2013_06_30.CreateCachediSCSIVolume

Description

This operation creates a cached volume on a specified cached gateway. This operation is supported only for the gateway-cached volume architecture. Cache storage must be allocated to the gateway before you can create a cached volume. Use the AddCache operation to add cache storage to a gateway. In the request, you must specify the gateway, size of the volume in bytes, the iSCSI target name, an IP address on which to expose the target, and a unique client token. In response, AWS Storage Gateway creates the volume and returns information about it such as the volume Amazon Resource Name (ARN), its size, and the iSCSI target ARN that initiators can use to connect to the volume target. Example Request The following example shows a request that specifies that a local disk of a gateway be configured as a cached volume. POST HTTP1.1 Host: storagegateway.us-east-1.amazonaws.com Content-Type: application/x-amz-json-1.1 Authorization: AWS4-HMAC-SHA256 Credential=AKIAIOSFODNN7EXAMPLE20120425us-east-1storagegatewayaws4_request, SignedHeaders=content-type;host;x-amz-date;x-amz-target, Signature=9cd5a3584d1d67d57e61f120f35102d6b3649066abdd4bf4bbcf05bd9f2f8fe2 x-amz-date: 20120912T120000Z x-amz-target: StorageGateway_20120630.CreateCachediSCSIVolume { ClientToken: "cachedvol112233", GatewayARN: "arn:aws:storagegateway:us-east-1:111122223333:gateway/mygateway", NetworkInterfaceId: "10.1.1.1", TargetName: "myvolume", VolumeSizeInBytes: 536870912000 } HTTP/1.1 200 OK x-amzn-RequestId: gur28r2rqlgb8vvs0mq17hlgij1q8glle1qeu3kpgg6f0kstauu0 Date: Wed, 12 Sep 2012 12:00:02 GMT Content-Type: application/x-amz-json-1.1 Content-length: 263 { TargetARN: "arn:aws:storagegateway:us-east-1:111122223333:gatewaymygatewaytarget/iqn.1997-05.com.amazon:myvolume", VolumeARN: "arn:aws:storagegateway:us-east-1:111122223333:gatewaymygatewayvolume/vol-1122AABB" }.

Request

data CreateCachediSCSIVolume

Request constructor

mkCreateCachediSCSIVolume

Arguments

:: Text
ccscsivGatewayARN

-> Integer
ccscsivVolumeSizeInBytes

-> Text
ccscsivTargetName

-> Text
ccscsivNetworkInterfaceId

-> Text
ccscsivClientToken

-> CreateCachediSCSIVolume

Smart constructor for the minimum required parameters to construct a valid CreateCachediSCSIVolume request.

Request lenses

ccscsivGatewayARN :: Lens' CreateCachediSCSIVolume Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

ccscsivVolumeSizeInBytes :: Lens' CreateCachediSCSIVolume Integer

ccscsivSnapshotId :: Lens' CreateCachediSCSIVolume (Maybe Text)

ccscsivTargetName :: Lens' CreateCachediSCSIVolume Text

ccscsivNetworkInterfaceId :: Lens' CreateCachediSCSIVolume Text

ccscsivClientToken :: Lens' CreateCachediSCSIVolume Text

Response

data CreateCachediSCSIVolumeResponse

Response lenses

ccscsivrVolumeARN :: Lens' CreateCachediSCSIVolumeResponse (Maybe Text)

ccscsivrTargetARN :: Lens' CreateCachediSCSIVolumeResponse (Maybe Text)