Network.AWS.StorageGateway.V2013_06_30.Types

Description

AWS Storage Gateway is a service that connects an on-premises software appliance with cloud-based storage to provide seamless and secure integration between your on-premises IT environment and AWS's storage infrastructure.

Documentation

data StorageGateway

Supported version (2013-06-30) of the AWS Storage Gateway service.

data ErrorCode

Additional information about the error.

Constructors

ErrorCodeActivationKeyExpired
ActivationKeyExpired

ErrorCodeActivationKeyInvalid
ActivationKeyInvalid

ErrorCodeActivationKeyNotFound
ActivationKeyNotFound

ErrorCodeAuthenticationFailure
AuthenticationFailure

ErrorCodeBandwidthThrottleScheduleNotFound
BandwidthThrottleScheduleNotFound

ErrorCodeBlocked
Blocked

ErrorCodeCannotExportSnapshot
CannotExportSnapshot

ErrorCodeChapCredentialNotFound
ChapCredentialNotFound

ErrorCodeDiskAlreadyAllocated
DiskAlreadyAllocated

ErrorCodeDiskDoesNotExist
DiskDoesNotExist

ErrorCodeDiskSizeGreaterThanVolumeMaxSize
DiskSizeGreaterThanVolumeMaxSize

ErrorCodeDiskSizeLessThanVolumeSize
DiskSizeLessThanVolumeSize

ErrorCodeDiskSizeNotGigAligned
DiskSizeNotGigAligned

ErrorCodeDuplicateCertificateInfo
DuplicateCertificateInfo

ErrorCodeDuplicateSchedule
DuplicateSchedule

ErrorCodeEndpointNotFound
EndpointNotFound

ErrorCodeGatewayInternalError
GatewayInternalError

ErrorCodeGatewayNotConnected
GatewayNotConnected

ErrorCodeGatewayNotFound
GatewayNotFound

ErrorCodeGatewayProxyNetworkConnectionBusy
GatewayProxyNetworkConnectionBusy

ErrorCodeIAMNotSupported
IAMNotSupported

ErrorCodeInitiatorInvalid
InitiatorInvalid

ErrorCodeInitiatorNotFound
InitiatorNotFound

ErrorCodeInternalError
InternalError

ErrorCodeInvalidEndpoint
InvalidEndpoint

ErrorCodeInvalidGateway
InvalidGateway

ErrorCodeInvalidParameters
InvalidParameters

ErrorCodeInvalidSchedule
InvalidSchedule

ErrorCodeLocalStorageLimitExceeded
LocalStorageLimitExceeded

ErrorCodeLunAlreadyAllocated
LunAlreadyAllocated

ErrorCodeLunInvalid
LunInvalid

ErrorCodeMaximumContentLengthExceeded
MaximumContentLengthExceeded

ErrorCodeMaximumTapeCartridgeCountExceeded
MaximumTapeCartridgeCountExceeded

ErrorCodeMaximumVolumeCountExceeded
MaximumVolumeCountExceeded

ErrorCodeNetworkConfigurationChanged
NetworkConfigurationChanged

ErrorCodeNoDisksAvailable
NoDisksAvailable

ErrorCodeNotImplemented
NotImplemented

ErrorCodeNotSupported
NotSupported

ErrorCodeOperationAborted
OperationAborted

ErrorCodeOutdatedGateway
OutdatedGateway

ErrorCodeParametersNotImplemented
ParametersNotImplemented

ErrorCodeRegionInvalid
RegionInvalid

ErrorCodeRequestTimeout
RequestTimeout

ErrorCodeServiceUnavailable
ServiceUnavailable

ErrorCodeSnapshotDeleted
SnapshotDeleted

ErrorCodeSnapshotIdInvalid
SnapshotIdInvalid

ErrorCodeSnapshotInProgress
SnapshotInProgress

ErrorCodeSnapshotNotFound
SnapshotNotFound

ErrorCodeSnapshotScheduleNotFound
SnapshotScheduleNotFound

ErrorCodeStagingAreaFull
StagingAreaFull

ErrorCodeStorageFailure
StorageFailure

ErrorCodeTapeCartridgeNotFound
TapeCartridgeNotFound

ErrorCodeTargetAlreadyExists
TargetAlreadyExists

ErrorCodeTargetInvalid
TargetInvalid

ErrorCodeTargetNotFound
TargetNotFound

ErrorCodeUnauthorizedOperation
UnauthorizedOperation

ErrorCodeVolumeAlreadyExists
VolumeAlreadyExists

ErrorCodeVolumeIdInvalid
VolumeIdInvalid

ErrorCodeVolumeInUse
VolumeInUse

ErrorCodeVolumeNotFound
VolumeNotFound

ErrorCodeVolumeNotReady
VolumeNotReady

data CachediSCSIVolumeInformation

Constructors

CachediSCSIVolumeInformation

Fields

_cscsiviVolumeiSCSIAttributes :: Maybe VolumeiSCSIAttributes

Lists iSCSI information about a volume.

_cscsiviVolumeStatus :: Maybe Text

_cscsiviSourceSnapshotId :: Maybe Text

_cscsiviVolumeARN :: Maybe Text

_cscsiviVolumeProgress :: Maybe Double

_cscsiviVolumeSizeInBytes :: Maybe Integer

_cscsiviVolumeId :: Maybe Text

_cscsiviVolumeType :: Maybe Text

data ChapInfo

Describes Challenge-Handshake Authentication Protocol (CHAP) information that supports authentication between your gateway and iSCSI initiators.

Constructors

ChapInfo

Fields

_ciTargetARN :: Maybe Text

The Amazon Resource Name (ARN) of the volume. Valid Values: 50 to 500 lowercase letters, numbers, periods (.), and hyphens (-).

_ciSecretToAuthenticateInitiator :: Maybe Text

The secret key that the initiator (e.g. Windows client) must provide to participate in mutual CHAP with the target.

_ciInitiatorName :: Maybe Text

The iSCSI initiator that connects to the target.

_ciSecretToAuthenticateTarget :: Maybe Text

The secret key that the target must provide to participate in mutual CHAP with the initiator (e.g. Windows client).

data DeviceiSCSIAttributes

Constructors

DeviceiSCSIAttributes

Fields

_dscsiaTargetARN :: Maybe Text

_dscsiaChapEnabled :: Maybe Bool

_dscsiaNetworkInterfaceId :: Maybe Text

_dscsiaNetworkInterfacePort :: Maybe Integer

data DiskInformation

Constructors

DiskInformation

Fields

_dwDiskAllocationResource :: Maybe Text

_dwDiskAllocationType :: Maybe Text

_dwDiskNode :: Maybe Text

_dwDiskPath :: Maybe Text

_dwDiskSizeInBytes :: Maybe Integer

_dwDiskId :: Maybe Text

data GatewayInformation

Constructors

GatewayInformation

Fields

_gjGatewayARN :: Maybe Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

_gjGatewayType :: Maybe Text

data NetworkInterface

Describes a gateway's network interface.

Constructors

NetworkInterface

Fields

_niIpv6Address :: Maybe Text

The Internet Protocol version 6 (IPv6) address of the interface. Currently not supported.

_niMacAddress :: Maybe Text

The Media Access Control (MAC) address of the interface. This is currently unsupported and will not be returned in output.

_niIpv4Address :: Maybe Text

The Internet Protocol version 4 (IPv4) address of the interface.

data StorageGatewayError

A StorageGatewayError that provides more detail about the cause of the error.

Constructors

StorageGatewayError

Fields

_sgeErrorDetails :: Map Text Text

Human-readable text that provides detail about the error that occured.

_sgeErrorCode :: Maybe ErrorCode

Additional information about the error.

data StorediSCSIVolumeInformation

Constructors

StorediSCSIVolumeInformation

Fields

_sscsiviVolumeiSCSIAttributes :: Maybe VolumeiSCSIAttributes

Lists iSCSI information about a volume.

_sscsiviVolumeStatus :: Maybe Text

_sscsiviSourceSnapshotId :: Maybe Text

_sscsiviPreservedExistingData :: Maybe Bool

_sscsiviVolumeARN :: Maybe Text

_sscsiviVolumeProgress :: Maybe Double

_sscsiviVolumeSizeInBytes :: Maybe Integer

_sscsiviVolumeId :: Maybe Text

_sscsiviVolumeDiskId :: Maybe Text

_sscsiviVolumeType :: Maybe Text

data Tape

Constructors

Tape

Fields

_vTapeBarcode :: Maybe Text

_vTapeStatus :: Maybe Text

_vTapeARN :: Maybe Text

_vProgress :: Maybe Double

_vTapeSizeInBytes :: Maybe Integer

_vVTLDevice :: Maybe Text

data TapeArchive

Constructors

TapeArchive

Fields

_tcTapeBarcode :: Maybe Text

_tcTapeStatus :: Maybe Text

_tcTapeARN :: Maybe Text

_tcTapeSizeInBytes :: Maybe Integer

_tcCompletionTime :: Maybe ISO8601

_tcRetrievedTo :: Maybe Text

The Amazon Resource Name (ARN) of the gateway. Use the ListGateways operation to return a list of gateways for your account and region.

data TapeRecoveryPointInfo

Constructors

TapeRecoveryPointInfo

Fields

_trpjTapeRecoveryPointTime :: Maybe ISO8601

_trpjTapeARN :: Maybe Text

_trpjTapeSizeInBytes :: Maybe Integer

data VTLDevice

Constructors

VTLDevice

Fields

_vtleDeviceiSCSIAttributes :: Maybe DeviceiSCSIAttributes

_vtleVTLDeviceARN :: Maybe Text

data VolumeInformation

Constructors

VolumeInformation

Fields

_vlVolumeARN :: Maybe Text

_vlVolumeType :: Maybe Text

data VolumeRecoveryPointInfo

Constructors

VolumeRecoveryPointInfo

Fields

_vrpjVolumeRecoveryPointTime :: Maybe Text

_vrpjVolumeARN :: Maybe Text

_vrpjVolumeSizeInBytes :: Maybe Integer

_vrpjVolumeUsageInBytes :: Maybe Integer

data VolumeiSCSIAttributes

Lists iSCSI information about a volume.

Constructors

VolumeiSCSIAttributes

Fields

_vscsiaLunNumber :: Maybe Integer

The logical disk number.

_vscsiaTargetARN :: Maybe Text

The Amazon Resource Name (ARN) of the volume target.

_vscsiaChapEnabled :: Maybe Bool

Indicates whether mutual CHAP is enabled for the iSCSI target.

_vscsiaNetworkInterfaceId :: Maybe Text

The network interface identifier.

_vscsiaNetworkInterfacePort :: Maybe Integer

The port used to communicate with iSCSI targets.

cscsiviVolumeiSCSIAttributes :: Lens' CachediSCSIVolumeInformation (Maybe VolumeiSCSIAttributes)

cscsiviVolumeType :: Lens' CachediSCSIVolumeInformation (Maybe Text)

cscsiviVolumeStatus :: Lens' CachediSCSIVolumeInformation (Maybe Text)

cscsiviVolumeSizeInBytes :: Lens' CachediSCSIVolumeInformation (Maybe Integer)

cscsiviVolumeProgress :: Lens' CachediSCSIVolumeInformation (Maybe Double)

cscsiviVolumeId :: Lens' CachediSCSIVolumeInformation (Maybe Text)

cscsiviVolumeARN :: Lens' CachediSCSIVolumeInformation (Maybe Text)

cscsiviSourceSnapshotId :: Lens' CachediSCSIVolumeInformation (Maybe Text)

ciTargetARN :: Lens' ChapInfo (Maybe Text)

ciSecretToAuthenticateTarget :: Lens' ChapInfo (Maybe Text)

ciSecretToAuthenticateInitiator :: Lens' ChapInfo (Maybe Text)

ciInitiatorName :: Lens' ChapInfo (Maybe Text)

dscsiaTargetARN :: Lens' DeviceiSCSIAttributes (Maybe Text)

dscsiaNetworkInterfacePort :: Lens' DeviceiSCSIAttributes (Maybe Integer)

dscsiaNetworkInterfaceId :: Lens' DeviceiSCSIAttributes (Maybe Text)

dscsiaChapEnabled :: Lens' DeviceiSCSIAttributes (Maybe Bool)

dwDiskSizeInBytes :: Lens' DiskInformation (Maybe Integer)

dwDiskPath :: Lens' DiskInformation (Maybe Text)

dwDiskNode :: Lens' DiskInformation (Maybe Text)

dwDiskId :: Lens' DiskInformation (Maybe Text)

dwDiskAllocationType :: Lens' DiskInformation (Maybe Text)

dwDiskAllocationResource :: Lens' DiskInformation (Maybe Text)

gjGatewayType :: Lens' GatewayInformation (Maybe Text)

gjGatewayARN :: Lens' GatewayInformation (Maybe Text)

niMacAddress :: Lens' NetworkInterface (Maybe Text)

niIpv6Address :: Lens' NetworkInterface (Maybe Text)

niIpv4Address :: Lens' NetworkInterface (Maybe Text)

sgeErrorDetails :: Lens' StorageGatewayError (Map Text Text)

sgeErrorCode :: Lens' StorageGatewayError (Maybe ErrorCode)

sscsiviVolumeiSCSIAttributes :: Lens' StorediSCSIVolumeInformation (Maybe VolumeiSCSIAttributes)

sscsiviVolumeType :: Lens' StorediSCSIVolumeInformation (Maybe Text)

sscsiviVolumeStatus :: Lens' StorediSCSIVolumeInformation (Maybe Text)

sscsiviVolumeSizeInBytes :: Lens' StorediSCSIVolumeInformation (Maybe Integer)

sscsiviVolumeProgress :: Lens' StorediSCSIVolumeInformation (Maybe Double)

sscsiviVolumeId :: Lens' StorediSCSIVolumeInformation (Maybe Text)

sscsiviVolumeDiskId :: Lens' StorediSCSIVolumeInformation (Maybe Text)

sscsiviVolumeARN :: Lens' StorediSCSIVolumeInformation (Maybe Text)

sscsiviSourceSnapshotId :: Lens' StorediSCSIVolumeInformation (Maybe Text)

sscsiviPreservedExistingData :: Lens' StorediSCSIVolumeInformation (Maybe Bool)

vVTLDevice :: Lens' Tape (Maybe Text)

vTapeStatus :: Lens' Tape (Maybe Text)

vTapeSizeInBytes :: Lens' Tape (Maybe Integer)

vTapeBarcode :: Lens' Tape (Maybe Text)

vTapeARN :: Lens' Tape (Maybe Text)

vProgress :: Lens' Tape (Maybe Double)

tcTapeStatus :: Lens' TapeArchive (Maybe Text)

tcTapeSizeInBytes :: Lens' TapeArchive (Maybe Integer)

tcTapeBarcode :: Lens' TapeArchive (Maybe Text)

tcTapeARN :: Lens' TapeArchive (Maybe Text)

tcRetrievedTo :: Lens' TapeArchive (Maybe Text)

tcCompletionTime :: Lens' TapeArchive (Maybe ISO8601)

trpjTapeSizeInBytes :: Lens' TapeRecoveryPointInfo (Maybe Integer)

trpjTapeRecoveryPointTime :: Lens' TapeRecoveryPointInfo (Maybe ISO8601)

trpjTapeARN :: Lens' TapeRecoveryPointInfo (Maybe Text)

vtleVTLDeviceARN :: Lens' VTLDevice (Maybe Text)

vtleDeviceiSCSIAttributes :: Lens' VTLDevice (Maybe DeviceiSCSIAttributes)

vlVolumeType :: Lens' VolumeInformation (Maybe Text)

vlVolumeARN :: Lens' VolumeInformation (Maybe Text)

vrpjVolumeUsageInBytes :: Lens' VolumeRecoveryPointInfo (Maybe Integer)

vrpjVolumeSizeInBytes :: Lens' VolumeRecoveryPointInfo (Maybe Integer)

vrpjVolumeRecoveryPointTime :: Lens' VolumeRecoveryPointInfo (Maybe Text)

vrpjVolumeARN :: Lens' VolumeRecoveryPointInfo (Maybe Text)

vscsiaTargetARN :: Lens' VolumeiSCSIAttributes (Maybe Text)

vscsiaNetworkInterfacePort :: Lens' VolumeiSCSIAttributes (Maybe Integer)

vscsiaNetworkInterfaceId :: Lens' VolumeiSCSIAttributes (Maybe Text)

vscsiaLunNumber :: Lens' VolumeiSCSIAttributes (Maybe Integer)

vscsiaChapEnabled :: Lens' VolumeiSCSIAttributes (Maybe Bool)